% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/z_examples.R
\docType{data}
\name{additional.residency.results}
\alias{additional.residency.results}
\title{Example residency results}
\format{
A list of outputs from residency()
}
\source{
Data collected by the authors.
}
\usage{
additional.residency.results
}
\description{
A list with residency results to append to example.results, so some examples can run.
}
\keyword{internal}
