% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{advEfficiency}
\alias{advEfficiency}
\title{Calculate beta estimations for efficiency}
\usage{
advEfficiency(
  x,
  labels = NULL,
  n = 10000,
  q = c(0.025, 0.5, 0.975),
  force.grid = NULL,
  paired = TRUE,
  title = ""
)
}
\arguments{
\item{x}{An efficiency object from actel (\code{overall.CJS}, \code{intra.array.CJS[[array]]} or \code{efficiency} objects)}

\item{labels}{a vector of strings to substitute default plot labels}

\item{n}{Number of draws for each array.}

\item{q}{The quantile values to be calculated. Defaults to code{c(0.025, 0.5, 0.975)} (i.e. median and 95\% CI)}

\item{force.grid}{A vector of format c(nrow, ncol) that allows the user to define the number of rows and columns to distribute the plots in.}

\item{paired}{Logical: For efficiency derived from residency analyses, should min and max estimates for an array be paired together?}

\item{title}{A title for the plot (feeds into ggplot's labs title parameter).}
}
\value{
A dataframe with the requred quantile values and a plot of the efficiency distributions.
}
\description{
Calculate beta estimations for efficiency
}
