% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{appendTo}
\alias{appendTo}
\title{Append to ...}
\usage{
appendTo(recipient, line, tag)
}
\arguments{
\item{recipient}{'Screen' displays the message on screen, 'Report' appends the message to 'temp_log.txt', 'Warning' appends the message to 'temp_warnings.txt', 'UD' appends the message to 'temp_UD.txt', 'Comment' appends the message to 'temp_comments.txt'. The same message may be appended to multiple recipients at once.}

\item{line}{The text to be appended.}

\item{tag}{the tag number to which the comment belongs. Only used when recipient = 'Comment'.}
}
\value{
No return value, called for side effects.
}
\description{
Appends a note/comment to the specified recipient, which in turn corresponds to a temporary helper file.
}
\keyword{internal}
