% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assembleEfficiency.R
\name{assembleEfficiency}
\alias{assembleEfficiency}
\title{Efficiency calculation wrapper}
\usage{
assembleEfficiency(sections, spatial, detections.list, movements,
  minimum.detections)
}
\arguments{
\item{sections}{The sections in which the study was divided. Must be coincident with the names given to the ALS arrays. (i.e. if an array is 'River1', then the respective section is 'River')}

\item{spatial}{A list of the spatial objects, created by assembleSpatial.}

\item{detections.list}{A list of the detections split by each target tag, created by splitDetections.}

\item{movements}{A list of movements for each target tag, created by groupMovements.}

\item{minimum.detections}{The mininum number of times a tag must have been recorded during the study period for it to be considered a true tag and not random noise.}
}
\value{
A list containing 1) the intersection efficiency, 2) the inter-array efficiency and 3) the the intra-array efficiency.
}
\description{
Collects the input from interSection, inter-array and intra-array into a list.
}
\keyword{internal}
