% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CJS.R
\name{assembleGroupCJS}
\alias{assembleGroupCJS}
\title{Assembles CJS tables for all groups}
\usage{
assembleGroupCJS(mat, CJS, arrays)
}
\arguments{
\item{mat}{the original presence/absence matrices}

\item{CJS}{A list of CJS calculated for each group x array combinations}

\item{arrays}{a list of arrays}
}
\value{
A CJS results table for each group
}
\description{
Assembles CJS tables for all groups
}
\keyword{internal}
