% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CJS.R
\name{assembleMatrices}
\alias{assembleMatrices}
\title{Assemble detection matrices}
\usage{
assembleMatrices(spatial, simple.movements, status.df)
}
\arguments{
\item{simple.movements}{A list of valid-only movements for each fish}
}
\value{
A list of detection matrices split by groups and release sites
}
\description{
Assemble detection matrices
}
\keyword{internal}
