% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assembleOutput.R
\name{assembleOutput}
\alias{assembleOutput}
\title{Create status.df}
\usage{
assembleOutput(timetable, bio, movements, spatial, sections, dist.mat,
  invalid.dist, tz.study.area)
}
\arguments{
\item{timetable}{A table of the entering and leaving points for each section per target tag, created by assembleTimetable.}

\item{bio}{A table with the tags and biometrics of the studied fish.}

\item{movements}{A list of movements for each target tag, created by groupMovements.}

\item{spatial}{A list of the spatial objects, created by assembleSpatial.}

\item{sections}{The sections in which the study was divided. Must be coincident with the names given to the ALS arrays. (i.e. if an array is 'River1', then the respective section is 'River')}

\item{dist.mat}{A matrix of the distances between the deployed ALS.}

\item{invalid.dist}{Wether or not the distance matrix suplied is valid for the study area.}

\item{tz.study.area}{The timezone of the study area. Necessary to conver the ALS time data, which is in UTC.}
}
\value{
A dataframe containing all the final data for each fish.
}
\description{
Combines the timetable and the original biometrics.
}
\keyword{internal}
