% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assembleOverview.R
\name{assembleOverview}
\alias{assembleOverview}
\title{Create group.overview}
\usage{
assembleOverview(status.df, sections, dist.mat, invalid.dist)
}
\arguments{
\item{status.df}{A dataframe with all the final data for each fish, created by assembleOutput.}

\item{sections}{The sections in which the study was divided. Must be coincident with the names given to the ALS arrays. (i.e. if an array is 'River1', then the respective section is 'River')}

\item{dist.mat}{A matrix of the distances between the deployed ALS.}

\item{invalid.dist}{Wether or not the distance matrix suplied is valid for the study area.}
}
\value{
A dataframe containing the survival per group of fish present in the biometrics.
}
\description{
Produces a table with the survival per group of fish present in the biometrics.
}
\keyword{internal}
