% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assembleProgression.R
\name{assembleProgression}
\alias{assembleProgression}
\title{Calculate the progression}
\usage{
assembleProgression(status.df, spatial)
}
\arguments{
\item{status.df}{A dataframe with all the final data for each fish, created by assembleOutput.}

\item{spatial}{A list of the spatial objects, created by assembleSpatial.}
}
\value{
A dataframe containing the number of fish that crossed each ALS array.
}
\description{
Produces a table with the number of fish that crossed each ALS array.
}
\keyword{internal}
