% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/residency.R
\name{assembleResidency}
\alias{assembleResidency}
\title{Collect summary information for the residency analysis}
\usage{
assembleResidency(secmoves, movements, sections)
}
\arguments{
\item{secmoves}{the section-movements}

\item{movements}{the array-movements}

\item{sections}{The sections in which the study was divided. Must be coincident with the names given to the ALS arrays. (i.e. if an array is 'River1', then the respective section is 'River')}
}
\value{
A residency summary table
}
\description{
Collect summary information for the residency analysis
}
\keyword{internal}
