% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/z_deprecated.R
\name{assembleSpatial}
\alias{assembleSpatial}
\title{Process spatial elements}
\usage{
assembleSpatial(file, bio, sections = NULL)
}
\arguments{
\item{file}{an input file with spatial data.}

\item{bio}{A table with the tags and biometrics of the studied fish.}

\item{sections}{The sections in which the study was divided. Must be coincident with the names given to the ALS arrays. (i.e. if an array is 'River1', then the respective section is 'River')}
}
\value{
A list of 1) stations, 2) release sites, 3) ALS columns in the spatial file, 4) the Number of ASL, 5) The ALS serial numbers and 6) the array order.
}
\description{
Creates a list containing multiple spatial elements required throughout the analyses
}
\keyword{internal}
