% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/migration.R
\name{assembleTimetable}
\alias{assembleTimetable}
\title{Create the timetable}
\usage{
assembleTimetable(
  vm,
  all.moves,
  sections,
  arrays,
  dist.mat,
  invalid.dist,
  speed.method,
  if.last.skip.section,
  success.arrays
)
}
\arguments{
\item{sections}{The sections in which the study area is divided. Must be
coincident with the names given to the arrays. See the vignettes for more
details.}

\item{arrays}{A list containing information for each array.}

\item{dist.mat}{A matrix of the distances between the deployed ALS.}

\item{invalid.dist}{Whether or not the distance matrix supplied is valid for the study area.}

\item{speed.method}{Can take two forms: 'last to first' or 'first to first'.
If 'last to first' (default), the last detection on a given array is matched
to the first detection on the next array to perform the calculations.
If 'first to first', the first detection on a given array is matched to the
first detection on the next array to perform the calculations.}

\item{if.last.skip.section}{Logical: Should a fish detected at the last array
of a given section be considered to have disappeared in the next section?}

\item{success.arrays}{The arrays that mark the end of the study area. If a
fish crosses one of these arrays, it is considered to have successfully
migrated through the study area.}

\item{movements}{A list of movements for each target tag, created by groupMovements.}
}
\value{
A table of the entering and leaving points for each section per target tag
}
\description{
Crawls trough the movement events of each fish to find when it entered and left each section of the study area.
}
\keyword{internal}
