% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/movements.R
\name{assembleValidSecMoves}
\alias{assembleValidSecMoves}
\title{Wrapper for sectionMovements}
\usage{
assembleValidSecMoves(valid.moves, spatial, valid.dist)
}
\arguments{
\item{valid.moves}{A list of movements for each tag.}

\item{spatial}{The spatial list.}

\item{valid.dist}{Logical: Is a valid distances matrix being used?}
}
\value{
A list of valid movements
}
\description{
Wrapper for sectionMovements
}
\keyword{internal}
