% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transform.R
\name{assignArrays}
\alias{assignArrays}
\title{Match the detection's receiver to its respective array.}
\usage{
assignArrays(input, spatial)
}
\arguments{
\item{input}{list of detections}
}
\description{
Match the detection's receiver to its respective array.
}
\keyword{internal}
