% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadFunctions.R
\name{bindDetections}
\alias{bindDetections}
\title{Bind detections}
\usage{
bindDetections(data.files, file.list)
}
\arguments{
\item{data.files}{The list of processed files, supplied by loadDetections.}

\item{file.list}{The list of valid files, supplied by loadDetections.}
}
\value{
A dataframe of standardized detections from all the input files.
}
\description{
Combines all the standardized detections into one data frame.
}
\keyword{internal}
