% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CJS.R
\name{blameArrays}
\alias{blameArrays}
\title{Find which arrays to blame for a jump in movement events}
\usage{
blameArrays(from, to, paths)
}
\arguments{
\item{from}{The array where the tag started}

\item{to}{The array where the tag was next detected}

\item{paths}{a list containing the shortest paths between arrays with distance > 1}
}
\value{
A list containing information on the arrays that failed
}
\description{
Find which arrays to blame for a jump in movement events
}
\keyword{internal}
