% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/z_examples.R
\name{blankWorkspace}
\alias{blankWorkspace}
\title{Create a Blank Workspace}
\usage{
blankWorkspace(dir, force = FALSE)
}
\arguments{
\item{dir}{The name of the target directory. Will be created if not present.}
}
\value{
No return value, called for side effects
}
\description{
Produces template files and folders required to run the \code{\link{explore}},
\code{\link{migration}} and \code{\link{residency}} functions.
}
\examples{
\donttest{
# running blankWorkspace deploys template
# files to a directory specified by the user
blankWorkspace(paste0(tempdir(), "/createWorkspace_example"))
}

}
