% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CJS.R
\name{breakMatricesByArray}
\alias{breakMatricesByArray}
\title{Break the detection matrices per array}
\usage{
breakMatricesByArray(m, arrays, type = c("peers", "all"), verbose = TRUE)
}
\arguments{
\item{m}{the original presence/absence matrices}

\item{arrays}{a list of arrays}
}
\value{
a list of matrices, split by array
}
\description{
Break the detection matrices per array
}
\keyword{internal}
