% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check.R
\name{check_args}
\alias{check_args}
\title{check.R arguments}
\arguments{
\item{arrays}{a list containing information for each array}

\item{bio}{A table with the tags and biometrics of the studied fish.}

\item{detections.list}{A list of the detections split by each target tag, created by splitDetections.}

\item{dotmat}{The matrix of distances between arrays}

\item{fish}{The fish being analysed}

\item{movements, moves}{The movements table}

\item{release}{The release location of the fish}

\item{secmoves}{the section movements list}

\item{spatial}{The spatial data frame, as loaded by loadSpatial}

\item{valid.movements}{The valid movements table}
}
\description{
check.R arguments
}
\keyword{internal}
