% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check.R
\name{checkDeploymentStations}
\alias{checkDeploymentStations}
\title{Confirm that the station names in the deployments table match those listed in the spatial file}
\usage{
checkDeploymentStations(input, spatial)
}
\arguments{
\item{input}{the table of deployments}

\item{spatial}{The spatial data frame, as loaded by loadSpatial}
}
\description{
Confirm that the station names in the deployments table match those listed in the spatial file
}
\keyword{internal}
