% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check.R
\name{checkDeploymentTimes}
\alias{checkDeploymentTimes}
\title{Confirm that receivers were not re-deployed before being retrieved}
\usage{
checkDeploymentTimes(input)
}
\arguments{
\item{input}{the table of deployments}
}
\description{
Confirm that receivers were not re-deployed before being retrieved
}
\keyword{internal}
