% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check.R
\name{checkDupSignals}
\alias{checkDupSignals}
\title{Check if there are duplicated signals in the detected tags.}
\usage{
checkDupSignals(input, bio)
}
\arguments{
\item{input}{list of detections}

\item{bio}{A table with the tags and biometrics of the studied fish.}

\item{tag.list}{list of the target signals}
}
\value{
No return value, called for side effects.
}
\description{
Check if there are duplicated signals in the detected tags.
}
\keyword{internal}
