% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/z_deprecated.R
\name{checkEmptyReceivers}
\alias{checkEmptyReceivers}
\title{Check for unknown receivers in the raw detections}
\usage{
checkEmptyReceivers(spatial, detections)
}
\arguments{
\item{detections}{A data frame with all the detections. Supplied by loadDetections.}
}
\description{
Check for unknown receivers in the raw detections
}
\keyword{internal}
