% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check.R
\name{checkFirstMove}
\alias{checkFirstMove}
\title{Check for movements upstream of the release site.}
\usage{
checkFirstMove(
  movements,
  tag,
  bio,
  detections,
  dot_list,
  spatial,
  GUI,
  save.tables.locally,
  n
)
}
\arguments{
\item{movements}{The movements table for a specific tag.}

\item{tag}{The tag being analysed.}

\item{bio}{A table with the tags and biometrics of the studied animals.}

\item{detections}{The detections data.frame for a specific tag.}

\item{spatial}{A list of spatial objects in the study area.}

\item{GUI}{One of "needed", "always" or "never". If "needed", a new window is
opened to inspect the movements only if the movements table is too big to be
displayed in R's console. If "always", a graphical interface is always created
when the possibility to invalidate events emerges. If "never", a graphical
interface is never invoked. In this case, if the table to be displayed does
not fit in R's console, a temporary file will be saved and the user will be
prompted to open and examine that file. Defaults to "needed".}

\item{n}{A string indicating the overall progress.}
}
\value{
A data frame with the movements for the target tag with
updated 'Valid' column.
}
\description{
Check for movements upstream of the release site.
}
\keyword{internal}
