% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check.R
\name{checkInactiveness}
\alias{checkInactiveness}
\title{Find if a fish is standing still in an array}
\usage{
checkInactiveness(movements, inactive.warning, inactive.error,
  invalid.dist, dist.mat)
}
\arguments{
\item{movements}{the movements list}

\item{inactive.warning}{Numeric value. If a fish spends a number of days greater or equal to this argument in a given place, a warning is issued. Defaults to NULL (i.e. no speed checks are performed, with a message)}

\item{inactive.error}{Numeric value. If a fish spends a number of days greater or equal to this argument in a given place, user intervention is suggested. Defaults to NULL (i.e. no errors are issued)}

\item{invalid.dist}{Whether or not the distance matrix supplied is valid for the study area.}

\item{dist.mat}{A matrix of the distances between the deployed ALS.}

\item{valid.movements}{a temporary valid.movements list}
}
\value{
The valid movements list with valid/invalid rows
}
\description{
Find if a fish is standing still in an array
}
\keyword{internal}
