% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check.R
\name{checkIssue79}
\alias{checkIssue79}
\title{warn users if they are about to run into an unfixed bug.}
\usage{
checkIssue79(arrays, spatial)
}
\value{
No return value, called for side effects.
}
\description{
warn users if they are about to run into an unfixed bug.
}
\keyword{internal}
