% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check.R
\name{checkJumpDistance}
\alias{checkJumpDistance}
\title{Check if fish are jumping over arrays}
\usage{
checkJumpDistance(
  movements,
  fish,
  release,
  dotmat,
  jump.warning = 2,
  jump.error = 3
)
}
\arguments{
\item{movements}{A list of movements for each target tag, created by groupMovements.}

\item{fish}{the fish being analysed}

\item{release}{the release array of the fish being analysed}

\item{dotmat}{A dot distance matrix}

\item{jump.warning}{Integer value. If a fish crosses ´jump.warning´ arrays without being detected, a warning is issued.}

\item{jump.error}{Integer value. If a fish crosses ´jump.error´ arrays without being detected, user intervention is suggested.}
}
\value{
a checked movements list
}
\description{
Check if fish are jumping over arrays
}
\keyword{internal}
