% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check.R
\name{checkLinearity}
\alias{checkLinearity}
\title{Check that the tag linearly moved along the sections}
\usage{
checkLinearity(
  secmoves,
  tag,
  spatial,
  dot_list,
  GUI,
  save.tables.locally,
  n,
  back.warning = "all",
  back.error = "all"
)
}
\arguments{
\item{secmoves}{the section movements list.}

\item{tag}{The tag being analysed.}

\item{spatial}{A list of spatial objects in the study area.}

\item{GUI}{One of "needed", "always" or "never". If "needed", a new window is
opened to inspect the movements only if the movements table is too big to be
displayed in R's console. If "always", a graphical interface is always created
when the possibility to invalidate events emerges. If "never", a graphical
interface is never invoked. In this case, if the table to be displayed does
not fit in R's console, a temporary file will be saved and the user will be
prompted to open and examine that file. Defaults to "needed".}

\item{save.tables.locally}{Logical: If a table must be temporarily stored into a file
for user inspection, should it be saved in the current working directory, or
in R's temporary folder?}

\item{n}{A string indicating the overall progress.}

\item{back.warning}{If a tag moves backwards a number of arrays equal or
greater than \code{back.warning}, a warning is issued. Defaults
to 2. To disable back warnings, set to Inf.
Must be equal to or lower than \code{back.error}.}

\item{back.error}{If a tag moves backwards a number of arrays equal or
greater than \code{back.error}, user intervention is suggested.
Defaults to 3. To disable user intervention suggestions, set to Inf.}
}
\value{
A list of section movements with updated 'Valid' columns
}
\description{
Check that the tag linearly moved along the sections
}
\keyword{internal}
