% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check.R
\name{checkMinimumN}
\alias{checkMinimumN}
\title{Check that the fish have enough detections to be valid}
\usage{
checkMinimumN(movements, minimum.detections, fish)
}
\arguments{
\item{movements}{the movements list}

\item{minimum.detections}{The minimum number of times a tag must have been recorded during the study period for it to be considered a true tag and not random noise.}

\item{fish}{the fish being analysed}
}
\value{
The movements with valid/invalid notes
}
\description{
Check that the fish have enough detections to be valid
}
\keyword{internal}
