% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check.R
\name{checkNoDetections}
\alias{checkNoDetections}
\title{Check if there are detections matching the target tags.}
\usage{
checkNoDetections(input, bio)
}
\arguments{
\item{input}{The list of detections}

\item{bio}{A table with the tags and biometrics of the studied animals.}
}
\value{
No return value, called for side effects.
}
\description{
Check if there are detections matching the target tags.
}
\keyword{internal}
