% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check.R
\name{checkPath}
\alias{checkPath}
\title{Check path validity}
\usage{
checkPath(my.home, path)
}
\arguments{
\item{path}{Path to the folder containing the data. If the R session is already running in the target folder, path may be left as NA}
}
\description{
Confirms that the target directory exists.
}
\keyword{internal}
