% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assembleTimetable.R
\name{checkPrevious}
\alias{checkPrevious}
\title{Find previous valid last event index}
\usage{
checkPrevious(last.events, j)
}
\arguments{
\item{last.events}{A vector of the movement numbers corresponding to each last event. Is supplied by the function findLastEvents.}

\item{j}{the row being analysed}
}
\value{
The previous valid last event index
}
\description{
Find previous valid last event index
}
\keyword{internal}
