% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check.R
\name{checkSectionsUnique}
\alias{checkSectionsUnique}
\title{Check that section names are not duplicated and that sections are not contained
within other section names (e.g. section "N" and section "NW").}
\usage{
checkSectionsUnique(sections)
}
\arguments{
\item{sections}{The sections in which the study area is divided. Must be
coincident with the names given to the arrays. See the vignettes for more
details.}
}
\value{
No return value, called for side effects.
}
\description{
Check that section names are not duplicated and that sections are not contained
within other section names (e.g. section "N" and section "NW").
}
\keyword{internal}
