% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check.R
\name{checkSMovesN}
\alias{checkSMovesN}
\title{Verify number of detections in section movements}
\usage{
checkSMovesN(secmoves)
}
\arguments{
\item{secmoves}{the section movements list}
}
\value{
the section movements with valid/invalid notes
}
\description{
Verify number of detections in section movements
}
\keyword{internal}
