% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check.R
\name{checkSpeeds}
\alias{checkSpeeds}
\title{check fish speeds against defined thresholds (in m/s)}
\usage{
checkSpeeds(movements, fish, valid.movements, speed.warning, speed.error)
}
\arguments{
\item{movements}{The movements table}

\item{fish}{The fish being analysed}

\item{valid.movements}{The valid movements table}

\item{speed.warning}{If a fish moves at a speed equal or greater than
\code{speed.warning} (in metres per second), a warning is issued. If left
NULL (default), no warnings are issued.}

\item{speed.error}{If a fish moves at a speed equal or greater than
\code{speed.error} (in metres per second), user intervention is suggested.
If left NULL (default), user intervention is never suggested.}
}
\value{
The valid movements list with valid/invalid rows
}
\description{
check fish speeds against defined thresholds (in m/s)
}
\keyword{internal}
