% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check.R
\name{checkUpstream}
\alias{checkUpstream}
\title{Check for movements upstream of the release site.}
\usage{
checkUpstream(movements, bio, spatial, arrays)
}
\arguments{
\item{bio}{A table with the tags and biometrics of the studied fish.}
}
\value{
The updated movements
}
\description{
Check for movements upstream of the release site.
}
\keyword{internal}
