% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{circularSection}
\alias{circularSection}
\title{Draw a section on the outside of the circle}
\usage{
circularSection(
  from,
  to,
  units = "hours",
  template = "clock24",
  limits = c(1, 0),
  fill = "white",
  border = "black"
)
}
\arguments{
\item{from}{value where the section should start}

\item{to}{value where the section should end}

\item{units}{units of the from and to variables, defaults to "hours"}

\item{template}{variable to feed into the circular package base functions}

\item{limits}{two values controlling the vertical start and end points of the section}

\item{fill}{The colour of the section}

\item{border}{The colour of the section's border}
}
\value{
No return value, adds to an existing plot.
}
\description{
Draw a section on the outside of the circle
}
\keyword{internal}
