% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CJS.R
\name{cjs_args}
\alias{cjs_args}
\title{CJS.R arguments}
\arguments{
\item{replicates}{A list of arrays containing the respective replicate stations.}

\item{CJS}{The overall CJS results.}

\item{detections.list}{A list of the detections split by each target tag.}

\item{spatial}{The spatial data frame.}

\item{efficiency}{The efficiency results.}

\item{dot_list}{A list containing information on the study area configuration.}

\item{mat, m}{The presence/absence matrices.}

\item{input}{A presence/absence matrix.}

\item{movements, moves}{The movements table.}

\item{status.df}{The main results table.}

\item{estimate}{An estimate of the last array's efficiency, between 0 and 1.}

\item{fixed.efficiency}{A vector of fixed efficiency estimates [0, 1]. \code{length(fixed.efficiency)} must match \code{ncol(input)}.}

\item{silent}{Logical: Should messages be printed? This argument is mainly intended for function calls running within actel's analyses.}
}
\description{
CJS.R arguments
}
\keyword{internal}
