% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assembleEfficiency.R
\name{combineCJS}
\alias{combineCJS}
\title{Split CJS matrix and calculate separate CJS}
\usage{
combineCJS(..., estimate = NULL)
}
\arguments{
\item{...}{The CJS matrixes to be joined}

\item{estimate}{an estimate of the last array's efficiency}
}
\value{
the combined CJS
}
\description{
Split CJS matrix and calculate separate CJS
}
