% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helperFunctions.R
\name{commentCheck}
\alias{commentCheck}
\title{Write in comments}
\usage{
commentCheck(line, tag)
}
\arguments{
\item{line}{The text of the interaction in which the user may or may not request a comment}

\item{tag}{The tag number currently being analysed}
}
\description{
Checks if the user has invoqued the comment command for a specific fish, and stores the comment.
}
\keyword{internal}
