% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load.R
\name{compileDetections}
\alias{compileDetections}
\title{Combine ALS detections}
\usage{
compileDetections(path = "detections", start.time = NULL,
  stop.time = NULL, tz)
}
\arguments{
\item{path}{the path(s) to the detection files}

\item{start.time}{Detection data prior to this date is not analysed. Improves processing time when loading large amounts of detection data.}

\item{stop.time}{Detection data posterior to this date is not analysed. Improves processing time when loading large amounts of detection data.}

\item{tz}{The time-zone of the study area. Necessary to convert the ALS time data, which is in UTC.}
}
\value{
A dataframe with all the detections
}
\description{
Finds the detections' files and processes them.
}
\keyword{internal}
