% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadFunctions.R
\name{convertCodes}
\alias{convertCodes}
\title{Convert codespaces}
\usage{
convertCodes(input)
}
\arguments{
\item{input}{A dataframe of standardized detections.}
}
\value{
A dataframe with standardized code spaces.
}
\description{
Unifies CodeSpace names, to avoid having different names depending on ALS vendor.
}
\keyword{internal}
