% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transform.R
\name{convertTimesToCircular}
\alias{convertTimesToCircular}
\title{Convert times data frame into a list of circular objects}
\usage{
convertTimesToCircular(times)
}
\arguments{
\item{times}{A data frame with the time stamps for reach fish and array}
}
\value{
A list of circular objects
}
\description{
Convert times data frame into a list of circular objects
}
\keyword{internal}
