% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{copyOfRosediagRad}
\alias{copyOfRosediagRad}
\title{circular:::RosediagRad}
\usage{
copyOfRosediagRad(
  x,
  zero,
  rotation,
  bins,
  upper,
  radii.scale,
  prop,
  border,
  col,
  ...
)
}
\arguments{
\item{x}{a vector, matrix or data.frame. The object is coerced to class circular.}

\item{zero}{the zero of the plot. Ignored if plot.info or template are provided.}

\item{rotation}{the rotation of the plot. Ignored if plot.info or template are provided.}

\item{bins}{number of arcs to partition the circle with.}

\item{upper}{logical: if TRUE, the rose diagram cells are "upper"-closed intervals.}

\item{radii.scale}{make possible to choose sector radius form: square-root of relative frequency (sqrt, default) or conventional linear scale (linear).}

\item{prop}{numerical constant determining the radii of the sectors. By default, prop = 1.}

\item{border}{the colour to draw the border. The default, NULL, means to use par("fg"). Use border = NA to omit borders.}

\item{col}{the colour for filling the rose diagram. The default, NULL, is to leave rose diagram unfilled. The values are recycled if needed.}
}
\value{
No return value, adds to an existing plot.
}
\description{
This function is an EXACT copy of the function RosediagRad() in the circular package.
As the function is not exported by the original package, I have copied it here to resolve
the note thrown by devtools::check()
}
\details{
For more details about the original function, visit the circular package homepage at \url{https://github.com/cran/circular}
}
\keyword{internal}
