% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/migration.R
\name{countBackMoves}
\alias{countBackMoves}
\title{Count backwards movements}
\usage{
countBackMoves(movements, arrays)
}
\arguments{
\item{movements}{A list of movements for each target tag, created by groupMovements.}

\item{arrays}{a list of arrays}
}
\value{
The number of backwards movements and the maximum consecutive backwards movements
}
\description{
Count backwards movements
}
\keyword{internal}
