% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assembleTimetable.R
\name{countUpMoves}
\alias{countUpMoves}
\title{Count backwards movements}
\usage{
countUpMoves(movements, spatial)
}
\arguments{
\item{movements}{A list of movements for each target tag, created by groupMovements.}

\item{spatial}{A list of the spatial objects, created by assembleSpatial.}
}
\value{
The number of backwards movements and the maximum consecutive backwards movements
}
\description{
Count backwards movements
}
\keyword{internal}
