% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{createEventRanges}
\alias{createEventRanges}
\title{collapse event indexes into ranges}
\usage{
createEventRanges(x)
}
\arguments{
\item{x}{a numerical vector}
}
\value{
a string of isolated events and event ranges
}
\description{
collapse event indexes into ranges
}
\keyword{internal}
