% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/user_interaction.R
\name{createNewEvents}
\alias{createNewEvents}
\title{Upon invalidating detections, recombines the remaining valid detections
into new events, and merges them with the remaining events.}
\usage{
createNewEvents(displayed.moves, all.moves, detections, event)
}
\arguments{
\item{displayed.moves}{The valid movements table for a specific tag.}

\item{detections}{The detections data.frame for a specific tag.}

\item{event}{The event selected for expansion.}
}
\value{
A data frame containing the movements for the target tag
}
\description{
Upon invalidating detections, recombines the remaining valid detections
into new events, and merges them with the remaining events.
}
\keyword{internal}
