% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load.R
\name{createStandards}
\alias{createStandards}
\title{Standardize serial numbers, stations and arrays in the detections}
\usage{
createStandards(detections, spatial, deployments, discard.orphans = FALSE)
}
\arguments{
\item{detections}{a data frame of detections}

\item{spatial}{A list of spatial objects in the study area}

\item{deployments}{a list of deployments}
}
\value{
A data frame with standardized station names.
}
\description{
Matches the ALS serial number to the deployments to rename the serial number.
The corresponding deployment is then used to update the Standard Station name and the array based in the spatial object.
}
\keyword{internal}
