% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load.R
\name{createUniqueSerials}
\alias{createUniqueSerials}
\title{Include the deployment in the serial number of the receive}
\usage{
createUniqueSerials(input)
}
\arguments{
\item{input}{A data frame with the deployments}
}
\value{
A list of deployments, with unique serial numbers per deployment.
}
\description{
Include the deployment in the serial number of the receive
}
\keyword{internal}
