% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/residency.R
\name{dailyRatios}
\alias{dailyRatios}
\title{calculate daily ratios per fish}
\usage{
dailyRatios(res)
}
\arguments{
\item{res}{a residency list}
}
\value{
A list containing the daily ratios for each fish.
}
\description{
calculate daily ratios per fish
}
\keyword{internal}
