% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/residency.R
\name{dailyRatiosIndOut}
\alias{dailyRatiosIndOut}
\title{compile output of dailyRatios}
\usage{
dailyRatiosIndOut(input)
}
\arguments{
\item{input}{a list containing the output of findSecondsPerSection for each day}
}
\value{
the daily ratios table for the fish
}
\description{
compile output of dailyRatios
}
\keyword{internal}
