% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{darken}
\alias{darken}
\title{darken colours}
\usage{
darken(color, factor = 1.4)
}
\arguments{
\item{color}{The colour to be darkened}

\item{factor}{The level of darkening}
}
\value{
The darker colour code
}
\description{
Copied from https://gist.github.com/Jfortin1/72ef064469d1703c6b30
}
\keyword{internal}
