% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{dataToList}
\alias{dataToList}
\title{Import \code{actel} Results a List}
\usage{
dataToList(source)
}
\arguments{
\item{source}{A RData file containing actel results from a previous run}
}
\description{
Import \code{actel} Results a List
}
