% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helperFunctions.R
\name{decimalTime}
\alias{decimalTime}
\title{Convert hh:mm:ss time to hh.hhh}
\usage{
decimalTime(input)
}
\arguments{
\item{input}{Single string or a vector of strings containing hours:minutes or hours:minutes:seconds}
}
\value{
Decimal hour equivalent (single value or vector)
}
\description{
Wrapper for timeConverter
}
\keyword{internal}
