% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{deleteHelpers}
\alias{deleteHelpers}
\title{Delete temporary files}
\usage{
deleteHelpers(emergency = FALSE)
}
\arguments{
\item{emergency}{Logical: Should temp_debug.txt be kept?}
}
\description{
At the end of the function actel or emergencyBreak, removes temporary files.
}
\keyword{internal}
