% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/z_deprecated.R
\name{deprecated_loadSpatial}
\alias{deprecated_loadSpatial}
\title{Load Spatial file and Check the structure}
\usage{
deprecated_loadSpatial(file)
}
\arguments{
\item{file}{an input file with spatial data.}
}
\value{
The spatial dataframe
}
\description{
Load Spatial file and Check the structure
}
\keyword{internal}
