% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/widget_setups.R
\name{detectionsSingleWidget}
\alias{detectionsSingleWidget}
\title{Detections Widget (Single table version)}
\usage{
detectionsSingleWidget(event, tag, to.print, silent)
}
\arguments{
\item{event}{The event selected for expansion.}

\item{tag}{The tag being analysed.}

\item{to.print}{The subset of detections to be displayed.}
}
\value{
A vector of detection validities.
}
\description{
Detections Widget (Single table version)
}
\keyword{internal}
