% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distanceMatrix.R
\name{distanceMatrix}
\alias{distanceMatrix}
\title{Calculate distance matrix}
\usage{
distanceMatrix(t.layer, starters, targets = starters, EPSGcode,
  coord.x = "x", coord.y = "y", PointIDCol = NA, actel = TRUE)
}
\arguments{
\item{t.layer}{A .RData file containing a transition layer.}

\item{starters}{The points from which to start measuring the distance.}

\item{targets}{The points to which a way must be found.}

\item{EPSGcode}{The EPSG code corresponding to the coordinate system of the imput data. All inputs must be in the same metric system. DO NOT use degree-based coordinates.}

\item{coord.x, coord.y}{The names of the columns containing the x and y information. Must be identical in the starters and targets.}

\item{PointIDCol}{The name of the column containing the IDs of the points to be used as starters and targets. Must be identical in both files.}

\item{actel}{logical: if TRUE, a distance matrix optimized for actel will be saved in the working directory.}
}
\value{
The distance matrix
}
\description{
Adapted from Grant Adams' script "distance to closest mpa". (grant.adams@eagles.usm.edu)
https://cran.r-project.org/web/packages/gdistance/vignettes/gdistance1.pdf
}
