% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{distancesMatrix}
\alias{distancesMatrix}
\title{Calculate Distances Matrix}
\usage{
distancesMatrix(
  t.layer = "transition.layer.RData",
  starters = NULL,
  targets = starters,
  EPSGcode,
  coord.x = "x",
  coord.y = "y",
  id.col = NULL,
  actel = TRUE
)
}
\arguments{
\item{t.layer}{An RData file containing a transition layer.}

\item{starters}{The points from which to start measuring the distance.}

\item{targets}{The points to which a way must be found.}

\item{EPSGcode}{The EPSG code corresponding to the coordinate system of the input data. All inputs must be in the same metric system. DO NOT use degree-based coordinates.}

\item{coord.x, coord.y}{The names of the columns containing the x and y information. Must be identical in the starters and targets.}

\item{id.col}{The name of the column containing the IDs of the points to be used as starters and targets. Must be identical in both files.}

\item{actel}{Logical: Should the distance matrix be optimized for actel and saved in the working directory?}
}
\value{
The distances matrix. If actel = TRUE, the distance matrix is also stored in a 'distances.csv' file.
}
\description{
Using a previously created transition layer (see \code{\link{transitionLayer}}), calculates the distances
between spatial points. Adapted from Grant Adams' script "distance to closest mpa".
}
