% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{divisors}
\alias{divisors}
\title{Calculate the absolute values by which X can be divided}
\usage{
divisors(x)
}
\arguments{
\item{x}{the value to be divided}
}
\value{
a vector of values that, when dividing x, return integers
}
\description{
Calculate the absolute values by which X can be divided
}
