% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transform.R
\name{dotMatrix}
\alias{dotMatrix}
\title{Create numerical distances between dot elements}
\usage{
dotMatrix(input)
}
\arguments{
\item{input}{a dot data frame}
}
\value{
a matrix of distances between arrays
}
\description{
Create numerical distances between dot elements
}
