% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load.R
\name{dotPaths}
\alias{dotPaths}
\title{Find arrays valid for efficiency calculation}
\usage{
dotPaths(input, dotmat, disregard.parallels)
}
\arguments{
\item{input}{A dot list}

\item{dotmat}{A dot distance matrix}

\item{disregard.parallels}{Logical:  Should the presence of parallel arrays
invalidate potential efficiency peers? See the vignettes for more details.}
}
\value{
The input list, with an extra element for each array with valid efficiency peers
}
\description{
Find arrays valid for efficiency calculation
}
\keyword{internal}
