% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CJSFunctions.R
\name{dualArray}
\alias{dualArray}
\title{Calculate estimated last-array efficiency}
\usage{
dualArray(input, silent = TRUE)
}
\arguments{
\item{input}{a detection matrix}

\item{silent}{logical: if TRUE, suppress messages.}
}
\value{
A summary of the CJS results
}
\description{
The calculations are based on 'Using mark-recapture models to estimate survival from telemetry data' by Perry et al. 2012
}
