% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/splitDetections.R
\name{dupSignalsCheck}
\alias{dupSignalsCheck}
\title{Check if there are duplicated signals in the detected tags.}
\usage{
dupSignalsCheck(input, bio, tag.list)
}
\arguments{
\item{input}{list of detections}

\item{bio}{A table with the tags and biometrics of the studied fish.}

\item{tag.list}{list of the target signals}
}
\description{
Check if there are duplicated signals in the detected tags.
}
\keyword{internal}
