% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CJS.R
\name{efficiencyMatrix}
\alias{efficiencyMatrix}
\title{Compile inter-array detection matrix}
\usage{
efficiencyMatrix(movements, dot_list)
}
\arguments{
\item{dot_list}{A list containing information on the study area configuration.}
}
\value{
A matrix of detection histories per tag.
}
\description{
Compile inter-array detection matrix
}
\keyword{internal}
