% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{emergencyBreak}
\alias{emergencyBreak}
\title{Standard procedure when aborting}
\usage{
emergencyBreak()
}
\description{
Saves a copy of the report in a permanent file, where the user decisions taken up to the abort are recorded.
}
\keyword{internal}
