% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{emptyMatrix}
\alias{emptyMatrix}
\title{Create an empty distances matrix}
\usage{
emptyMatrix()
}
\description{
Creates a matrix based on the local 'spatial.csv' file and saves it to 'distances.csv' so the
user can manually fill it.
}
