% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unknownReceiversCheck.R
\name{emptyReceiversCheck}
\alias{emptyReceiversCheck}
\title{Check for unknown receivers in the raw detections}
\usage{
emptyReceiversCheck(spatial, detections)
}
\arguments{
\item{spatial}{A list of the spatial objects, created by assembleSpatial.}
}
\description{
Check for unknown receivers in the raw detections
}
