% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{event}
\alias{event}
\title{Log a new event}
\usage{
event(..., type, tag)
}
\arguments{
\item{...}{The text fragments that compose the event message.}

\item{type}{A vector with one of more of several options:
\itemize{
\item 'crash' special case to be used only together with on.exit(). Meant
to capture function failures not expressly handled by our coded stops.
\item 'stop' appends the message to temp_log.txt then stops the execution.
\item 'warning' appends the message to temp_warnings.txt,
\item 'screen' displays the message on screen,
\item 'report' appends the message to temp_log.txt,
\item 'ud' appends the message to temp_UD.txt,
\item 'comment' appends the message to temp_comments.txt,
\item 'debug' appends the message only to the debug log. NOTE: does not have to be
explicitly called; everything goes into the debug log. Use when you
\emph{only} want the event to be logged in as debug.
}
The same message may be appended to multiple types at once.
Preferred order is "stop", "warning", "screen", "report".}

\item{tag}{Only relevant one 'comment' is one of the types. Put the
transmitter ID here so the comment is attached to the right animal.}
}
\value{
No return value, called for side effects.
}
\description{
Logs useful information as the analysis progresses. May be called to abort
the run if type includes "stop"
}
\keyword{internal}
