% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/migration.R
\name{eventOrderCheck}
\alias{eventOrderCheck}
\title{Check event order}
\usage{
eventOrderCheck(i, last.events, sections, movements)
}
\arguments{
\item{i}{The tag number currently under scrutiny. Only used for messaging purposes.}

\item{sections}{The sections in which the study was divided. Must be coincident with the names given to the ALS arrays. (i.e. if an array is 'River1', then the respective section is 'River')}

\item{movements}{A list of movements for each target tag, created by groupMovements.}
}
\value{
A vector of the movement numbers corresponding to each last event.
}
\description{
Looks for evidence of backwards movements.
}
\keyword{internal}
