% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/widget_setups.R
\name{eventsTabbedWidget}
\alias{eventsTabbedWidget}
\title{Event Widget (Tabbed version)}
\usage{
eventsTabbedWidget(
  tag,
  displayed.moves,
  all.moves,
  detections,
  trigger,
  first.time,
  type
)
}
\arguments{
\item{tag}{The tag being analysed.}

\item{displayed.moves}{The valid movements table for a specific tag.}

\item{all.moves}{The complete movements table for a specific tag.}

\item{detections}{The detections data.frame for a specific tag.}

\item{trigger}{The message/warning that triggered the interaction}

\item{first.time}{Logical: Is this the first time this widget is running for this tag?}

\item{type}{The type of events (Array or Section)}
}
\value{
The movements list, a vector of event validities, and a note on whether or not the widget should be restarted.
}
\description{
Event Widget (Tabbed version)
}
\keyword{internal}
