% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/z_examples.R
\name{example.biometrics}
\alias{example.biometrics}
\title{Example biometric data}
\format{
A data frame with 60 rows and 7 variables:
\describe{
\item{Release.date}{The date and time of release}
\item{Release.site}{The release site}
\item{Serial.nr}{The serial number of the implanted tag}
\item{Signal}{The Signal emitted by the implanted tag}
\item{Group}{The group of the tagged fish}
\item{Total.Length.mm}{The length of the tagged fish}
\item{Mass.g}{The mass of the tagged fish}
}
}
\source{
Data collected by the authors.
}
\description{
A dataset containing the tag numbers and respective biometrics of 60 fish.
}
\keyword{internal}
