% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/z_examples.R
\name{example.deployments}
\alias{example.deployments}
\title{Example deployment data}
\format{
A data frame with 17 rows and 4 variables:
\describe{
\item{Receiver}{The receiver serial number}
\item{Station.name}{The name given to the receiver}
\item{Start}{The date and time of the deployment}
\item{Stop}{The date and time of the retrieval}
}
}
\source{
Data collected by the authors.
}
\description{
A dataset containing the deployed receivers.
}
\keyword{internal}
