% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/z_examples.R
\docType{data}
\name{example.distances}
\alias{example.distances}
\title{Example distances matrix}
\format{
A 18 * 18 matrix
}
\source{
Data collected by the authors.
}
\usage{
example.distances
}
\description{
A matrix containing the distances between stations and release sites
}
\keyword{internal}
