% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/z_examples.R
\docType{data}
\name{example.spatial}
\alias{example.spatial}
\title{Example spatial data}
\format{
A data frame with 18 rows and 6 variables:
\describe{
\item{Station.name}{The name of the ALS or release site}
\item{Latitude}{The latitude of the ALS or release site in WGS84}
\item{Longitude}{The longitude of the ALS or release site in WGS84}
\item{x}{The x coordinate of the ALS or release site in EPSG 32632}
\item{y}{The y coordinate of the ALS or release site in EPSG 32632}
\item{Array}{The Array to which the ALS belongs, or the first ALS array downstream of the release site.}
\item{Type}{The type of spatial object (must be either Hydrophone or Release)}
}
}
\source{
Data collected by the authors.
}
\usage{
example.spatial
}
\description{
A dataset containing the positions of the deployed ALS and release site.
}
\keyword{internal}
