% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/splitDetections.R
\name{excludeTags}
\alias{excludeTags}
\title{Collect summary information on the tags detected but that are not part of the study.}
\usage{
excludeTags(input, exclude.tags, silent)
}
\arguments{
\item{input}{list of detections}

\item{exclude.tags}{A list of tags that should be excluded from the detection data before any analyses are performed. Intended to be used if stray tags from a different code space but with the same signal as a target tag are detected in the study area.}

\item{silent}{logical: if TRUE, suppress messages.}
}
\description{
Collect summary information on the tags detected but that are not part of the study.
}
\keyword{internal}
