% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{expFirstArray}
\alias{expFirstArray}
\title{extract the string of expected first arrays for a tag}
\usage{
expFirstArray(tag, bio, spatial)
}
\arguments{
\item{tag}{the tag number to check}

\item{bio}{the biometrics table}

\item{spatial}{the spatial list}
}
\value{
a string of arrays where the animal should have been seen first.
}
\description{
extract the string of expected first arrays for a tag
}
\keyword{internal}
