% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/explore.R
\name{explore}
\alias{explore}
\title{Actel: Acoustic telemetry data sorting}
\usage{
explore(
  path = NULL,
  max.interval = 60,
  maximum.time = 60,
  speed.method = c("last to first", "first to first"),
  speed.warning = NULL,
  speed.error = NULL,
  tz.study.area = NULL,
  tz,
  start.time = NULL,
  start.timestamp = NULL,
  stop.time = NULL,
  end.timestamp = NULL,
  override = NULL,
  minimum.detections = 2,
  report = TRUE,
  exclude.tags = NULL,
  jump.warning = 2,
  jump.error = 3,
  inactive.warning = NULL,
  inactive.error = NULL,
  debug = FALSE
)
}
\arguments{
\item{path}{Path to the folder containing the data. If the R session is already running in the target folder, path may be left as NULL}

\item{max.interval}{The number of minutes that must pass between detections for a new event to be created.}

\item{maximum.time}{Deprecated. See "max.interval"}

\item{speed.method}{One of 'last to first' or 'first to first'. In the former, the last detection on a given array/section is matched to the first detection on the next array/section (default). If changed to 'first to first', the first detection on two consecutive arrays/sections are used to perform the calculations.}

\item{speed.warning}{Numeric value (m/s). If a fish moves at a speed greater or equal to this argument, a warning is issued. Defaults to NULL (i.e. no speed checks are performed, with a message)}

\item{speed.error}{Numeric value (m/s). If a fish moves at a speed greater or equal to this argument, user intervention is suggested. Defaults to NULL (i.e. no errors are issued)}

\item{tz.study.area}{Deprecated. See "tz"}

\item{tz}{The time-zone of the study area. Necessary to convert the ALS time data, which is in UTC.}

\item{start.time}{Detection data prior to this date is not analysed. Improves processing time when loading large amounts of detection data.}

\item{start.timestamp}{Deprecated. See "start.time"}

\item{stop.time}{Detection data posterior to this date is not analysed. Improves processing time when loading large amounts of detection data.}

\item{end.timestamp}{Deprecated. See "stop.time"}

\item{override}{A list of tags for which the user intends to manually define entering and leaving points for each study section.}

\item{minimum.detections}{The minimum number of times a tag must have been recorded during the study period for it to be considered a true tag and not random noise.}

\item{report}{Whether graphics, tables and html report files should be created. Defaults to TRUE. Allows automatic compiling of an html report after the analysis.}

\item{exclude.tags}{A list of tags that should be excluded from the detection data before any analyses are performed. Intended to be used if stray tags from a different code space but with the same signal as a target tag are detected in the study area.}

\item{jump.warning}{Integer value. If a fish crosses ´jump.warning´ arrays without being detected, a warning is issued.}

\item{jump.error}{Integer value. If a fish crosses ´jump.error´ arrays without being detected, user intervention is suggested.}

\item{inactive.warning}{Numeric value (days). If a fish spends a number of days greater or equal to this argument in a given place, a warning is issued. Defaults to NULL (i.e. no speed checks are performed, with a message)}

\item{inactive.error}{Numeric value (days). If a fish spends a number of days greater or equal to this argument in a given place, user intervention is suggested. Defaults to NULL (i.e. no errors are issued)}

\item{debug}{If TRUE, temporary files are not deleted at the end of the analysis. Defaults to FALSE.}
}
\value{
A list containing 1) the detections used during the analysis, 2) the movement events, 3) the status dataframe, 4) the survival overview per group, 5) the progression through the study area, 6) the ALS array/sections' efficiency, 7) the list of spatial objects used during the analysis.
}
\description{
The actel package provides a systematic way of analysing fish migration data.
Its main function, actel, collects the input present in the target folder and analyses the telemetry data.
It is strongly recommended to read the package vignettes before attempting to run the analyses. You can find the vignettes by running browseVignettes('actel') .
}
