% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{extractSignals}
\alias{extractSignals}
\title{Extract signals from transmitter names}
\usage{
extractSignals(input)
}
\arguments{
\item{input}{A vector of transmitter names}
}
\value{
A vector of transmitter signals
}
\description{
Extract signals from transmitter names
}
\examples{
# create dummy string
x <- c("R64K-1234", "A69-1303-12")

# run extractSignals
extractSignals(x)

}
