% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{fact}
\alias{fact}
\title{nearsq helper}
\usage{
fact(n)
}
\description{
Obtained here: https://stackoverflow.com/questions/32017327/calculate-the-optimal-grid-layout-dimensions-for-a-given-amount-of-plots-in-r
}
\keyword{internal}
