% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load.R
\name{findDirectChains}
\alias{findDirectChains}
\title{Find all arrays linked to an array in a given direction}
\usage{
findDirectChains(input, dotmat, type = c("before", "after"))
}
\arguments{
\item{input}{An array list}

\item{type}{The direction in which to expand the chain ("before" or "after")}
}
\value{
The array list with all linked arrays.
}
\description{
Find all arrays linked to an array in a given direction
}
\keyword{internal}
