% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadFunctions.R
\name{findFiles}
\alias{findFiles}
\title{Find file names}
\usage{
findFiles(path = NULL, pattern = NULL)
}
\arguments{
\item{path}{the path(s) to the detection files}
}
\value{
A vector of the file names.
}
\description{
Find file names
}
\keyword{internal}
