% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/migration.R
\name{findFirstEvents}
\alias{findFirstEvents}
\title{Find first events}
\usage{
findFirstEvents(i, last.events, movements, sections, cautious.assignment)
}
\arguments{
\item{i}{The tag number currently under scrutiny. Only used for messaging purposes.}

\item{last.events}{A vector of the movement numbers corresponding to each last event. Is supplied by the function findLastEvents.}

\item{movements}{A list of movements for each target tag, created by groupMovements.}

\item{sections}{The sections in which the study was divided. Must be coincident with the names given to the ALS arrays. (i.e. if an array is 'River1', then the respective section is 'River')}

\item{cautious.assignment}{If TRUE, actel avoids assigning events with one detection as first and/or last events of a section.}
}
\value{
A list containing both the first and last events.
}
\description{
Finds the first event for each section in the study area that has a last event.
}
\keyword{internal}
