% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/migration.R
\name{findFirstRow}
\alias{findFirstRow}
\title{Find starting row}
\usage{
findFirstRow(l, last.events, movements, sections)
}
\arguments{
\item{l}{The section currently being analysed. Supplied by findFirstEvents}

\item{last.events}{A vector of the movement numbers corresponding to each last event. Is supplied by the function findLastEvents.}

\item{movements}{A list of movements for each target tag, created by groupMovements.}

\item{sections}{The sections in which the study was divided. Must be coincident with the names given to the ALS arrays. (i.e. if an array is 'River1', then the respective section is 'River')}
}
\value{
The row number
}
\description{
Finds the row from which the function findFirstEvents should start looking for the first event.
}
\keyword{internal}
