% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assembleTimetable.R
\name{findLastEvents}
\alias{findLastEvents}
\title{Find last events}
\usage{
findLastEvents(i, movements, sections, minimum.detections, processing.type,
  cautious.assignment)
}
\arguments{
\item{i}{The tag number currently under scrutiny. Only used for messaging purposes.}

\item{movements}{A list of movements for each target tag, created by groupMovements.}

\item{sections}{The sections in which the study was divided. Must be coincident with the names given to the ALS arrays. (i.e. if an array is 'River1', then the respective section is 'River')}

\item{minimum.detections}{The mininum number of times a tag must have been recorded during the study period for it to be considered a true tag and not random noise.}

\item{processing.type}{the current processing type for the fish. Is supplied by the function assembleTimetable.}

\item{cautious.assignment}{If TRUE, actel avoids assigning events with one detection as first and/or last events of a section.}
}
\value{
A vector of the movement numbers corresponding to each last event.
}
\description{
Finds the last event for each section in the study area.
}
\keyword{internal}
