% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load.R
\name{findPeers}
\alias{findPeers}
\title{Find efficiency peers for each array}
\usage{
findPeers(input, dotmat, type = c("before", "after"),
  disregard.parallels)
}
\arguments{
\item{input}{An array list}

\item{type}{The type of peers to be found ("before" or "after")}
}
\value{
The array list with efficiency peers
}
\description{
Find efficiency peers for each array
}
\keyword{internal}
