% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load.R
\name{findShortestChains}
\alias{findShortestChains}
\title{Find the shortest paths between arrays}
\usage{
findShortestChains(input)
}
\arguments{
\item{input}{An array list}
}
\value{
The array list with all paths between arrays with distance > 1.
}
\description{
Find the shortest paths between arrays
}
\keyword{internal}
