% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/residency.R
\name{firstArrayFailure}
\alias{firstArrayFailure}
\title{Determine if the first array after release has failed}
\usage{
firstArrayFailure(tag, bio, spatial, first.array, dot_list)
}
\arguments{
\item{tag}{The tag being analysed}

\item{bio}{A table with the tags and biometrics of the studied animals.}

\item{spatial}{A list of spatial objects in the study area}

\item{first.array}{The array of the first valid movement}
}
\value{
NULL if no arrays failed, or a list of arrays that failed.
}
\description{
Determine if the first array after release has failed
}
\keyword{internal}
