% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/printFunctions.R
\name{folderCheck}
\alias{folderCheck}
\title{Check report compatibility}
\usage{
folderCheck(report, redraw)
}
\value{
A TRUE/FALSE decision
}
\description{
Creates a "Report" folder if necessary and silently activates ggplot2 and reshape2 to avoid startup messages
}
\keyword{internal}
