% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CJSFunctions.R
\name{getEstimate}
\alias{getEstimate}
\title{Get efficiency estimate for last array}
\usage{
getEstimate(spatial, detections.list, replicate)
}
\arguments{
\item{spatial}{A list of the spatial objects, created by assembleSpatial.}

\item{detections.list}{A list of the detections split by each target tag, created by splitDetections.}
}
\value{
the modified CJS model for the last array
}
\description{
Get efficiency estimate for last array
}
\keyword{internal}
