% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CJSFunctions.R
\name{getGroupCJS}
\alias{getGroupCJS}
\title{Calculate CJS for each group}
\usage{
getGroupCJS(the.matrices, status.df, fixed.efficiency = NULL)
}
\arguments{
\item{the.matrices}{A list of detection matrices}

\item{status.df}{A dataframe with all the final data for each fish, created by assembleOutput.}

\item{fixed.efficiency}{A vector of fixed efficiency estimates from a more complete CJS model.}
}
\value{
A list of CJS results
}
\description{
Calculate CJS for each group
}
\keyword{internal}
