% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CJSFunctions.R
\name{getReleaseCJS}
\alias{getReleaseCJS}
\title{Calculate CJS per release site}
\usage{
getReleaseCJS(the.matrices, status.df)
}
\arguments{
\item{the.matrices}{A list of detection matrices}

\item{status.df}{A dataframe with all the final data for each fish, created by assembleOutput.}
}
\value{
A list of CJS results
}
\description{
Calculate CJS per release site
}
\keyword{internal}
