% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/residency.R
\name{getResidency}
\alias{getResidency}
\title{Assemble residency tables per fish}
\usage{
getResidency(movements, spatial)
}
\arguments{
\item{movements}{the movements list}

\item{spatial}{A list of spatial objects in the study area}
}
\value{
a list of residency tables
}
\description{
Assemble residency tables per fish
}
\keyword{internal}
