% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.R
\name{getSpeeds}
\alias{getSpeeds}
\title{Extrat speeds from the analysis results}
\usage{
getSpeeds(
  input,
  type = c("all", "forward", "backward"),
  only.direct = FALSE,
  n.events = c("first", "all", "last")
)
}
\arguments{
\item{input}{An actel results object generated by \code{\link{explore}}, \code{\link{migration}} or \code{\link{residency}}.}

\item{type}{The type of movements to record. One of "all", "forward", or "backward". In the two last options,
only the forward or backwards (relatively to the study area structure) movement speeds are returned.}

\item{only.direct}{only extract speeds between arrays that are directly connected}

\item{n.events}{The events to record. One of "first", "all", or "last".}
}
\value{
A data frame with the following columns:
\itemize{
\item Fish: The tag of the fish who performed the recorded speed
\item Event: The valid event where the speed was recorded
\item From.array: The array from which the fish left
\item From.station: The statino from which the fish left
\item To.array: The array to which the fish arrived
\item To.station: The station to which the fish arrived
\item Speed: The speed recorded in the described movement
}
}
\description{
Extrat speeds from the analysis results
}
\examples{
# using the example results loaded with actel
getSpeeds(example.results)

# You can specify which direction of movement to extract with 'type'
getSpeeds(example.results, type = "forward")
# or
getSpeeds(example.results, type = "backward")

# and also how many events per fish (this won't change the output 
# with the example.results, only because these results are minimal).
getSpeeds(example.results, n.events = "first")
# or
getSpeeds(example.results, n.events = "all")
# or
getSpeeds(example.results, n.events = "last")

}
