% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.R
\name{getSpeedsChecker}
\alias{getSpeedsChecker}
\title{Helper function of getSpeeds}
\usage{
getSpeedsChecker(
  rows,
  moves,
  first,
  dot_list,
  side = c("neighbours", "all.before", "all.after")
)
}
\arguments{
\item{rows}{a vector of rows to check.}

\item{moves}{the respective movements table.}

\item{first}{a list of valid arrays for the first movement.}

\item{dot_list}{the list with the study area structure.}

\item{side}{which pool of array-relatives to check against.}
}
\value{
an updated vector of rows to consider for extraction.
}
\description{
Checks if each row should be considered for extraction or not.
}
\keyword{internal}
