% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.R
\name{getSpeedsCompiler}
\alias{getSpeedsCompiler}
\title{Helper function of getSpeeds}
\usage{
getSpeedsCompiler(rows, moves, tag, release, speed.method, n.events)
}
\arguments{
\item{rows}{a vector of rows to check.}

\item{moves}{the respective movements table.}

\item{tag}{the transmitter code being analysed.}

\item{release}{the release site of the tag.}

\item{n.events}{The events to record. One of "first", "all", or "last".}
}
\value{
a data.frame with the speeds for the target tag
}
\description{
Compiles the final output table for each tag
}
\keyword{internal}
