% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CJSFunctions.R
\name{getSplitCJS}
\alias{getSplitCJS}
\title{Calculate CJS for each group.release combination}
\usage{
getSplitCJS(the.matrices, fixed.efficiency = NULL)
}
\arguments{
\item{the.matrices}{A list of detection matrices}

\item{fixed.efficiency}{A vector of fixed efficiency estimates from a more complete CJS model.}
}
\value{
A list of CJS results
}
\description{
Calculate CJS for each group.release combination
}
\keyword{internal}
