% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/circularFunctions.R
\name{getTimes}
\alias{getTimes}
\title{Extract timestamp of valid entry or exit in each array}
\usage{
getTimes(simple.movements, spatial, tz.study.area, type = c("Arrival",
  "Departure"))
}
\arguments{
\item{simple.movements}{A list of valid-only movements for each fish}

\item{tz.study.area}{The time-zone of the study area. Necessary to convert the ALS time data, which is in UTC.}

\item{type}{The point to be recorded: one of "Arrival" or "Departure".}
}
\value{
A data frame with the timestamps for each fish (rows) and array (columns)
}
\description{
Extract timestamp of valid entry or exit in each array
}
\keyword{internal}
