% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{gg_colour_hue}
\alias{gg_colour_hue}
\title{Generate default ggplot colours}
\usage{
gg_colour_hue(n)
}
\arguments{
\item{n}{The number of colours to be generated}
}
\value{
a vector of colours with the same length as n
}
\description{
Copied from: \url{https://stackoverflow.com/questions/8197559/emulate-ggplot2-default-color-palette}.
}
\keyword{internal}
