% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/residency.R
\name{globalRatios}
\alias{globalRatios}
\title{Calculate number/percentage of fish at each location for each day}
\usage{
globalRatios(positions)
}
\arguments{
\item{positions}{a positions table, supplied by dailyPositions}
}
\value{
A list with 1) a table containing the absolute number of fish at each location per day,
and 2) the respective percentage table.
}
\description{
Calculate number/percentage of fish at each location for each day
}
\keyword{internal}
