% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/user_interaction.R
\name{graphicalInvalidate_detections}
\alias{graphicalInvalidate_detections}
\title{Opens a new window that allows the user to determine movement event invalidity}
\usage{
graphicalInvalidate_detections(
  dets,
  displayed.moves,
  all.moves,
  event,
  tag,
  silent = FALSE
)
}
\arguments{
\item{tag}{The tag being analysed}

\item{trigger}{The warning/message that triggered the interaction}
}
\value{
A data frame with the movement events for the target tag and an updated 'Valid' column.
}
\description{
Opens a new window that allows the user to determine movement event invalidity
}
\keyword{internal}
