% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/user_interaction.R
\name{graphicalInvalidateDetections}
\alias{graphicalInvalidateDetections}
\title{Opens a new window that allows the user to determine detection invalidity}
\usage{
graphicalInvalidateDetections(
  detections,
  displayed.moves,
  all.moves,
  event,
  tag,
  silent = FALSE
)
}
\arguments{
\item{detections}{The detections data.frame for a specific tag.}

\item{displayed.moves}{The valid movements table for a specific tag.}

\item{event}{The event selected for expansion.}

\item{tag}{The tag being analysed.}
}
\value{
A data frame with the movement events for the target tag and an updated 'Valid' column.
}
\description{
Opens a new window that allows the user to determine detection invalidity
}
\keyword{internal}
