% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/groupMovements.R
\name{groupMovements}
\alias{groupMovements}
\title{Group movements}
\usage{
groupMovements(detections.list, bio, spatial, speed.method, maximum.time,
  tz.study.area, dist.mat, invalid.dist)
}
\arguments{
\item{detections.list}{A list of the detections split by each target tag, created by splitDetections.}

\item{bio}{A table with the tags and biometrics of the studied fish.}

\item{spatial}{A list of the spatial objects, created by assembleSpatial.}

\item{speed.method}{One of 'last to first' or 'first to first'. In the former, the last detection on a given array/section is matched to the first detection on the next array/section (default). If changed to 'first to first', the first detection on two consecutive arrays/sections are used to perform the calculations.}

\item{maximum.time}{The number of minutes that must pass between detections for a new event to be created.}

\item{tz.study.area}{The timezone of the study area. Necessary to conver the ALS time data, which is in UTC.}

\item{dist.mat}{A matrix of the distances between the deployed ALS.}

\item{invalid.dist}{Wether or not the distance matrix suplied is valid for the study area.}
}
\value{
A list of movement events for each fish.
}
\description{
Crawls trough the detections of each fish and groups them based on ALS arrays and time requirements.
}
\keyword{internal}
