% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CJS.R
\name{includeIntraArrayEstimates}
\alias{includeIntraArrayEstimates}
\title{Incorporate intra-array estimates in the overall CJS object}
\usage{
includeIntraArrayEstimates(m, CJS)
}
\arguments{
\item{m}{A list of dual matrices}

\item{CJS}{The overall CJS results}
}
\value{
A list containing the updated overall CJS and the intra-array CJS results
}
\description{
Incorporate intra-array estimates in the overall CJS object
}
\keyword{internal}
