% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assembleEfficiency.R
\name{includeMissing}
\alias{includeMissing}
\title{Include fish that were never detected}
\usage{
includeMissing(x, status.df)
}
\arguments{
\item{x}{an efficiency matrix}

\item{status.df}{A dataframe with all the final data for each fish, created by assembleOutput.}
}
\value{
a matrix of detection histories per fish, including fish that were never detected.
}
\description{
Include fish that were never detected
}
\keyword{internal}
