% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check.R
\name{includeUnknownReceiver}
\alias{includeUnknownReceiver}
\title{Temporarily include missing receivers in the spatial object}
\usage{
includeUnknownReceiver(spatial, deployments, unknown.receivers)
}
\arguments{
\item{spatial}{The spatial data frame, as loaded by loadSpatial}

\item{unknown.receivers}{serial number of the receivers to be included}
}
\description{
Temporarily include missing receivers in the spatial object
}
\keyword{internal}
