% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assembleEfficiency.R
\name{interArray}
\alias{interArray}
\title{Calculate inter-array efficiency}
\usage{
interArray(spatial, movements, minimum.detections)
}
\arguments{
\item{spatial}{A list of the spatial objects, created by assembleSpatial.}

\item{movements}{A list of movements for each target tag, created by groupMovements.}

\item{minimum.detections}{The mininum number of times a tag must have been recorded during the study period for it to be considered a true tag and not random noise.}
}
\value{
a table of inter-array efficiency estimates.
}
\description{
Calculate inter-array efficiency
}
\keyword{internal}
