% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assembleEfficiency.R
\name{interSection}
\alias{interSection}
\title{Calculate inter-section efficiency}
\usage{
interSection(sections, movements, minimum.detections)
}
\arguments{
\item{sections}{The sections in which the study was divided. Must be coincident with the names given to the ALS arrays. (i.e. if an array is 'River1', then the respective section is 'River')}

\item{movements}{A list of movements for each target tag, created by groupMovements.}

\item{minimum.detections}{The mininum number of times a tag must have been recorded during the study period for it to be considered a true tag and not random noise.}
}
\value{
a table of inter-section efficiency estimates.
}
\description{
Calculate inter-section efficiency
}
\keyword{internal}
