% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assembleEfficiency.R
\name{intraArray}
\alias{intraArray}
\title{Calculate intra-array efficiency}
\usage{
intraArray(spatial, detections.list)
}
\arguments{
\item{spatial}{A list of the spatial objects, created by assembleSpatial.}

\item{detections.list}{A list of the detections split by each target tag, created by splitDetections.}
}
\value{
a table of intra-array efficiency estimates.
}
\description{
Calculate intra-array efficiency
}
\keyword{internal}
