% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/user_interaction.R
\name{invalidateDetections}
\alias{invalidateDetections}
\title{Allow the user to determine a given set of detections invalid}
\usage{
invalidateDetections(displayed.moves, all.moves, detections, tag, event)
}
\arguments{
\item{displayed.moves}{The valid movements table for a specific tag.}

\item{detections}{The detections data.frame for a specific tag.}

\item{tag}{The tag being analysed.}

\item{event}{The event selected for expansion.}
}
\value{
A data frame with the movement events for the target tag and an updated 'Valid' column.
}
\description{
Allow the user to determine a given set of detections invalid
}
\keyword{internal}
