% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check.R
\name{invalidateEvents}
\alias{invalidateEvents}
\title{Allow the user to determine a given movement event invalid}
\usage{
invalidateEvents(movements, tag)
}
\arguments{
\item{movements}{The movements table}

\item{tag}{The tag being analysed}
}
\value{
A data frame with the movement events for the target tag and an updated 'Valid' column.
}
\description{
Allow the user to determine a given movement event invalid
}
\keyword{internal}
