% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load.R
\name{labelUnknowns}
\alias{labelUnknowns}
\title{Temporarily include unknown receivers with target detections in the study area}
\usage{
labelUnknowns(detections.list)
}
\arguments{
\item{detections.list}{a list of the detections for each target tag}
}
\value{
The detections.list with the unknown receivers labelled
}
\description{
Temporarily include unknown receivers with target detections in the study area
}
\keyword{internal}
