% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load.R
\name{liveArrayTimes}
\alias{liveArrayTimes}
\title{Assign live times to arrays}
\usage{
liveArrayTimes(arrays, deployments, spatial)
}
\arguments{
\item{arrays}{The array list}

\item{deployments}{the deployments list}

\item{spatial}{The spatial list}
}
\value{
an expanded array list
}
\description{
Assign live times to arrays
}
\keyword{internal}
