% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load.R
\name{loadDetections}
\alias{loadDetections}
\title{Load ALS detections}
\usage{
loadDetections(start.time = NULL, stop.time = NULL, tz, force = FALSE)
}
\arguments{
\item{start.time}{Detection data prior to this date is not analysed. Improves processing time when loading large amounts of detection data.}

\item{stop.time}{Detection data posterior to this date is not analysed. Improves processing time when loading large amounts of detection data.}

\item{tz}{The time-zone of the study area. Necessary to convert the ALS time data, which is in UTC.}
}
\value{
A dataframe with all the detections
}
\description{
If there are previously compiled detections present, offers the chance to reuse. Otherwise triggers combineDetections.
}
\keyword{internal}
