% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load.R
\name{loadDistances}
\alias{loadDistances}
\title{Load distances matrix}
\usage{
loadDistances(spatial)
}
\arguments{
\item{spatial}{A list of spatial objects in the study area}
}
\value{
a list containing the distances matrix and a TRUE/FALSE value indicating whether or not that distances matrix is valid for the target study area.
}
\description{
Load distances matrix
}
\keyword{internal}
