% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load.R
\name{loadDot}
\alias{loadDot}
\title{Load spatial.dot}
\usage{
loadDot(
  string = NULL,
  input = NULL,
  spatial,
  sections = NULL,
  disregard.parallels
)
}
\arguments{
\item{string}{A dot string}

\item{input}{The name of the dot file}

\item{spatial}{A spatial data frame}

\item{sections}{The sections in which the study area is divided. Must be
coincident with the names given to the arrays. See the vignettes for more
details.}

\item{disregard.parallels}{Logical:  Should the presence of parallel arrays
invalidate potential efficiency peers? See the vignettes for more details.}
}
\value{
The dot list
}
\description{
Load spatial.dot
}
\keyword{internal}
