% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load.R
\name{loadStudyData}
\alias{loadStudyData}
\title{Load, structure and check the inputs}
\usage{
loadStudyData(
  tz,
  override = NULL,
  start.time,
  stop.time,
  sections = NULL,
  exclude.tags,
  disregard.parallels = TRUE
)
}
\arguments{
\item{tz}{The time zone of the study area. Must match one of the values
present in \code{\link[base]{timezones}}.}

\item{override}{A vector of tags for which the user intends to manually
define which movement events are valid and invalid.}

\item{start.time}{Detection data prior to the timestamp set in
\code{start.time} (in YYYY-MM-DD HH:MM:SS format) is not considered during
the analysis.}

\item{stop.time}{Detection data posterior to the timestamp set in
\code{stop.time} (in YYYY-MM-DD HH:MM:SS format) is not considered during
the analysis.}

\item{sections}{The sections in which the study area is divided. Must be
coincident with the names given to the arrays. See the vignettes for more
details.}

\item{exclude.tags}{A vector of tags that should be excluded from the
detection data before any analyses are performed. Intended to be used if
stray tags from a different code space but with the same signal as a target
tag are detected in the study area.}

\item{disregard.parallels}{Logical:  Should the presence of parallel arrays
invalidate potential efficiency peers? See the vignettes for more details.}
}
\value{
a list of objects to be used for data analysis
}
\description{
Load, structure and check the inputs
}
\keyword{internal}
