% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{lowestSignal}
\alias{lowestSignal}
\title{cleanly extract lowest signal from signals string}
\usage{
lowestSignal(x)
}
\arguments{
\item{x}{}
}
\value{
a vector of lowest signals
}
\description{
cleanly extract lowest signal from signals string
}
\keyword{internal}
