% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{match.POSIXt}
\alias{match.POSIXt}
\title{Match POSIX values}
\usage{
match.POSIXt(this, there)
}
\arguments{
\item{this}{the vector of posix to be match}

\item{there}{the vector of posix to be matched against}
}
\value{
a vector with the matches
}
\description{
Match POSIX values
}
\keyword{internal}
