% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{matchl}
\alias{matchl}
\title{TRUE/FALSE wrapper for match}
\usage{
matchl(input, match)
}
\arguments{
\item{input}{vector of elements to be matched.}

\item{match}{vector of elements where to look for the input.}
}
\value{
A logical vector.
}
\description{
Looks for matches of a list on a target, but returns TRUE/FALSE instead of the position where the match was found.
}
\keyword{internal}
