% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{mbPrintEfficiency}
\alias{mbPrintEfficiency}
\title{Print efficiency fragment}
\usage{
mbPrintEfficiency(overall.CJS, intra.CJS)
}
\arguments{
\item{overall.CJS}{a single CJS with all the groups and release sites merged}

\item{intra.CJS}{The output of the getEstimate calculations.}
}
\description{
Prints the ALS inter-array efficiency for inclusion in printRmd.
}
\keyword{internal}
