% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{mbPrintProgression}
\alias{mbPrintProgression}
\title{Print progression diagram}
\usage{
mbPrintProgression(dot, sections, overall.CJS, spatial, status.df)
}
\arguments{
\item{dot}{a dot data frame}

\item{sections}{The sections in which the study was divided. Must be coincident with the names given to the ALS arrays. (i.e. if an array is 'River1', then the respective section is 'River')}

\item{overall.CJS}{a single CJS with all the groups and release sites merged}

\item{spatial}{A list of spatial objects in the study area}

\item{status.df}{A data frame with the final migration results}
}
\description{
Print progression diagram
}
\keyword{internal}
