% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CJS.R
\name{mbSplitCJS}
\alias{mbSplitCJS}
\title{Calculate CJS for each group.release combination}
\usage{
mbSplitCJS(m, fixed.efficiency = NULL)
}
\arguments{
\item{m}{The presence/absence matrices.}

\item{fixed.efficiency}{A vector of fixed efficiency estimates from a more complete CJS model.}
}
\value{
A list of CJS results
}
\description{
Calculate CJS for each group.release combination
}
\keyword{internal}
