% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{minuteTime}
\alias{minuteTime}
\title{Convert numeric time to hh:mm:ss}
\usage{
minuteTime(x, format = c("h", "m", "s"), seconds = TRUE)
}
\arguments{
\item{input}{Single string or a vector of strings containing hours:minutes or hours:minutes:seconds}
}
\value{
Decimal hour equivalent (single value or vector)
}
\description{
Convert numeric time to hh:mm:ss
}
