% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/movements.R
\name{move_args}
\alias{move_args}
\title{movements.R arguments}
\arguments{
\item{arrmoves}{the array movements list}

\item{secmoves}{the section movements list}

\item{bio}{The biometrics table.}

\item{detections.list}{A list of the detections split by each target tag.}

\item{dist.mat}{A matrix of the distances between the deployed ALS.}

\item{spatial}{The spatial list.}

\item{arrays}{A list containing information for each array.}

\item{movements}{the movement dataframe for the tag being analysed.}

\item{tag}{The tag ID of the animal currently being analysed}

\item{type}{The type of movements being analysed. One of "array" or "section".}

\item{valid.dist}{Logical: Is a valid distances matrix being used?}
}
\description{
movements.R arguments
}
\keyword{internal}
