% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{moveHelpers}
\alias{moveHelpers}
\title{Move helper files to new directory}
\usage{
moveHelpers(my.home)
}
\arguments{
\item{my.home}{The working directory where the main function was triggered.}
}
\description{
Move helper files to new directory
}
\keyword{internal}
