% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/movements.R
\name{movementTimes}
\alias{movementTimes}
\title{Calculate movement times}
\usage{
movementTimes(movements, silent = TRUE)
}
\arguments{
\item{movements}{A list of movements for each target tag, created by groupMovements.}

\item{silent}{logical: If TRUE, debug messages are issued (only works within actel)}
}
\value{
The movement data frame with time and speed calculations
}
\description{
Determines the duration of an event and the time from an event to the next.
}
