% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{na.as.false}
\alias{na.as.false}
\title{Consider NA's as FALSE}
\usage{
na.as.false(input)
}
\arguments{
\item{input}{vector containing NA's.}
}
\value{
A logical vector.
}
\description{
Aimed to be used in a vector of TRUE/FALSE's, where NA's are present and should be considered as false.
}
\keyword{internal}
