% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{nearsq}
\alias{nearsq}
\title{find optimum plotting grid}
\usage{
nearsq(n, tol = 5/3 + 0.001)
}
\arguments{
\item{n}{number of plots to fit.}
}
\description{
obtained here: https://stackoverflow.com/questions/32017327/calculate-the-optimal-grid-layout-dimensions-for-a-given-amount-of-plots-in-r
}
\keyword{internal}
