% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unknownReceiversCheck.R
\name{new_includeUnknownReceiver}
\alias{new_includeUnknownReceiver}
\title{Temporarily include missing receivers in the spatial object}
\usage{
new_includeUnknownReceiver(spatial, deployments, unknown.receivers)
}
\arguments{
\item{unknown.receivers}{serial number of the receivers to be included}
}
\description{
Temporarily include missing receivers in the spatial object
}
\keyword{internal}
