% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/splitDetections.R
\name{noDetectionsCheck}
\alias{noDetectionsCheck}
\title{Check if there are detections matching the target tags.}
\usage{
noDetectionsCheck(input, bio)
}
\arguments{
\item{input}{list of detections}

\item{bio}{A table with the tags and biometrics of the studied fish.}
}
\description{
Check if there are detections matching the target tags.
}
\keyword{internal}
