% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CJS.R
\name{oneWayMoves}
\alias{oneWayMoves}
\title{Trim movements table to contain only uni-directional movements}
\usage{
oneWayMoves(movements, arrays)
}
\arguments{
\item{movements}{a table of movements}

\item{arrays}{a list of array information}
}
\value{
the uni-directional movement table
}
\description{
Trim movements table to contain only uni-directional movements
}
\keyword{internal}
