% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assembleTimetable.R
\name{overrideDefaults}
\alias{overrideDefaults}
\title{Override automated logics}
\usage{
overrideDefaults(i, movements, sections)
}
\arguments{
\item{i}{The tag number currently under scrutiny. Only used for messaging purposes.}

\item{movements}{A list of movements for each target tag, created by groupMovements.}

\item{sections}{The sections in which the study was divided. Must be coincident with the names given to the ALS arrays. (i.e. if an array is 'River1', then the respective section is 'River')}
}
\value{
The first and last events for each section for the target fish.
}
\description{
If the user has called for an override for a specific tag, this function will allow for a fully manual choice of events.
}
\keyword{internal}
