% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/migration.R
\name{overrideEventCheck}
\alias{overrideEventCheck}
\title{Check event validity during manual mode}
\usage{
overrideEventCheck(type, the.event, up.boundary, t, last.events, j,
  sections, movements)
}
\arguments{
\item{type}{The type of event being checked, only affects message display}

\item{the.event}{The event being evaluated (chosen by the user during overrideDefaults)}

\item{up.boundary}{The last event for the target section}

\item{t}{The previous valid last event index}

\item{last.events}{A vector of the movement numbers corresponding to each last event. Is supplied by the function findLastEvents.}

\item{j}{The current section index}

\item{sections}{The sections in which the study was divided. Must be coincident with the names given to the ALS arrays. (i.e. if an array is 'River1', then the respective section is 'River')}

\item{movements}{A list of movements for each target tag, created by groupMovements.}
}
\value{
TRUE if the event is valid, FALSE if the event is not valid
}
\description{
Check event validity during manual mode
}
\keyword{internal}
