% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{parse_arg}
\alias{parse_arg}
\title{Parse argument and its value as a string}
\usage{
parse_arg(arg, arg_val)
}
\arguments{
\item{arg}{the argument to be parsed}

\item{arg_val}{an optional argument, to use as the value of the argument
used in "arg". Used when arg is, in itself, a complex object provided
by the user. E.g. a datapack, or an argument of preload.}
}
\value{
A string showing the argument and its value, as it would have been
inputted into the R console.
}
\description{
used to store the calls of the main actel functions as a string
}
\keyword{internal}
