% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{parse_list}
\alias{parse_list}
\title{Helper to parse list arguments}
\usage{
parse_list(arg)
}
\arguments{
\item{arg}{the argument to be parsed}
}
\value{
A string showing the list as it would have been
inputted into the R console.
}
\description{
Used inside \code{\link[=parse_arg]{parse_arg()}} to properly parse list arguments
}
\keyword{internal}
