% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preload.R
\name{preload}
\alias{preload}
\title{Load a dataset before running an analysis}
\usage{
preload(
  biometrics,
  spatial,
  deployments,
  detections,
  dot,
  distances,
  tz,
  sections = NULL,
  exclude.tags = NULL,
  disregard.parallels = FALSE,
  discard.orphans = FALSE
)
}
\arguments{
\item{biometrics}{A data frame containing biometric information.}

\item{spatial}{A data frame containing spatial information.}

\item{deployments}{A data frame containing deployment information.}

\item{detections}{A data frame containing the detections.}

\item{dot}{A DOT string of the array configuration.}

\item{distances}{A distances matrix between arrays. See \code{\link{distancesMatrix}}.}

\item{tz}{The time zone of the study area. Must match one of the values
present in \code{\link[base]{timezones}}.}

\item{sections}{The sections in which the study area is divided. Must be
coincident with the names given to the arrays. See the vignettes for more
details.}

\item{exclude.tags}{A vector of tags that should be excluded from the
detection data before any analyses are performed. Intended to be used if
stray tags from a different code space but with the same signal as a target
tag are detected in the study area.}

\item{disregard.parallels}{Logical:  Should the presence of parallel arrays
invalidate potential efficiency peers? See the vignettes for more details.}

\item{discard.orphans}{Logical: Should actel automatically discard
detections that do not fall within receiver deployment periods, or that
were recorded before the respective fish were released?}
}
\value{
...
}
\description{
This function allows the user to prepare a set of R objects to be run through
an \code{\link{explore}}, \code{\link{migration}} or \code{\link{residency}}
analysis.
}
