% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preload.R
\name{preloadDetections}
\alias{preloadDetections}
\title{Run quality checks on preloaded detections}
\usage{
preloadDetections(input, tz)
}
\arguments{
\item{input}{The detection data frame}

\item{tz}{The time zone of the study area. Must match one of the values
present in \code{\link[base]{timezones}}.}
}
\value{
A data frame with all the detections
}
\description{
Run quality checks on preloaded detections
}
\keyword{internal}
