% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{prepareAdditions}
\alias{prepareAdditions}
\title{Summarise places were more fish were released}
\usage{
prepareAdditions(group.CJS, progression)
}
\arguments{
\item{progression}{the progression data frame, provided by compileProgressionDataFrame}
}
\value{
The additions data frame, or NULL if additions should be aborted.
}
\description{
The "additions" will give the labels
}
\keyword{internal}
