% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/printFunctions.R
\name{prepareFinalTotals}
\alias{prepareFinalTotals}
\title{Prepare final totals in ggplot-friendly style}
\usage{
prepareFinalTotals(group.CJS, totals, progression)
}
\arguments{
\item{group.CJS}{a list of CJS's for each group, with release sites merged}

\item{totals}{the totals dataframe, provided by prepareTotals}

\item{progression}{the progression dataframe, provided by compileProgressionDataFrame}
}
\value{
the final total dataframe
}
\description{
the "the.final.totals" will provide the data for the red dashed line
}
\keyword{internal}
