% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/printFunctions.R
\name{printArrayOverview}
\alias{printArrayOverview}
\title{print Rmd fragment for inclusion in the report}
\usage{
printArrayOverview(array.overview)
}
\arguments{
\item{array.overview}{a list of abolute detection numbers for each group}
}
\value{
Rmd fragment
}
\description{
print Rmd fragment for inclusion in the report
}
\keyword{internal}
