% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/printFunctions.R
\name{printCircular}
\alias{printCircular}
\title{Print circular graphics for each array}
\usage{
printCircular(times, bio)
}
\arguments{
\item{times}{a (list of) circular object(s)}
}
\value{
A rmd string to be attached to the report.
}
\description{
Prints the time of first entry point on each of the arrays. Functions adapted from the circular R package.
}
\keyword{internal}
