% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{printDot}
\alias{printDot}
\title{Print DOT diagram}
\usage{
printDot(dot, sections = NULL, spatial)
}
\arguments{
\item{dot}{a dot data frame}

\item{sections}{The sections in which the study was divided. Must be coincident with the names given to the ALS arrays. (i.e. if an array is 'River1', then the respective section is 'River')}
}
\description{
Print DOT diagram
}
\keyword{internal}
