% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/printFunctions.R
\name{printEfficiency}
\alias{printEfficiency}
\title{Print efficiency fragment}
\usage{
printEfficiency(efficiency)
}
\arguments{
\item{efficiency}{A list containing the efficiency. Supplied by assembleEfficiency.}
}
\description{
Prints the ALS inter-array efficiency for inclusion in printRmd.
}
\keyword{internal}
