% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{printGlobalRatios}
\alias{printGlobalRatios}
\title{print the distribution of fish per location}
\usage{
printGlobalRatios(ratios)
}
\arguments{
\item{ratios}{the global ratios}
}
\description{
print the distribution of fish per location
}
\keyword{internal}
