% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{printIndividualResidency}
\alias{printIndividualResidency}
\title{print the individual locations per day}
\usage{
printIndividualResidency(ratios, dayrange)
}
\arguments{
\item{ratios}{the daily ratios}

\item{dayrange}{the overall first and last detection dates}
}
\value{
a Rmarkdown string to be included in the report
}
\description{
print the individual locations per day
}
\keyword{internal}
