% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{printIndividuals}
\alias{printIndividuals}
\title{Print individual graphics}
\usage{
printIndividuals(redraw, detections.list, bio, status.df = NULL, tz,
  movements, valid.movements = NULL, extension = "png")
}
\arguments{
\item{detections.list}{A list of the detections split by each target tag, created by splitDetections.}

\item{bio}{A table with the tags and biometrics of the studied fish.}

\item{tz}{The time-zone of the study area. Necessary to convert the ALS time data, which is in UTC.}

\item{movements}{A list of movements for each target tag, created by groupMovements.}

\item{extension}{the format of the generated graphics}
}
\value{
String to be included in printRmd
}
\description{
Prints the individual detections for each fish, overlaying the points in time considered crucial during the analysis.
}
\keyword{internal}
