% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{printLastArray}
\alias{printLastArray}
\title{Print a simple barplot with the number of tags last seen at each section}
\usage{
printLastArray(status.df)
}
\value{
No return value, called to plot and save graphic.
}
\description{
Print a simple barplot with the number of tags last seen at each section
}
\keyword{internal}
