% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{printLastSeen}
\alias{printLastSeen}
\title{Print a simple barplot with the number of fish last seen at each section}
\usage{
printLastSeen(input, sections)
}
\arguments{
\item{input}{a table with the last seen data}

\item{sections}{the order of the sections}
}
\value{
No return value, called to plot and save graphic.
}
\description{
Print a simple barplot with the number of fish last seen at each section
}
\keyword{internal}
