% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/printFunctions.R
\name{printLaTeX}
\alias{printLaTeX}
\title{Print LaTeX report}
\usage{
printLaTeX(name.fragment, header.fragment, biometric.fragment,
  survival.graph.size, LaTeX.individual.plots)
}
\arguments{
\item{name.fragment}{LaTeX string specifying the type of report for the title.}

\item{header.fragment}{LaTeX string specifying the type of report for the header.}

\item{biometric.fragment}{LaTeX string specifying the biometric graphics drawn.}

\item{survival.graph.size}{LaTeX string specifying the type size of the survival graphics.}

\item{LaTeX.individual.plots}{LaTeX string specifying the name of the individual plots.}
}
\description{
Creates the file 'LaTeX_report.tex', from which it is possible to create a pdf report.
}
\keyword{internal}
