% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysisMigration.R
\name{printMigrationRmd}
\alias{printMigrationRmd}
\title{Print Rmd report}
\usage{
printMigrationRmd(name.fragment, header.fragment, biometric.fragment,
  efficiency.fragment, array.overview.fragment, survival.graph.size,
  individual.plots, circular.plots, spatial, deployments)
}
\arguments{
\item{name.fragment}{Rmarkdown string specifying the type of report for the title.}

\item{header.fragment}{Rmarkdown string specifying the type of report for the header.}

\item{biometric.fragment}{Rmarkdown string specifying the biometric graphics drawn.}

\item{efficiency.fragment}{Rmarkdown string specifying the efficiency results.}

\item{array.overview.fragment}{Rmarkdown string specifying the array overview results.}

\item{survival.graph.size}{Rmarkdown string specifying the type size of the survival graphics.}

\item{individual.plots}{Rmarkdown string specifying the name of the individual plots.}

\item{circular.plots}{Rmarkdown string specifying the name of the circular plots.}
}
\description{
Creates a Rmd report and converts it to hmtl.
}
\keyword{internal}
