% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/printFunctions.R
\name{printProgression}
\alias{printProgression}
\title{Print progression graphic}
\usage{
printProgression(progression, success.arrays, spatial)
}
\arguments{
\item{progression}{A dataframe containing the number of fish that crossed each ALS array. Supplied by assembleProgression.}

\item{success.arrays}{The ALS arrays that mark the end of the study area. If a fish crosses one of these arrays, it is considered to have successfully migrated through the area.}

\item{spatial}{A list of the spatial objects, created by assembleSpatial.}
}
\description{
Prints progression graphics per fish group and release site.
}
\keyword{internal}
