% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/residency.R
\name{printResidencyRmd}
\alias{printResidencyRmd}
\title{Print Rmd report}
\usage{
printResidencyRmd(biometric.fragment, efficiency.fragment,
  individual.detection.plots, individual.residency.plots,
  array.circular.plots, section.arrival.circular.plots,
  section.departure.circular.plots, spatial, deployments, detections,
  valid.detections, last.seen, last.seen.graph.size)
}
\arguments{
\item{biometric.fragment}{Rmarkdown string specifying the biometric graphics drawn.}

\item{efficiency.fragment}{Rmarkdown string specifying the efficiency results.}
}
\description{
Creates a Rmd report and converts it to hmtl.
}
\keyword{internal}
