% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{printSectionTimes}
\alias{printSectionTimes}
\title{Print arrival and departure times per section}
\usage{
printSectionTimes(section.times, bio, detections)
}
\arguments{
\item{section.times}{A list of entry and exit times for each section}

\item{bio}{A table with the tags and biometrics of the studied animals.}

\item{detections}{A valid detections list}
}
\value{
No return value, called to plot and save graphic.
}
\description{
Print arrival and departure times per section
}
\keyword{internal}
