% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{printSensorData}
\alias{printSensorData}
\title{Print sensor data for each individual tag}
\usage{
printSensorData(
  detections,
  spatial,
  rsp.info,
  colour.by = c("auto", "stations", "arrays"),
  extension = "png"
)
}
\arguments{
\item{detections}{A valid detections list}

\item{extension}{the format of the generated graphics}
}
\value{
A string of file locations in rmd syntax, to be included in printRmd
}
\description{
Print sensor data for each individual tag
}
\keyword{internal}
