% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/printFunctions.R
\name{printStations}
\alias{printStations}
\title{Print stations table}
\usage{
printStations(spatial)
}
\arguments{
\item{spatial}{A list of the spatial objects, created by assembleSpatial.}
}
\description{
Prints the spatial object in LaTeX format for inclusion in .printLaTeX.
}
\keyword{internal}
