% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/printFunctions.R
\name{printSurvivalGraphic}
\alias{printSurvivalGraphic}
\title{Print survival graphic}
\usage{
printSurvivalGraphic(section.overview)
}
\arguments{
\item{section.overview}{A dataframe containing the survival per group of fish present in the biometrics. Supplied by assembleOverview.}
}
\description{
Prints survival graphics per fish group.
}
\keyword{internal}
