% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/printFunctions.R
\name{printSurvivalTable}
\alias{printSurvivalTable}
\title{Print survival table}
\usage{
printSurvivalTable(group.overview)
}
\arguments{
\item{group.overview}{A dataframe containing the survival per group of fish present in the biometrics. Supplied by assembleOverview.}
}
\description{
Prints the group.overview object in LaTeX format for inclusion in printLaTeX.
}
\keyword{internal}
