% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load.R
\name{processInnovaseaFile}
\alias{processInnovaseaFile}
\title{Innovasea files}
\usage{
processInnovaseaFile(input, file_type)
}
\arguments{
\item{input}{the detections data frame.}

\item{file_type}{the detected file type.}
}
\value{
A data frame of standardized detections from the input file.
}
\description{
Processes Innovasea ALS files
}
\keyword{internal}
