% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load.R
\name{processStandardFile}
\alias{processStandardFile}
\title{Standard detections file created for actel}
\usage{
processStandardFile(input)
}
\arguments{
\item{input}{the detections data frame.}
}
\value{
A data frame of standardized detections from the input file.
}
\description{
Standard detections file created for actel
}
\keyword{internal}
