% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load.R
\name{processThelmaNewFile}
\alias{processThelmaNewFile}
\title{Thelma new export files}
\usage{
processThelmaNewFile(input)
}
\arguments{
\item{input}{the detections data frame.}
}
\value{
A data frame of standardized detections from the input file.
}
\description{
Processes Thelma ALS files.
}
\keyword{internal}
