% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load.R
\name{processThelmaOldFile}
\alias{processThelmaOldFile}
\title{Thelma old export files}
\usage{
processThelmaOldFile(input)
}
\arguments{
\item{input}{the detections data frame.}
}
\value{
A data frame of standardized detections from the input file.
}
\description{
Processes Thelma ALS files.
}
\keyword{internal}
