% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadFunctions.R
\name{processVemcoFile}
\alias{processVemcoFile}
\title{Vemco files}
\usage{
processVemcoFile(input)
}
\arguments{
\item{input}{the file name, supplied by findFiles.}
}
\value{
A dataframe of standardized detections from the input file.
}
\description{
Processes Vemco ALS files
}
\keyword{internal}
