% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assembleEfficiency.R
\name{recalculateCJS}
\alias{recalculateCJS}
\title{Combine CJS model results}
\usage{
recalculateCJS(input, estimate = NULL)
}
\arguments{
\item{input}{A list of CJS results to be combined}

\item{estimate}{an estimate of the last array's efficiency}
}
\value{
the combined CJS
}
\description{
Combine CJS model results
}
\keyword{internal}
