% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{recoverLog}
\alias{recoverLog}
\title{Recover latest actel crash log}
\usage{
recoverLog(file, overwrite = FALSE)
}
\arguments{
\item{file}{Name of the file to which the log should be saved.}

\item{overwrite}{Logical: If 'file' already exists, should its content be overwritten?}
}
\value{
No return value, called for side effects.
}
\description{
Recover latest actel crash log
}
\examples{
\dontshow{
sink(paste0(tempdir(), "/latest_actel_error_log.txt"))
cat(
"This is an example file
-
-
-
Timestamp: ", Sys.Date(), "
Function: example()")
sink()
}

recoverLog(file = paste0(tempdir(), "/new_log.txt"))

\dontshow{
file.remove(paste0(tempdir(), "/latest_actel_error_log.txt"))
file.remove(paste0(tempdir(), "/new_log.txt"))
}

}
