% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/groupMovements.R
\name{refineMovements}
\alias{refineMovements}
\title{Removes discarded events}
\usage{
refineMovements(fish, sections, status.df, movements)
}
\arguments{
\item{fish}{The transmitter to be refined}

\item{sections}{The sections in which the study was divided. Must be coincident with the names given to the ALS arrays. (i.e. if an array is 'River1', then the respective section is 'River')}

\item{movements}{A list of movements for each target tag, created by groupMovements.}
}
\value{
The movement dataframe containing only valid events
}
\description{
Refine the movement events based on the actel results.
}
