% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/residency.R
\name{residency}
\alias{residency}
\title{Residency Analysis}
\usage{
residency(
  path = NULL,
  tz,
  sections,
  success.arrays = NULL,
  max.interval = 60,
  minimum.detections = 2,
  start.time = NULL,
  stop.time = NULL,
  speed.method = c("last to first", "first to first"),
  speed.warning = NULL,
  speed.error = NULL,
  jump.warning = 2,
  jump.error = 3,
  inactive.warning = NULL,
  inactive.error = NULL,
  exclude.tags = NULL,
  override = NULL,
  report = TRUE,
  section.minimum = 2,
  replicates = NULL,
  debug = FALSE,
  maximum.time = 60,
  tz.study.area = NULL,
  start.timestamp = NULL,
  end.timestamp = NULL
)
}
\arguments{
\item{path}{Path to the folder containing the data. If left NULL (default),
the analysis runs in the current folder.}

\item{tz}{The time zone of the study area. Must match one of the values
present in \code{\link[base]{timezones}}.}

\item{sections}{The sections in which the study area is divided. Must be
coincident with the names given to the arrays. See the vignettes for more
details.}

\item{success.arrays}{The arrays that mark the end of the study area. If a
fish crosses one of these arrays, it is considered to have successfully
migrated through the study area.}

\item{max.interval}{The number of minutes that must pass between detections
for a new event to be created. Defaults to 60.}

\item{minimum.detections}{For tags with only one movement event, defines the
minimum number of times a tag must have been recorded during the study
period for it to be considered true detections and not random noise.
Defaults to 2.}

\item{start.time}{Detection data prior to the timestamp set in
\code{start.time} (in YYYY-MM-DD HH:MM:SS format) is not considered during
the analysis.}

\item{stop.time}{Detection data posterior to the timestamp set in
\code{stop.time} (in YYYY-MM-DD HH:MM:SS format) is not considered during
the analysis.}

\item{speed.method}{Can take two forms: 'last to first' or 'first to first'.
If 'last to first' (default), the last detection on a given array is matched
to the first detection on the next array to perform the calculations.
If 'first to first', the first detection on a given array is matched to the
first detection on the next array to perform the calculations.}

\item{speed.warning}{If a fish moves at a speed equal or greater than
\code{speed.warning} (in metres per second), a warning is issued. If left
NULL (default), no warnings are issued.}

\item{speed.error}{If a fish moves at a speed equal or greater than
\code{speed.error} (in metres per second), user intervention is suggested.
If left NULL (default), user intervention is never suggested.}

\item{jump.warning}{If a fish crosses a number of arrays equal or greater
than \code{jump.error} without being detected, a warning is issued. If left
NULL (default), no warnings are issued.}

\item{jump.error}{If a fish crosses a number of arrays equal or greater than
\code{jump.error} without being detected, user intervention is suggested.
If left NULL (default), user intervention is never suggested.}

\item{inactive.warning}{If a fish spends a number of days equal or greater
than \code{inactive.error} in a given array at the tail of the respective
detections, a warning is issued. If left NULL (default), no warnings are
issued.}

\item{inactive.error}{If a fish spends a number of days equal or greater than
\code{inactive.error} in a given array at the tail of the respective
detections, user intervention is suggested. If left NULL (default), user
intervention is never suggested.}

\item{exclude.tags}{A vector of tags that should be excluded from the
detection data before any analyses are performed. Intended to be used if
stray tags from a different code space but with the same signal as a target
tag are detected in the study area.}

\item{override}{A vector of tags for which the user intends to manually
define which movement events are valid and invalid.}

\item{report}{Logical. Should an HTML report be created at the end of the
analysis?}

\item{section.minimum}{If a fish has less than \code{section.minimum}
consecutive detections in a section, a warning is issued. Defaults to 2.}

\item{replicates}{A list containing, for each array to which intra-array
efficiency is to be calculated: The standard names of the stations to be
used as a replicate. See the vignettes for more details.}

\item{debug}{Logical: Should temporary files be kept at the end of the
analysis?}

\item{maximum.time}{DEPRECATED. See \code{max.interval}.}

\item{tz.study.area}{DEPRECATED. See \code{tz}.}

\item{start.timestamp}{DEPRECATED. See \code{start.time}.}

\item{end.timestamp}{DEPRECATED See argument \code{stop.time}}
}
\value{
A list containing:
\itemize{
\item \code{detections}: All detections for each target fish;
\item \code{valid.detections}: Valid detections for each target fish;
\item \code{spatial}: The spatial information used during the analysis;
\item \code{deployments}: The deployments of each receiver;
\item \code{arrays}: The array details used during the analysis;
\item \code{movements}: All movement events for each target fish;
\item \code{valid.movements}: Valid movemenet events for each target fish;
\item \code{section.movements}: Valid section shifts for each target fish;
\item \code{status.df}: Summary information for each fish, including the
following columns:
\itemize{
\item \emph{Times.entered.[section]}: Total number of times the fish
entered a given section
\item \emph{Average.entry.[section]}: Average entry time at a given
section
\item \emph{Average.time.[section]}: Average time the fish spent in a
given section during each visit
\item \emph{Average.departure.[section]}: Average departure time from
a given section
\item \emph{Total.time.[section]}: Total time spent in a given section
\item \emph{Very.last.array}: Last array where the fish was detected
\item \emph{Very.last.time}: Time of the last valid detection
\item \emph{Status}: Fate assigned to the fish
\item \emph{Valid.detections}: Number of valid detections
\item \emph{Invalid.detections}: Number of invalid detections
\item \emph{Valid.events}: Number of valid events
\item \emph{Invalid.events}: Number of invalid events
\item \emph{P.type}: Type of processing:
\itemize{
\item 'Skipped' if no data was found for the fish,
\item 'Auto' if no user interaction was required,
\item 'Manual' if user interaction was suggested and the user made
changes to the validity of the events,
\item 'Overridden' if the user listed the fish in the
\code{override} argument.
}
\item \emph{Comments}: Comments left by the user during the analysis
}
\item \code{last.seen}: Summary table of the number of fish last seen in
each study area section;
\item \code{array.times}: Table containing ALL the entry times of each fish
in each array;
\item \code{section.times}: Table containing all the entry times of each
fish in each section;
\item \code{residency.list}: Places of residency between first and last
valid detection for each fish;
\item \code{daily.ratios}: Daily location per section (both in seconds spent
and in percentage of day) for each fish;
\item \code{daily.positions}: Summary table showing the location where each
fish spent the most time per day;
\item \code{global.ratios}: Summary tables showing the number of active fish
(and respective percentages) present at each location per day;
\item \code{efficiency}: Results of the inter-array Multi-way efficiency
calculations (see vignettes for more details);
\item \code{intra.array.CJS}: Results of the intra-array CJS calculations;
\item \code{rsp.info}: Appendix information for the RSP package;
\item \code{dist.mat}: The distance matrix used in the analysis (if a valid
distance matrix was supplied)
}
}
\description{
The \code{\link{residency}} analysis runs the same initial checks as
\code{\link{explore}}, but, similarly to \code{\link{migration}}, explores
particular points of the fish behaviour. If you want to know where your fish
were in each day of the study, how many fish were in each section each day,
and other residency-focused variables, this is the analysis you are looking
for!
}
\seealso{
\code{\link{explore}}, \code{\link{migration}}
}
