% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/residency.R
\name{resPositions}
\alias{resPositions}
\title{Find the location where the tag spent most time per day}
\usage{
resPositions(ratios, timestep = c("days", "hours"))
}
\arguments{
\item{ratios}{the daily ratios}
}
\value{
A data frame containing the section in which each tag spent more time per day.
}
\description{
Find the location where the tag spent most time per day
}
\keyword{internal}
