% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/residency.R
\name{resRatios}
\alias{resRatios}
\title{calculate residency ratios per fish}
\usage{
resRatios(res, timestep = c("days", "hours"), tz)
}
\arguments{
\item{res}{a residency list}

\item{timestep}{the size of each frame}
}
\value{
A list containing the residency ratios for each fish.
}
\description{
calculate residency ratios per fish
}
\keyword{internal}
