% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/residency.R
\name{resRatiosIndOut}
\alias{resRatiosIndOut}
\title{compile output of dailyRatios}
\usage{
resRatiosIndOut(input, slots, tz, tag)
}
\arguments{
\item{input}{a list containing the output of findSecondsPerSection for each day}

\item{slots}{the names of the time slots}
}
\value{
A data frame with the daily ratios table for the target tag
}
\description{
compile output of dailyRatios
}
\keyword{internal}
