% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{ringsRel}
\alias{ringsRel}
\title{Draw rings at relative points}
\usage{
ringsRel(
  plot.params,
  border,
  rings.lty,
  ring.text,
  ring.text.pos,
  ring.text.cex
)
}
\arguments{
\item{border}{the colour to draw the border. The default, NULL, means to use par("fg"). Use border = NA to omit borders.}

\item{rings.lty}{line type of the rings, See help on par.}

\item{ring.text}{logical: if notes should be displayed.}

\item{ring.text.pos}{The position of the rings' text. Ignored if ring.text is set to FALSE.}

\item{ring.text.cex}{The size of the ring's text. Ignored if ring.text is set to FALSE.}
}
\value{
No return value, adds to an existing plot.
}
\description{
Adapted from RosediagRad() (from the circular package) to draw rings inside the circular plot. Called if rings = TRUE in \code{\link{myRoseDiag}}
}
\details{
For more details about the circular package, visit its homepage at \url{https://github.com/cran/circular}
}
\keyword{internal}
