% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helperFunctions.R
\name{roundDown}
\alias{roundDown}
\title{Forcefully round a number down}
\usage{
roundDown(input, to = 10)
}
\arguments{
\item{input}{The value to be rounded}

\item{to}{The level of rounding to be applied (i.e. to=10 will round 14.8 to 10; to=1 will round i to 14)}
}
\value{
The rounded value
}
\description{
Forces the rounding of the input to the next lower rounded value.
}
\keyword{internal}
