% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assembleSpatial.R
\name{setSpatialStandards}
\alias{setSpatialStandards}
\title{Create Standard Names for spatial elements}
\usage{
setSpatialStandards(input)
}
\arguments{
\item{input}{A dataframe with spatial information.}
}
\value{
A dataframe with the same information as the input plus the Standard names.
}
\description{
Includes standard names and also reprints 'spatial.csv'
}
\keyword{internal}
