% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/groupMovements.R
\name{simplifyMovements}
\alias{simplifyMovements}
\title{Removes discarded events}
\usage{
simplifyMovements(fish = NULL, movements, status.df, sections)
}
\arguments{
\item{fish}{The transmitter to be refined. If left empty, all fish present in the movements will be simplified.}

\item{movements}{A list of movements for each target tag, created by groupMovements.}

\item{status.df}{A dataframe with all the final data for each fish, created by assembleOutput.}

\item{sections}{The sections in which the study was divided. Must be coincident with the names given to the ALS arrays. (i.e. if an array is 'River1', then the respective section is 'River')}
}
\value{
The movement dataframe containing only valid events
}
\description{
Simplify the movement events based on the actel results.
}
