% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/movementSpeeds.R
\name{speedReleaseToFirst}
\alias{speedReleaseToFirst}
\title{Calculate time and speed since release.}
\usage{
speedReleaseToFirst(fish, bio, movements, dist.mat, invalid.dist = FALSE,
  silent = TRUE)
}
\arguments{
\item{bio}{A table with the tags and biometrics of the studied fish.}

\item{movements}{A list of movements for each target tag, created by groupMovements.}

\item{dist.mat}{A matrix of the distances between the deployed ALS.}

\item{invalid.dist}{Whether or not the distance matrix supplied is valid for the study area.}

\item{silent}{logical: If TRUE, debug messages are issued (only works within actel)}
}
\value{
The movement dataframe containing the missing information.
}
\description{
Calculate time and speed since release.
}
