% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CJS.R
\name{splitCJS}
\alias{splitCJS}
\title{Split CJS matrix and calculate separate CJS}
\usage{
splitCJS(input, by, estimate = NULL, fixed.efficiency = TRUE)
}
\arguments{
\item{input}{The CJS matrix to be split}

\item{by}{a grouping vector with the same length has the number of rows in input}

\item{estimate}{an estimate of the last array's efficiency}

\item{fixed.efficiency}{logical: If TRUE, a CJS is run with the entire dataset and the efficiency estimates of that CJS are used in the split CJS's.}
}
\value{
The split CJS
}
\description{
Split CJS matrix and calculate separate CJS
}
