% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/splitDetections.R
\name{splitDetections}
\alias{splitDetections}
\title{Split detections by tag}
\usage{
splitDetections(detections, bio, spatial, exclude.tags = NULL,
  silent = FALSE)
}
\arguments{
\item{detections}{A dataframe with all the detections. Supplied by loadDetections.}

\item{bio}{A table with the tags and biometrics of the studied fish.}

\item{exclude.tags}{A list of tags that should be excluded from the detection data before any analyses are performed. Intended to be used if stray tags from a different code space but with the same signal as a target tag are detected in the study area.}

\item{silent}{logical: if TRUE, suppress messages.}
}
\value{
A list of detections for each tag.
}
\description{
Splits the detections' table by tags and selects only detections from target tags
}
\keyword{internal}
