% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{splitN}
\alias{splitN}
\title{Split a dataframe every nth row}
\usage{
splitN(x, n, row.names = FALSE)
}
\arguments{
\item{x}{the dataframe}

\item{n}{the number of rows to keep in each chunk}
}
\value{
A list of equal-sized dataframes
}
\description{
Idea from here: https://stackoverflow.com/questions/7060272/split-up-a-dataframe-by-number-of-rows
}
\keyword{internal}
