% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{splitSignals}
\alias{splitSignals}
\title{Split signals from multi-signal input}
\usage{
splitSignals(x)
}
\arguments{
\item{x}{}
}
\value{
a list of split signals
}
\description{
Split signals from multi-signal input
}
\keyword{internal}
