% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{stationName}
\alias{stationName}
\title{Find original station name}
\usage{
stationName(input, station)
}
\arguments{
\item{input}{The results of an actel analysis (either explore, migration or residency).}

\item{station}{The station standard name or number.}
}
\value{
The original station name
}
\description{
Find original station name
}
\examples{
stationName(example.results, 1)

# or

stationName(example.results, "St.2")

}
