% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{std.error.circular}
\alias{std.error.circular}
\title{Calculate the standard error of the mean}
\usage{
std.error.circular(x, na.rm = TRUE, silent = FALSE)
}
\arguments{
\item{x}{input data}

\item{na.rm}{Logical: Should missing values be removed?}

\item{silent}{Logical: Should the number of NA's removed be displayed?}
}
\value{
Standard Error of the Mean
}
\description{
Calculate the standard error of the mean
}
