% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{stopAndReport}
\alias{stopAndReport}
\title{stop function but paste error to the report too}
\usage{
stopAndReport(...)
}
\arguments{
\item{...}{parts of the error string}
}
\value{
No return value, called for side effects.
}
\description{
stop function but paste error to the report too
}
\keyword{internal}
