% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load.R
\name{storeStrays}
\alias{storeStrays}
\title{Store summary information on the stray tags detected in a permanent file.}
\usage{
storeStrays()
}
\description{
Store summary information on the stray tags detected in a permanent file.
}
\keyword{internal}
