% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{stripCodeSpaces-deprecated}
\alias{stripCodeSpaces-deprecated}
\alias{stripCodeSpaces}
\title{Remove Code Spaces from transmitter names}
\usage{
stripCodeSpaces(input)

stripCodeSpaces(input)
}
\arguments{
\item{input}{A vector of transmitter names}
}
\value{
A vector of transmitter signals
}
\description{
Remove Code Spaces from transmitter names
}
\section{\code{stripCodeSpaces}}{

For \code{stripCodeSpaces}, use \code{\link{extractSignals}}.
}

\seealso{
\code{\link{actel-deprecated}}
}
\keyword{internal}
