% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{substrRight}
\alias{substrRight}
\title{Subset a character string counting from the right end}
\usage{
substrRight(input, n)
}
\arguments{
\item{input}{Character string to be trimmed.}

\item{n}{Number of characters to be removed.}
}
\value{
A trimmed string or vector of trimmed strings.
}
\description{
Allows to quickly remove a given number of characters from the end of a string
}
\keyword{internal}
