% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/user_interaction.R
\name{tableInteraction}
\alias{tableInteraction}
\title{Handler for table interaction events}
\usage{
tableInteraction(
  moves,
  detections,
  tag,
  trigger,
  GUI,
  force = FALSE,
  save.tables.locally
)
}
\arguments{
\item{detections}{The detections data.frame for a specific tag.}

\item{tag}{The tag being analysed.}

\item{trigger}{The message/warning that triggered the interaction}

\item{GUI}{One of "needed", "always" or "never". If "needed", a new window is
opened to inspect the movements only if the movements table is too big to be
displayed in R's console. If "always", a graphical interface is always created
when the possibility to invalidate events emerges. If "never", a graphical
interface is never invoked. In this case, if the table to be displayed does
not fit in R's console, a temporary file will be saved and the user will be
prompted to open and examine that file. Defaults to "needed".}

\item{force}{Logical: If TRUE, the user is moved directly to indicating which movements should be invalidated.}

\item{save.tables.locally}{Logical: If a table must be temporarily stored into a file
for user inspection, should it be saved in the current working directory, or
in R's temporary folder?}
}
\value{
A data frame with the updated movements table
}
\description{
Handler for table interaction events
}
\keyword{internal}
