% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{timesToCircular}
\alias{timesToCircular}
\title{Convert a data frame with timestamps into a list of circular objects}
\usage{
timesToCircular(x, by.group = FALSE)
}
\arguments{
\item{x}{A data frame where the first column is an identifier, the second column is a grouping structure, and columns three and onwards are timestamps at different locations.}

\item{by.group}{Logical: Should the times at each location be divided by the group column (second column of x)?}
}
\value{
A list of circular objects
}
\description{
Convert a data frame with timestamps into a list of circular objects
}
