% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check.R
\name{transferValidity}
\alias{transferValidity}
\title{Transfer validity updates from valid movements to all movements}
\usage{
transferValidity(from, to)
}
\arguments{
\item{from}{the valid movements with newly invalidated moves}

\item{to}{the all movements table}
}
\value{
the all movements table with valid/invalid updated
}
\description{
Transfer validity updates from valid movements to all movements
}
\keyword{internal}
