% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load.R
\name{transformSpatial}
\alias{transformSpatial}
\title{Process spatial elements}
\usage{
transformSpatial(
  spatial,
  bio,
  arrays,
  dotmat,
  sections = NULL,
  first.array = NULL
)
}
\arguments{
\item{bio}{A table with the tags and biometrics of the studied fish.}

\item{first.array}{Either NULL or the top level array in the study area.}

\item{file}{an input file with spatial data.}
}
\value{
A list of 1) stations, 2) release sites, 3) ALS columns in the spatial file, 4) the Number of ASL, 5) The ALS serial numbers and 6) the array order.
}
\description{
Creates a list containing multiple spatial elements required throughout the analyses
}
\keyword{internal}
