% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distanceMatrix.R
\name{transitionLayer}
\alias{transitionLayer}
\title{Calculate distance matrix}
\usage{
transitionLayer(shape, x.res, y.res, EPSGcode, directions = c(16, 8, 4))
}
\arguments{
\item{shape}{A shape file from which to create the transition layer.}

\item{x.res, y.res}{If map is a shapefile, sets the resolution for the creation of the transition layer. Ignored if map is a pre-rendered transition layer.}

\item{directions}{If map is a shapefile,the number of directions considered for every single movement situation during cost calculation. Ignored if map is a pre-rendered transition layer.}
}
\description{
Adapted from Grant Adams' script "distance to closest mpa". (grant.adams@eagles.usm.edu)
https://cran.r-project.org/web/packages/gdistance/vignettes/gdistance1.pdf
}
