% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unknownReceiversCheck.R
\name{unknownReceiversCheckA}
\alias{unknownReceiversCheckA}
\title{Check for unknown receivers in the raw detections}
\usage{
unknownReceiversCheckA(detections, spatial)
}
\arguments{
\item{detections}{A dataframe with all the detections. Supplied by loadDetections.}

\item{spatial}{A list of the spatial objects, created by assembleSpatial.}
}
\description{
Check for unknown receivers in the raw detections
}
\keyword{internal}
