% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/z_deprecated.R
\name{unknownReceiversCheckB}
\alias{unknownReceiversCheckB}
\title{Check for target data in the unknown receivers}
\usage{
unknownReceiversCheckB(detections.list, spatial)
}
\arguments{
\item{detections.list}{A list of the detections split by each target tag, created by splitDetections.}
}
\value{
A list containing an updated spatial list and a TRUE/FALSE object indicating whether or not Standard station names must be reprocessed.
}
\description{
Check for target data in the unknown receivers
}
\keyword{internal}
