% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/residency.R
\name{updateAMValidity}
\alias{updateAMValidity}
\title{update array-movement validity based on the section-movements}
\usage{
updateAMValidity(arrmoves, secmoves)
}
\arguments{
\item{arrmoves}{the array movements}

\item{secmoves}{the section movements}
}
\value{
the updated array movements
}
\description{
update array-movement validity based on the section-movements
}
\keyword{internal}
