% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/migration.R
\name{updateMovementValidity}
\alias{updateMovementValidity}
\title{Update movement validity based on the chosen first and last events for each section}
\usage{
updateMovementValidity(movements, events, sections)
}
\arguments{
\item{movements}{A list of movements for each target tag, created by groupMovements.}

\item{events}{A list containing both the first and last events. Supplied by findFirstEvents.}

\item{sections}{The sections in which the study was divided. Must be coincident with the names given to the ALS arrays. (i.e. if an array is 'River1', then the respective section is 'River')}
}
\value{
the updated movement's table
}
\description{
Update movement validity based on the chosen first and last events for each section
}
\keyword{internal}
