% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{updateStudy}
\alias{updateStudy}
\title{Update study area for actel v.0.0.4}
\usage{
updateStudy(tz)
}
\arguments{
\item{tz}{The time-zone of the study area. Necessary to convert the ALS time data, which is in UTC.}
}
\description{
Converts the spatial.csv in itself + a deployments.csv file.
}
