% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/movements.R
\name{updateValidity}
\alias{updateValidity}
\title{update array-movement validity based on the section-movements}
\usage{
updateValidity(arrmoves, secmoves)
}
\arguments{
\item{arrmoves}{the array movements}

\item{secmoves}{the section movements}
}
\value{
A data frame with the array movements for the target fish, with an updated 'Valid' column.
}
\description{
update array-movement validity based on the section-movements
}
\keyword{internal}
