% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/movements.R
\name{upstreamCheck}
\alias{upstreamCheck}
\title{Check for movements upstream of the release site.}
\usage{
upstreamCheck(i, recipient, bio, spatial)
}
\arguments{
\item{i}{The tag being analysed.}

\item{recipient}{The movement events table for the tag.}

\item{bio}{A table with the tags and biometrics of the studied fish.}
}
\value{
The checked recipient
}
\description{
Check for movements upstream of the release site.
}
\keyword{internal}
