% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/user_interaction.R
\name{user_interaction_args}
\alias{user_interaction_args}
\title{user_interaction.R arguments}
\arguments{
\item{tag}{The tag being analysed.}

\item{detections}{The detections data.frame for a specific tag.}

\item{force}{Logical: If TRUE, the user is moved directly to indicating which movements should be invalidated.}

\item{trigger}{The message/warning that triggered the interaction}

\item{moves, all.moves}{The complete movements table for a specific tag.}

\item{displayed.moves}{The valid movements table for a specific tag.}

\item{event}{The event selected for expansion.}

\item{GUI}{One of "needed", "always" or "never". If "needed", a new window is
opened to inspect the movements only if the movements table is too big to be
displayed in R's console. If "always", a graphical interface is always created
when the possibility to invalidate events emerges. If "never", a graphical
interface is never invoked. In this case, if the table to be displayed does
not fit in R's console, a temporary file will be saved and the user will be
prompted to open and examine that file. Defaults to "needed".}

\item{save.tables.locally}{Logical: If a table must be temporarily stored into a file
for user inspection, should it be saved in the current working directory, or
in R's temporary folder?}
}
\description{
user_interaction.R arguments
}
\keyword{internal}
