% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/user_interaction.R
\name{userInput}
\alias{userInput}
\title{Wrap frequently used code to handle user input}
\usage{
userInput(question, choices, tag, hash)
}
\arguments{
\item{question}{The question to be asked}

\item{choices}{The accepted inputs. Leave empty for any input}

\item{tag}{the tag code (for comments only)}

\item{hash}{A string to attach to the decision in the UD. Ignored if input already has a hash string}
}
\description{
Wrap frequently used code to handle user input
}
\keyword{internal}
