% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/widget_setups.R
\name{widget_args}
\alias{widget_args}
\title{widget_setups.R arguments}
\arguments{
\item{displayed.moves}{The valid movements table for a specific tag.}

\item{all.moves}{The complete movements table for a specific tag.}

\item{detections}{The detections data.frame for a specific tag.}

\item{trigger}{The message/warning that triggered the interaction}

\item{first.time}{Logical: Is this the first time this widget is running for this tag?}

\item{type}{The type of events (Array or Section)}

\item{tag}{The tag being analysed.}

\item{event}{The event selected for expansion.}

\item{to.print}{The subset of detections to be displayed.}
}
\description{
widget_setups.R arguments
}
\keyword{internal}
