% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CJS.R
\name{assembleArrayCJS}
\alias{assembleArrayCJS}
\title{Combine the individual CJS's of each array into a single table}
\usage{
assembleArrayCJS(mat, CJS, dot_list, releases, silent = TRUE)
}
\arguments{
\item{CJS}{A list of CJS calculated for each array}

\item{dot_list}{A list containing information on the study area configuration.}

\item{silent}{Logical: Should messages be printed? This argument is mainly intended for function calls running within actel's analyses.}
}
\value{
A list containing the CJS absolute numbers and efficiency estimates
}
\description{
Combine the individual CJS's of each array into a single table
}
\keyword{internal}
