% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CJS.R
\name{assembleArrayCJS}
\alias{assembleArrayCJS}
\title{Combine the individual CJS's of each array into a single table}
\usage{
assembleArrayCJS(mat, CJS, arrays, releases)
}
\arguments{
\item{mat}{The presence/absence matrices.}

\item{CJS}{A list of CJS calculated for each array}

\item{arrays}{A list containing information for each array.}
}
\value{
A list containing the CJS absolute numbers and efficiency estimates
}
\description{
Combine the individual CJS's of each array into a single table
}
\keyword{internal}
