% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CJS.R
\name{assembleGroupCJS}
\alias{assembleGroupCJS}
\title{Assembles CJS tables for all groups}
\usage{
assembleGroupCJS(mat, CJS, dot_list, releases, intra.CJS = NULL)
}
\arguments{
\item{CJS}{A list of CJS calculated for each group x array combinations}

\item{dot_list}{A list containing information on the study area configuration.}
}
\value{
A data frame containing the CJS results for each group
}
\description{
Assembles CJS tables for all groups
}
\keyword{internal}
