% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CJS.R
\name{assembleGroupCJS}
\alias{assembleGroupCJS}
\title{Assembles CJS tables for all groups}
\usage{
assembleGroupCJS(mat, CJS, arrays, releases, intra.CJS = NULL)
}
\arguments{
\item{CJS}{A list of CJS calculated for each group x array combinations}

\item{arrays}{A list containing information for each array.}
}
\value{
A data frame containing the CJS results for each group
}
\description{
Assembles CJS tables for all groups
}
\keyword{internal}
