% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CJS.R
\name{assembleMatrices}
\alias{assembleMatrices}
\title{Assemble detection matrices}
\usage{
assembleMatrices(spatial, movements, status.df, arrays, paths, dotmat)
}
\arguments{
\item{movements}{A list of movements for each target tag, created by groupMovements.}

\item{simple.movements}{A list of valid-only movements for each fish}
}
\value{
A list of detection matrices split by groups and release sites
}
\description{
Assemble detection matrices
}
\keyword{internal}
