% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CJSFunctions.R
\name{assembleMatrices}
\alias{assembleMatrices}
\title{Assemble detection matrices}
\usage{
assembleMatrices(spatial, movements, minimum.detections, status.df)
}
\arguments{
\item{spatial}{A list of the spatial objects, created by assembleSpatial.}

\item{movements}{A list of movements for each target tag, created by groupMovements.}

\item{minimum.detections}{The minimum number of times a tag must have been recorded during the study period for it to be considered a true tag and not random noise.}

\item{status.df}{A dataframe with all the final data for each fish, created by assembleOutput.}
}
\value{
A list of detection matrices split by groups and release sites
}
\description{
Assemble detection matrices
}
\keyword{internal}
