% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CJSFunctions.R
\name{assembleMatrices}
\alias{assembleMatrices}
\title{Assemble detection matrices}
\usage{
assembleMatrices(spatial, simple.movements, minimum.detections, status.df)
}
\arguments{
\item{simple.movements}{A list of valid-only movements for each fish}

\item{minimum.detections}{The minimum number of times a tag must have been recorded during the study period for it to be considered a true tag and not random noise.}
}
\value{
A list of detection matrices split by groups and release sites
}
\description{
Assemble detection matrices
}
\keyword{internal}
