% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CJS.R
\name{assembleMatrices}
\alias{assembleMatrices}
\title{Assemble detection matrices}
\usage{
assembleMatrices(spatial, movements, status.df, dot_list)
}
\arguments{
\item{spatial}{The spatial data frame.}

\item{status.df}{The main results table.}

\item{dot_list}{A list containing information on the study area configuration.}
}
\value{
A list containing the detection matrices split by groups and
release sites
}
\description{
Assemble detection matrices
}
\keyword{internal}
