% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CJS.R
\name{assembleMatrices}
\alias{assembleMatrices}
\title{Assemble detection matrices}
\usage{
assembleMatrices(spatial, movements, status.df, arrays, paths, dotmat)
}
\arguments{
\item{spatial}{The spatial data frame.}

\item{status.df}{The main results table.}

\item{arrays}{A list containing information for each array.}

\item{paths}{A list containing the shortest paths between arrays with distance > 1.}

\item{dotmat}{The matrix of distances between arrays.}
}
\value{
A list containing the detection matrices split by groups and release sites
}
\description{
Assemble detection matrices
}
\keyword{internal}
