% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/migration.R
\name{assembleOutput}
\alias{assembleOutput}
\title{Create status.df}
\usage{
assembleOutput(timetable, bio, spatial, dist.mat, tz)
}
\arguments{
\item{timetable}{A table of the entering and leaving points for each section per target tag, created by assembleTimetable.}

\item{bio}{A table with the tags and biometrics of the studied animals.}

\item{spatial}{The spatial list.}

\item{dist.mat}{A matrix of the distances between the deployed ALS.}

\item{tz}{The time zone of the study area. Must match one of the values
present in \code{\link[base]{timezones}}.}
}
\value{
A data frame containing all the final data for each tag.
}
\description{
Combines the timetable and the original biometrics.
}
\keyword{internal}
