% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/residency.R
\name{assembleResidency}
\alias{assembleResidency}
\title{Collect summary information for the residency analysis}
\usage{
assembleResidency(secmoves, movements, sections)
}
\arguments{
\item{movements}{the array-movements (Valid and invalid)}

\item{sections}{The sections in which the study area is divided. Must be
coincident with the names given to the arrays. See the vignettes for more
details.}
}
\value{
A data frame with the compiled residency values
}
\description{
Collect summary information for the residency analysis
}
\keyword{internal}
