% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/residency.R
\name{assembleResidency}
\alias{assembleResidency}
\title{Collect summary information for the residency analysis}
\usage{
assembleResidency(secmoves, movements, spatial)
}
\arguments{
\item{movements}{the array-movements (Valid and invalid)}
}
\value{
A data frame with the compiled residency values
}
\description{
Collect summary information for the residency analysis
}
\keyword{internal}
