% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/migration.R
\name{assembleSectionOverview}
\alias{assembleSectionOverview}
\title{Create section.overview}
\usage{
assembleSectionOverview(status.df, sections)
}
\arguments{
\item{sections}{The sections in which the study was divided. Must be coincident with the names given to the ALS arrays. (i.e. if an array is 'River1', then the respective section is 'River')}
}
\value{
A data frame containing the survival per group of fish present in the biometrics.
}
\description{
Produces a table with the survival per group of fish present in the biometrics.
}
\keyword{internal}
