% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/migration.R
\name{assembleSectionOverview}
\alias{assembleSectionOverview}
\title{Create section.overview}
\usage{
assembleSectionOverview(status.df, spatial)
}
\arguments{
\item{spatial}{The spatial list.}
}
\value{
A data frame containing the survival per group of animals present in the biometrics.
}
\description{
Produces a table with the survival per group of animals present in the biometrics.
}
\keyword{internal}
