% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assembleTimetable.R
\name{assembleTimetable}
\alias{assembleTimetable}
\title{Create the timetable}
\usage{
assembleTimetable(movements, sections, spatial, minimum.detections,
  dist.mat, invalid.dist, speed.method, if.last.skip.section,
  success.arrays, override, cautious.assignment)
}
\arguments{
\item{movements}{A list of movements for each target tag, created by groupMovements.}

\item{sections}{The sections in which the study was divided. Must be coincident with the names given to the ALS arrays. (i.e. if an array is 'River1', then the respective section is 'River')}

\item{spatial}{A list of the spatial objects, created by assembleSpatial.}

\item{minimum.detections}{The minimum number of times a tag must have been recorded during the study period for it to be considered a true tag and not random noise.}

\item{dist.mat}{A matrix of the distances between the deployed ALS.}

\item{invalid.dist}{Whether or not the distance matrix supplied is valid for the study area.}

\item{speed.method}{One of 'last to first' or 'first to first'. In the former, the last detection on a given array/section is matched to the first detection on the next array/section (default). If changed to 'first to first', the first detection on two consecutive arrays/sections are used to perform the calculations.}

\item{if.last.skip.section}{Indicates whether a fish detected at the last array of a given section should be considered to have disappeared in the next section. Defaults to TRUE. I.e.: In a study with sections 'River' and 'Fjord', where 'River3' is the last river array, a fish last detected at River3 will be considered as 'disappeared in the Fjord'.}

\item{success.arrays}{The ALS arrays that mark the end of the study area. If a fish crosses one of these arrays, it is considered to have successfully migrated through the area.}

\item{override}{A list of tags for which the user intends to manually define entering and leaving points for each study section.}

\item{cautious.assignment}{If TRUE, actel avoids assigning events with one detection as first and/or last events of a section.}
}
\value{
A table of the entering and leaving points for each section per target tag
}
\description{
Crawls trough the movement events of each fish to find when it entered and left each section of the study area.
}
\keyword{internal}
