% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CJS.R
\name{breakMatricesByArray}
\alias{breakMatricesByArray}
\title{Break the detection matrices per array}
\usage{
breakMatricesByArray(m, dot_list, type = c("peers", "all"), verbose = TRUE)
}
\arguments{
\item{dot_list}{A list containing information on the study area configuration.}

\item{type}{The type of arrays to be matched}

\item{verbose}{Logical: Should \code{\link[=event]{event()}} be used to show and log messages?}
}
\value{
A list containing the split matrices for each array
}
\description{
Break the detection matrices per array
}
\keyword{internal}
