% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CJS.R
\name{breakMatricesByArray}
\alias{breakMatricesByArray}
\title{Break the detection matrices per array}
\usage{
breakMatricesByArray(m, arrays, type = c("peers", "all"), verbose = TRUE)
}
\arguments{
\item{arrays}{A list containing information for each array.}

\item{type}{The type of arrays to be matched}

\item{verbose}{Logical: Should appendTo be used?}
}
\value{
A list containing the split matrices for each array
}
\description{
Break the detection matrices per array
}
\keyword{internal}
