% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check.R
\name{check_args}
\alias{check_args}
\title{check.R arguments}
\arguments{
\item{arrays}{a list containing information for each array.}

\item{bio}{A table with the tags and biometrics of the studied animals.}

\item{deployments}{a list of deployments.}

\item{detections}{The detections data.frame for a specific tag.}

\item{detections.list}{A list of the detections split by each target tag,
created by \code{\link{splitDetections}}.}

\item{dist.mat}{A distances matrix.}

\item{discard.orphans}{Logical: Should actel automatically discard
detections that do not fall within receiver deployment periods, or that
were recorded before the respective animals were released?}

\item{dotmat}{The matrix of distances between arrays.}

\item{GUI}{One of "needed", "always" or "never". If "needed", a new window is
opened to inspect the movements only if the movements table is too big to be
displayed in R's console. If "always", a graphical interface is always created
when the possibility to invalidate events emerges. If "never", a graphical
interface is never invoked. In this case, if the table to be displayed does
not fit in R's console, a temporary file will be saved and the user will be
prompted to open and examine that file. Defaults to "needed".}

\item{movements}{The movements table for a specific tag.}

\item{n}{A string indicating the overall progress.}

\item{release}{The release location of the animal.}

\item{secmoves}{the section movements list.}

\item{spatial}{A list of spatial objects in the study area.}

\item{tag}{The tag being analysed.}

\item{unknown.receivers}{serial number of the receivers to be included.}

\item{valid.movements}{The valid movements table for a specific tag.}
}
\description{
check.R arguments
}
\keyword{internal}
