% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check.R
\name{checkDeploymentStations}
\alias{checkDeploymentStations}
\title{Confirm that the station names in the deployments table match those listed in the spatial file}
\usage{
checkDeploymentStations(input, spatial)
}
\arguments{
\item{input}{The table of deployments}

\item{spatial}{A list of spatial objects in the study area.}
}
\value{
A data frame with the deployment locations
}
\description{
Confirm that the station names in the deployments table match those listed in the spatial file
}
\keyword{internal}
