% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check.R
\name{checkDetectionsBeforeRelease}
\alias{checkDetectionsBeforeRelease}
\title{Check if there are detections for the target tags before release.}
\usage{
checkDetectionsBeforeRelease(input, bio, discard.orphans = FALSE)
}
\arguments{
\item{input}{The list of detections}

\item{bio}{A table with the tags and biometrics of the studied fish.}
}
\value{
A list containing the detections without invalid data.
}
\description{
Check if there are detections for the target tags before release.
}
\keyword{internal}
