% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check.R
\name{checkDetectionsBeforeRelease}
\alias{checkDetectionsBeforeRelease}
\title{Check if there are detections for the target tags before release.}
\usage{
checkDetectionsBeforeRelease(input, bio, discard.orphans = FALSE)
}
\arguments{
\item{input}{The list of detections}

\item{bio}{A table with the tags and biometrics of the studied animals.}

\item{discard.orphans}{Logical: Should actel automatically discard
detections that do not fall within receiver deployment periods, or that
were recorded before the respective animals were released?}
}
\value{
A list containing the detections without invalid data.
}
\description{
Check if there are detections for the target tags before release.
}
\keyword{internal}
