% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check.R
\name{checkDupDetections}
\alias{checkDupDetections}
\title{Check if there are duplicated detection in the input data}
\usage{
checkDupDetections(input)
}
\arguments{
\item{input}{The detections data frame}
}
\value{
The detections data frame, without duplicated detections.
}
\description{
Check if there are duplicated detection in the input data
}
\keyword{internal}
