% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check.R
\name{checkDupSignals}
\alias{checkDupSignals}
\title{Check if there are duplicated signals in the detected tags.}
\usage{
checkDupSignals(input, bio)
}
\arguments{
\item{input}{list of detections}

\item{bio}{the biometrics table}
}
\value{
No return value, called for side effects.
}
\description{
Check if there are duplicated signals in the detected tags.
}
\keyword{internal}
