% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check.R
\name{checkImpassables}
\alias{checkImpassables}
\title{Find out if a tag moved in an impossible direction}
\usage{
checkImpassables(movements, tag, detections, dotmat, GUI, save.tables.locally)
}
\arguments{
\item{movements}{The movements table for a specific tag.}

\item{tag}{The tag being analysed.}

\item{detections}{The detections data.frame for a specific tag.}

\item{dotmat}{The matrix of distances between arrays.}

\item{GUI}{One of "needed", "always" or "never". If "needed", a new window is
opened to inspect the movements only if the movements table is too big to be
displayed in R's console. If "always", a graphical interface is always created
when the possibility to invalidate events emerges. If "never", a graphical
interface is never invoked. In this case, if the table to be displayed does
not fit in R's console, a temporary file will be saved and the user will be
prompted to open and examine that file. Defaults to "needed".}
}
\value{
A list of movements with updated 'Valid' columns
}
\description{
Find out if a tag moved in an impossible direction
}
\keyword{internal}
