% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check.R
\name{checkImpassables}
\alias{checkImpassables}
\title{Find out if a fish moved in an impossible direction}
\usage{
checkImpassables(movements, fish, dotmat)
}
\arguments{
\item{movements}{the movements list}

\item{fish}{the fish being analysed}

\item{dotmat}{the matrix of distances between arrays}
}
\value{
the movements list with valid/invalid rows
}
\description{
Find out if a fish moved in an impossible direction
}
\keyword{internal}
