% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check.R
\name{checkInactiveness}
\alias{checkInactiveness}
\title{Find if a fish is standing still in an array}
\usage{
checkInactiveness(
  movements,
  fish,
  detections.list,
  inactive.warning,
  inactive.error,
  invalid.dist,
  dist.mat
)
}
\arguments{
\item{movements}{The movements table}

\item{fish}{The fish being analysed}

\item{detections.list}{A list of the detections split by each target tag, created by splitDetections.}

\item{inactive.warning}{If a fish spends a number of days equal or greater
than \code{inactive.error} in a given array at the tail of the respective
detections, a warning is issued. If left NULL (default), no warnings are
issued.}

\item{inactive.error}{If a fish spends a number of days equal or greater than
\code{inactive.error} in a given array at the tail of the respective
detections, user intervention is suggested. If left NULL (default), user
intervention is never suggested.}
}
\value{
The valid movements list with valid/invalid rows
}
\description{
Find if a fish is standing still in an array
}
\keyword{internal}
