% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check.R
\name{checkJumpDistance}
\alias{checkJumpDistance}
\title{Check if fish are jumping over arrays}
\usage{
checkJumpDistance(movements, fish, release, dotmat, jump.warning, jump.error)
}
\arguments{
\item{movements}{The movements table}

\item{fish}{The fish being analysed}

\item{release}{The release location of the fish}

\item{dotmat}{The matrix of distances between arrays}

\item{jump.warning}{If a fish crosses a number of arrays equal or greater
than \code{jump.error} without being detected, a warning is issued. If left
NULL (default), no warnings are issued.}

\item{jump.error}{If a fish crosses a number of arrays equal or greater than
\code{jump.error} without being detected, user intervention is suggested.
If left NULL (default), user intervention is never suggested.}
}
\value{
a checked movements list
}
\description{
Check if fish are jumping over arrays
}
\keyword{internal}
