% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check.R
\name{checkJumpDistance}
\alias{checkJumpDistance}
\title{Check if fish are jumping over arrays}
\usage{
checkJumpDistance(movements, bio, spatial, dotmat, jump.warning = 2,
  jump.error = 3)
}
\arguments{
\item{movements}{A list of movements for each target tag, created by groupMovements.}

\item{bio}{A table with the tags and biometrics of the studied fish.}

\item{spatial}{A list of spatial objects in the study area}

\item{dotmat}{A dot distance matrix}

\item{jump.warning}{Integer value. If a fish crosses ´jump.warning´ arrays without being detected, a warning is issued.}

\item{jump.error}{Integer value. If a fish crosses ´jump.error´ arrays without being detected, user intervention is suggested.}
}
\value{
a checked movements list
}
\description{
Check if fish are jumping over arrays
}
\keyword{internal}
