% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check.R
\name{checkLinearity}
\alias{checkLinearity}
\title{Check that the tag linearly moved along the sections}
\usage{
checkLinearity(
  secmoves,
  tag,
  spatial,
  arrays,
  GUI,
  save.tables.locally,
  n,
  back.warning = "all",
  back.error = "all"
)
}
\arguments{
\item{secmoves}{the section movements list.}

\item{tag}{The tag being analysed.}

\item{spatial}{A list of spatial objects in the study area.}

\item{arrays}{a list containing information for each array.}

\item{GUI}{One of "needed", "always" or "never". If "needed", a new window is
opened to inspect the movements only if the movements table is too big to be
displayed in R's console. If "always", a graphical interface is always created
when the possibility to invalidate events emerges. If "never", a graphical
interface is never invoked. In this case, if the table to be displayed does
not fit in R's console, a temporary file will be saved and the user will be
prompted to open and examine that file. Defaults to "needed".}

\item{save.tables.locally}{Logical: If a table must be temporarily stored into a file
for user inspection, should it be saved in the current working directory, or
in R's temporary folder?}

\item{n}{A string indicating the overall progress.}

\item{back.warning}{Warn the user about backwards inter-section movements.
One of three values: "all" will warn the user about S shaped turns (where
the animal moves backwards but then resumes normal movement), and U shaped
turns (where the animal moves backwards and disappears); "u" will warn the
about U-shaped turns (i.e. no warnings for S-shaped turns); "none" will
suppress all inter-section movement warnings.}

\item{back.error}{Pause the analysis so the user can inspect the movements
table that generated an inter-section backwards movement warning.
One of three values: "all" will prompt the user both for S shaped turns and
U shaped turns; "u" will prompt the user only for U-shaped turns, and "none"
will never pause the analysis for inter-section movements inspection.}
}
\value{
A list of section movements with updated 'Valid' columns
}
\description{
Check that the tag linearly moved along the sections
}
\keyword{internal}
