% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check.R
\name{checkLinearity}
\alias{checkLinearity}
\title{Check that the fish linearly moved along the sections}
\usage{
checkLinearity(secmoves, fish, sections, arrays)
}
\arguments{
\item{secmoves}{the section movements list}

\item{fish}{the fish being analysed}

\item{sections}{The sections in which the study was divided. Must be coincident with the names given to the ALS arrays. (i.e. if an array is 'River1', then the respective section is 'River')}

\item{arrays}{a list containing information for each array}
}
\value{
the section movements with valid/invalid notes
}
\description{
Check that the fish linearly moved along the sections
}
\keyword{internal}
