% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check.R
\name{checkMinimumN}
\alias{checkMinimumN}
\title{Check that the tag have enough detections to be valid}
\usage{
checkMinimumN(movements, minimum.detections, tag, n)
}
\arguments{
\item{movements}{The movements table for a specific tag.}

\item{minimum.detections}{For tags with only one movement event, defines the
minimum number of times a tag must have been recorded during the study
period for it to be considered true detections and not random noise.
Defaults to 2.}

\item{tag}{The tag being analysed.}

\item{n}{A string indicating the overall progress.}
}
\value{
A data frame containing the movements with valid/invalid notes
}
\description{
Check that the tag have enough detections to be valid
}
\keyword{internal}
