% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check.R
\name{checkPath}
\alias{checkPath}
\title{Check path validity}
\usage{
checkPath(my.home, path)
}
\arguments{
\item{path}{Path to the folder containing the data. If left NULL (default),
the analysis runs in the current folder.}
}
\description{
Confirms that the target directory exists.
}
\keyword{internal}
