% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check.R
\name{checkReport}
\alias{checkReport}
\title{Check report compatibility}
\usage{
checkReport(report)
}
\arguments{
\item{report}{Logical. Should an HTML report be created at the end of the
analysis?}
}
\value{
A TRUE/FALSE decision
}
\description{
Checks if pandoc is up and running and creates a "Report" folder, if necessary
}
\keyword{internal}
