% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check.R
\name{checkSMovesN}
\alias{checkSMovesN}
\title{Verify number of detections in section movements}
\usage{
checkSMovesN(secmoves, fish, section.minimum)
}
\arguments{
\item{secmoves}{the section movements list}

\item{fish}{The fish being analysed}

\item{section.minimum}{If a fish has less than \code{section.minimum}
consecutive detections in a section, a warning is issued. Defaults to 2.}
}
\value{
the section movements with valid/invalid notes
}
\description{
Verify number of detections in section movements
}
\keyword{internal}
