% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check.R
\name{checkSMovesN}
\alias{checkSMovesN}
\title{Verify number of detections in section movements}
\usage{
checkSMovesN(secmoves, section.minimum)
}
\arguments{
\item{secmoves}{the section movements list}

\item{section.minimum}{threshold consecutive number of detections in a section that awards confidence in the results. If the number of detections is lower than this argument, a warning is issued and user interaction is allowed.}
}
\value{
the section movements with valid/invalid notes
}
\description{
Verify number of detections in section movements
}
\keyword{internal}
