% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check.R
\name{checkSMovesN}
\alias{checkSMovesN}
\title{Verify number of detections in section movements}
\usage{
checkSMovesN(secmoves, fish, section.minimum, GUI)
}
\arguments{
\item{secmoves}{the section movements list}

\item{fish}{The fish being analysed}

\item{section.minimum}{If a fish has less than \code{section.minimum}
consecutive detections in a section, a warning is issued. Defaults to 2.}

\item{GUI}{One of "needed", "always" or "never". If "needed", a new window is
opened to inspect the movements only if the movements table is too big to be
displayed in R's console. If "always", a graphical interface is always created
when the possibility to invalidate events emerges. If "never", a graphical
interface is never invoked. In this case, if the table to be displayed does
not fit in R's console, a temporary file will be saved and the user will be
prompted to open and examine that file. Defaults to "needed".}
}
\value{
A list of section movements with updated 'Valid' columns
}
\description{
Verify number of detections in section movements
}
\keyword{internal}
