% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check.R
\name{checkSpeeds}
\alias{checkSpeeds}
\title{check fish speeds against defined thresholds (in m/s)}
\usage{
checkSpeeds(movements, fish, valid.movements, speed.warning, speed.error, GUI)
}
\arguments{
\item{movements}{The movements table}

\item{fish}{The fish being analysed}

\item{valid.movements}{The valid movements table}

\item{speed.warning}{If a fish moves at a speed equal or greater than
\code{speed.warning} (in metres per second), a warning is issued. If left
NULL (default), no warnings are issued.}

\item{speed.error}{If a fish moves at a speed equal or greater than
\code{speed.error} (in metres per second), user intervention is suggested.
If left NULL (default), user intervention is never suggested.}

\item{GUI}{One of "needed", "always" or "never". If "needed", a new window is
opened to inspect the movements only if the movements table is too big to be
displayed in R's console. If "always", a graphical interface is always created
when the possibility to invalidate events emerges. If "never", a graphical
interface is never invoked. In this case, if the table to be displayed does
not fit in R's console, a temporary file will be saved and the user will be
prompted to open and examine that file. Defaults to "needed".}
}
\value{
A list of movements with updated 'Valid' columns
}
\description{
check fish speeds against defined thresholds (in m/s)
}
\keyword{internal}
