% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check.R
\name{checkSpeeds}
\alias{checkSpeeds}
\title{check tag speeds against defined thresholds (in m/s)}
\usage{
checkSpeeds(
  movements,
  tag,
  detections,
  valid.movements,
  speed.warning,
  speed.error,
  GUI,
  save.tables.locally,
  n
)
}
\arguments{
\item{movements}{The movements table for a specific tag.}

\item{tag}{The tag being analysed.}

\item{detections}{The detections data.frame for a specific tag.}

\item{valid.movements}{The valid movements table for a specific tag.}

\item{speed.warning}{If a tag moves at a speed equal or greater than
\code{speed.warning} (in metres per second), a warning is issued. If left
NULL (default), no warnings are issued. Must be equal to or lower than \code{speed.error}}

\item{speed.error}{If a tag moves at a speed equal or greater than
\code{speed.error} (in metres per second), user intervention is suggested.
If left NULL (default), user intervention is never suggested.}

\item{GUI}{One of "needed", "always" or "never". If "needed", a new window is
opened to inspect the movements only if the movements table is too big to be
displayed in R's console. If "always", a graphical interface is always created
when the possibility to invalidate events emerges. If "never", a graphical
interface is never invoked. In this case, if the table to be displayed does
not fit in R's console, a temporary file will be saved and the user will be
prompted to open and examine that file. Defaults to "needed".}

\item{save.tables.locally}{Logical: If a table must be temporarily stored into a file
for user inspection, should it be saved in the current working directory, or
in R's temporary folder?}

\item{n}{A string indicating the overall progress.}
}
\value{
A list of movements with updated 'Valid' columns
}
\description{
check tag speeds against defined thresholds (in m/s)
}
\keyword{internal}
