% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check.R
\name{checkTagsInUnknownReceivers}
\alias{checkTagsInUnknownReceivers}
\title{Check for target data in the unknown receivers}
\usage{
checkTagsInUnknownReceivers(detections.list, deployments, spatial)
}
\arguments{
\item{detections.list}{A list of the detections split by each target tag, created by \code{splitDetections}.}

\item{spatial}{The spatial data frame, as loaded by loadSpatial}
}
\value{
A list containing an updated spatial list and a TRUE/FALSE object indicating whether or not Standard station names must be reprocessed.
}
\description{
Check for target data in the unknown receivers
}
\keyword{internal}
