% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check.R
\name{checkToken}
\alias{checkToken}
\title{Verify that the source data has been compiled using actel's preload function}
\usage{
checkToken(token, timestamp)
}
\arguments{
\item{token}{The token generated by preload}

\item{timestamp}{The time when the token was issued}
}
\value{
No return value. Called for side effects.
}
\description{
Verify that the source data has been compiled using actel's preload function
}
\keyword{internal}
