% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check.R
\name{checkUnknownReceivers}
\alias{checkUnknownReceivers}
\title{Find detections from unknown receivers}
\usage{
checkUnknownReceivers(input)
}
\arguments{
\item{input}{The detections data frame}
}
\description{
Find detections from unknown receivers
}
\keyword{internal}
