% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check.R
\name{checkUpstream}
\alias{checkUpstream}
\title{Check for movements upstream of the release site.}
\usage{
checkUpstream(movements, fish, release, arrays)
}
\value{
The updated movements
}
\description{
Check for movements upstream of the release site.
}
\keyword{internal}
