% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check.R
\name{checkUpstream}
\alias{checkUpstream}
\title{Check for movements upstream of the release site.}
\usage{
checkUpstream(movements, fish, release, arrays, GUI)
}
\arguments{
\item{movements}{The movements table}

\item{fish}{The fish being analysed}

\item{release}{The release location of the fish}

\item{arrays}{a list containing information for each array}

\item{GUI}{One of "needed", "always" or "never". If "needed", a new window is
opened to inspect the movements only if the movements table is too big to be
displayed in R's console. If "always", a graphical interface is always created
when the possibility to invalidate events emerges. If "never", a graphical
interface is never invoked. In this case, if the table to be displayed does
not fit in R's console, a temporary file will be saved and the user will be
prompted to open and examine that file. Defaults to "needed".}
}
\value{
A data frame with the movements for the target fish with updated 'Valid' column.
}
\description{
Check for movements upstream of the release site.
}
\keyword{internal}
