% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CJS.R
\name{cjs_args}
\alias{cjs_args}
\title{CJS.R arguments}
\arguments{
\item{replicates}{A list of arrays containing the respective replicate stations.}

\item{CJS}{The overall CJS results.}

\item{detections.list}{A list of the detections split by each target tag.}

\item{spatial}{The spatial data frame.}

\item{efficiency}{The efficiency results.}

\item{arrays}{A list containing information for each array.}

\item{mat, m}{The presence/absence matrices.}

\item{input}{A presence/Absence matrix.}

\item{movements, moves}{The movements table.}

\item{status.df}{The main results table.}

\item{dotmat}{The matrix of distances between arrays.}

\item{paths}{A list containing the shortest paths between arrays with distance > 1.}

\item{estimate}{an estimate of the last array's efficiency.}

\item{fixed.efficiency}{A vector of fixed efficiency estimates from a more complete CJS model.}

\item{silent}{Logical: Should messages be printed?}
}
\description{
CJS.R arguments
}
\keyword{internal}
