% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{clearWorkspace}
\alias{clearWorkspace}
\title{Delete previous analysis files from the current workspace. Input files are not deleted.}
\usage{
clearWorkspace(skip = NA)
}
\arguments{
\item{skip}{A vector of files to be ignored.}
}
\description{
Delete previous analysis files from the current workspace. Input files are not deleted.
}
