% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{collectStrays}
\alias{collectStrays}
\title{Collect summary information on the tags detected but that are not part of the study.}
\usage{
collectStrays(input, restart = FALSE)
}
\arguments{
\item{input}{list of detections for the tags to be excluded.}

\item{restart}{logical: if TRUE, remove file 'temp_strays.csv' from the working directory.}
}
\description{
Collect summary information on the tags detected but that are not part of the study.
}
\keyword{internal}
