% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{combine}
\alias{combine}
\title{Combine a list of vectors}
\usage{
combine(input)
}
\arguments{
\item{input}{a list of vectors with non-overlapping data.}
}
\value{
A vectorized combination of the data in the different list vectors.
}
\description{
Intended to combine vectors where, for each position, only one of the vectors contains data (i.e. the remaining are NA's).
}
\keyword{internal}
