% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CJS.R
\name{combineCJS}
\alias{combineCJS}
\title{Combine multiple CJS models}
\usage{
combineCJS(..., estimate = NULL, fixed.efficiency = NULL, silent = FALSE)
}
\arguments{
\item{...}{The detection matrices to be joined, or a list containing detection matrices}

\item{estimate}{an estimate of the last array's efficiency}

\item{fixed.efficiency}{A vector of fixed efficiency estimates from a more complete CJS model.}

\item{silent}{logical: if TRUE, suppress messages.}
}
\value{
The combined CJS
}
\description{
Combine multiple CJS models
}
