% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load.R
\name{compileDetections}
\alias{compileDetections}
\title{Combine ALS detections}
\usage{
compileDetections(path = "detections", start.timestamp = NULL,
  end.timestamp = NULL, tz.study.area)
}
\arguments{
\item{path}{the path(s) to the detection files}

\item{start.timestamp}{Detection data prior to this date is not analysed. Improves processing time when loading large amounts of detection data.}

\item{end.timestamp}{Detection data posterior to this date is not analysed. Improves processing time when loading large amounts of detection data.}

\item{tz.study.area}{The time-zone of the study area. Necessary to convert the ALS time data, which is in UTC.}
}
\value{
A dataframe with all the detections
}
\description{
Finds the detections' files and processes them.
}
\keyword{internal}
