% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{compileProgressionDataFrame}
\alias{compileProgressionDataFrame}
\title{Compile Progression in a ggplot-friendly fashion}
\usage{
compileProgressionDataFrame(status.df, group.CJS, detailed.absolutes, i,
  the.levels)
}
\arguments{
\item{detailed.absolutes}{a list of absolutes extracted from split.CJS}

\item{i}{the current iteration number}

\item{the.levels}{a vector containing levels for later use in ggplot}
}
\value{
An updated progression data frame and an updated the.levels
}
\description{
Used within printProgression
}
\keyword{internal}
