% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load.R
\name{convertCodes}
\alias{convertCodes}
\title{Convert code spaces}
\usage{
convertCodes(input)
}
\arguments{
\item{input}{A data frame of standardized detections.}
}
\value{
A data frame with standardized code spaces.
}
\description{
Unifies CodeSpace names, to avoid having different names depending on ALS vendor.
}
\keyword{internal}
