% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadFunctions.R
\name{convertTimes}
\alias{convertTimes}
\title{Convert Times}
\usage{
convertTimes(input, start.timestamp, end.timestamp, tz.study.area)
}
\arguments{
\item{input}{A dataframe of standardized detections.}

\item{start.timestamp}{Detection data prior to this date is not analysed. Improves processing time when loading large amounts of detection data.}

\item{end.timestamp}{Detection data posterior to this date is not analysed. Improves processing time when loading large amounts of detection data.}

\item{tz.study.area}{The time-zone of the study area. Necessary to convert the ALS time data, which is in UTC.}
}
\value{
A dataframe with corrected timestamps.
}
\description{
Converts the ALS timestamps (UTC) to the designated study area time zone. Can also trim the data by time.
}
\keyword{internal}
