% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load.R
\name{convertTimes}
\alias{convertTimes}
\title{Convert Times}
\usage{
convertTimes(input, start.time, stop.time, tz)
}
\arguments{
\item{input}{A data frame of standardized detections.}

\item{start.time}{Detection data prior to this date is not analysed. Improves processing time when loading large amounts of detection data.}

\item{stop.time}{Detection data posterior to this date is not analysed. Improves processing time when loading large amounts of detection data.}

\item{tz}{The time-zone of the study area. Necessary to convert the ALS time data, which is in UTC.}
}
\value{
A data frame with corrected timestamps.
}
\description{
Converts the ALS timestamps (UTC) to the designated study area time zone. Can also trim the data by time.
}
\keyword{internal}
