% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load.R
\name{convertTimes}
\alias{convertTimes}
\title{Convert Times}
\usage{
convertTimes(input, start.time, stop.time, tz)
}
\arguments{
\item{input}{A data frame of standardized detections.}

\item{start.time}{Detection data prior to the timestamp set in
\code{start.time} (in YYYY-MM-DD HH:MM:SS format) is not considered during
the analysis.}

\item{stop.time}{Detection data posterior to the timestamp set in
\code{stop.time} (in YYYY-MM-DD HH:MM:SS format) is not considered during
the analysis.}

\item{tz}{The time zone of the study area. Must match one of the values
present in \code{\link[base]{timezones}}.}
}
\value{
A data frame with corrected timestamps.
}
\description{
Converts the ALS timestamps (UTC) to the designated study area time zone. Can also trim the data by time.
}
\keyword{internal}
