% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/circularFunctions.R
\name{convertTimesToCircular}
\alias{convertTimesToCircular}
\title{Convert times dataframe into a list of circular objects}
\usage{
convertTimesToCircular(times)
}
\arguments{
\item{times}{A dataframe with the timestamps fo reach fish and array}
}
\value{
A list of circular objects
}
\description{
Convert times dataframe into a list of circular objects
}
\keyword{internal}
