% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CJS.R
\name{countArrayFailures}
\alias{countArrayFailures}
\title{Find and list arrays which failed during the movements of the tags}
\usage{
countArrayFailures(moves, paths, dotmat)
}
\arguments{
\item{moves}{the valid array movements}

\item{paths}{a list containing the shortest paths between arrays with distance > 1}
}
\value{
NULL if no arrays failed, or a list containing information on the arrays that failed
}
\description{
Find and list arrays which failed during the movements of the tags
}
\keyword{internal}
