% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/migration.R
\name{countBackMoves}
\alias{countBackMoves}
\title{Count backwards movements}
\usage{
countBackMoves(movements, dot_list)
}
\arguments{
\item{movements}{the movement dataframe for the tag being analysed.}

\item{dot_list}{A list containing information on the study area configuration.}
}
\value{
A list containing:
\itemize{
\item \code{sum.back.moves} The number of backwards movements for
the target tag
\item \code{max.back.moves} The maximum number of consecutive backwards
movements for the target tag
}
}
\description{
Count backwards movements
}
\keyword{internal}
