% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/migration.R
\name{countBackMoves}
\alias{countBackMoves}
\title{Count backwards movements}
\usage{
countBackMoves(movements, arrays)
}
\arguments{
\item{movements}{A list of movements for each target tag, created by groupMovements.}

\item{arrays}{A list containing information for each array.}
}
\value{
A list containing:
\itemize{
\item \code{sum.back.moves} The number of backwards movements for the target tag
\item \code{max.back.moves} The maximum number of consecutive backwards movements for the target tag
}
}
\description{
Count backwards movements
}
\keyword{internal}
