% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/z_examples.R
\name{createWorkspace}
\alias{createWorkspace}
\title{Create a default Workspace}
\usage{
createWorkspace(dir = "actel_workspace")
}
\description{
Produces the files and folders required to run the package, which the user can use as a template.
}
