% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/z_examples.R
\name{createWorkspace}
\alias{createWorkspace}
\title{Create a Default Workspace}
\usage{
createWorkspace(dir)
}
\arguments{
\item{dir}{The name of the target directory. Will be created if not present.}
}
\value{
No return value, called for side effects
}
\description{
Produces template files and folders required to run the \code{\link{explore}},
\code{\link{migration}} and \code{\link{residency}} functions.
}
\examples{
\donttest{
# running createWorkspace deploys template 
# files to a directory specified by the user
createWorkspace(paste0(tempdir(), "/createWorkspace_example"))
}

}
