% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/z_examples.R
\name{createWorkspace}
\alias{createWorkspace}
\title{Create a Default Workspace}
\usage{
createWorkspace(dir = "actel_workspace")
}
\arguments{
\item{dir}{The name of the target directory. Will be created if not present.}
}
\value{
No return value, called for side effects
}
\description{
Produces the files and folders required to run the package, which the user can use as a template.
Note: This function will create a new sub-directory in the current directory and write files into it.
}
\examples{
\donttest{
# running createWorkspace deploys template files for an analysis
# by default, createWorkspace creates a directory called 'actel_workspace'
# but this can be changed with the argument 'dir'.
createWorkspace(dir = tempdir())
}

}
