% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/z_examples.R
\name{createWorkspace}
\alias{createWorkspace}
\title{Deprecated function.}
\usage{
createWorkspace(dir, force = FALSE)
}
\arguments{
\item{dir}{The name of the target directory. Will be created if not present.}

\item{force}{logical. Defaults to FALSE. Prevents deploying files in a directory that already exists without explicit permission.}
}
\value{
No return value, called for side effects
}
\description{
Use blankWorkspace instead.
}
\examples{
\donttest{
# createWorkspace is deprecated. Use blankWorkspace instead.
}

}
