% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/residency.R
\name{dailyPositions}
\alias{dailyPositions}
\title{Find the location where the fish spent most time per day}
\usage{
dailyPositions(ratios)
}
\arguments{
\item{ratios}{the daily ratios}
}
\value{
a data frame with the transmitters as columns and the days as rows
}
\description{
Find the location where the fish spent most time per day
}
\keyword{internal}
