% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{dataToList}
\alias{dataToList}
\title{Import actel results into a single list object}
\usage{
dataToList(source)
}
\arguments{
\item{source}{A RData file containing actel results from a previous run}
}
\description{
Import actel results into a single list object
}
