% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{dataToList}
\alias{dataToList}
\title{Import RData in a list format}
\usage{
dataToList(source)
}
\arguments{
\item{source}{A RData file.}
}
\value{
A list containing the objects present in the source RData file.
}
\description{
Import RData in a list format
}
\examples{
\donttest{
dataTolist("Path/to/file.RData")
}

}
