% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{decimalTime}
\alias{decimalTime}
\title{Convert hh:mm:ss time to hh.hhh}
\usage{
decimalTime(input, unit = c("h", "m", "s"))
}
\arguments{
\item{input}{Single string or a vector of strings containing hours:minutes or hours:minutes:seconds.}

\item{unit}{the desired units of the output, one of "h" (hours), "m", (minutes) or "s" (seconds).}
}
\value{
A number or vector of numbers corresponding to the decimal hour equivalent of the character input.
}
\description{
Convert hh:mm:ss time to hh.hhh
}
\keyword{internal}
