% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{distancesMatrix}
\alias{distancesMatrix}
\title{Calculate Distances Matrix}
\usage{
distancesMatrix(
  t.layer,
  starters = NULL,
  targets = starters,
  EPSGcode,
  coord.x = "x",
  coord.y = "y",
  id.col = NULL,
  actel = TRUE
)
}
\arguments{
\item{t.layer}{A TransitionLayer object, generated by \code{transitionLayer}.}

\item{starters}{A data frame with the points from which to start measuring the distance. Ignored if actel = TRUE (default), as the 'spatial.csv' is loaded as starters.}

\item{targets}{A data frame with the points to which a way must be found. Ignored if actel = TRUE (default), as the 'spatial.csv' is loaded as targets.}

\item{EPSGcode}{The EPSG code corresponding to the coordinate system of the input data. All inputs must be in the same metric system. DO NOT use degree-based coordinates.}

\item{coord.x, coord.y}{The names of the columns containing the x and y coordinates. Must be identical in the starters and targets.}

\item{id.col}{The name of the column containing the IDs of the points to be used as starters and targets. Must be identical in both files.  Ignored if actel = TRUE (default), as the stations' standard names are used.}

\item{actel}{Logical: Should the distance matrix be optimized for actel? Defaults to TRUE.}
}
\value{
A matrix with the distances between each pair of points.
}
\description{
Using a previously created transition layer (see \code{\link{transitionLayer}}), calculates the distances
between spatial points. Adapted from Grant Adams' script "distance to closest mpa". if the argument 'actel'
is set to TRUE (default), an actel-compatible matrix is generated, and the user will be asked if they would
like to save the matrix as 'distances.csv' in the current directory.
}
\details{
It is highly recommended to read the vignette regarding distances matrix before running this function.
You can find it by running \code{vignette('a-2_distances_matrix', 'actel')} or \code{browseVignettes('actel')}
}
\examples{
\donttest{
# Assuming a shapefile was saved to 'shapefile.shp' in the current
# working directory
x <- transitionLayer("shapefile.shp", size = 20, EPSGcode = 32362)

# Assuming there is a 'spatial.txt' file in the current directory,
# with a 'x' and 'y' column specifying coordinates in the same system
# coordinate system as the original shapefile.
distancesMatrix(x, coord.x = 'x', coord.y = 'y', EPSGcode = 32632)

# Alternatively, if you want to calculate a distances matrix that is not
# optimised for future actel analyses, you can set 'actel' to FALSE and 
# determine the 'starters', 'targets' and 'id.col' manually.
}

}
