% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load.R
\name{dotList}
\alias{dotList}
\title{Break the dot data frame into a list}
\usage{
dotList(input, sections = NULL)
}
\arguments{
\item{input}{a dot data frame}

\item{sections}{The sections in which the study was divided. Must be coincident with the names given to the ALS arrays. (i.e. if an array is 'River1', then the respective section is 'River')}
}
\value{
A list containing, for each array, the arrays that connect to it and to which it connects.
}
\description{
Break the dot data frame into a list
}
\keyword{internal}
