% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load.R
\name{dotList}
\alias{dotList}
\title{Break the dot data frame into a list}
\usage{
dotList(input, sections = NULL)
}
\arguments{
\item{input}{a dot data frame}

\item{sections}{The sections in which the study area is divided. Must be
coincident with the names given to the arrays. See the vignettes for more
details.}
}
\value{
A list containing detailed information on the arrays
}
\description{
Break the dot data frame into a list
}
\keyword{internal}
