% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load.R
\name{dotList}
\alias{dotList}
\title{Break the dot data frame into a list}
\usage{
dotList(input, spatial)
}
\arguments{
\item{input}{a dot data frame}

\item{spatial}{The spatial data frame}
}
\value{
A list containing detailed information on the arrays
}
\description{
Break the dot data frame into a list
}
\keyword{internal}
