% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load.R
\name{dotMatrix}
\alias{dotMatrix}
\title{Create numerical distances between dot elements}
\usage{
dotMatrix(input)
}
\arguments{
\item{input}{a dot data frame}
}
\value{
A matrix of the distance (in number of arrays) between pairs of arrays
}
\description{
Create numerical distances between dot elements
}
\keyword{internal}
