% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transform.R
\name{dotPaths}
\alias{dotPaths}
\title{Find arrays valid for efficiency calculation}
\usage{
dotPaths(input, mat)
}
\arguments{
\item{input}{A dot list}

\item{mat}{A dot distance matrix}
}
\value{
The input list, with an extra element for each array with valid efficiency peers
}
\description{
Find arrays valid for efficiency calculation
}
\keyword{internal}
