% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load.R
\name{dotPaths}
\alias{dotPaths}
\title{Find arrays valid for efficiency calculation}
\usage{
dotPaths(input, disregard.parallels)
}
\arguments{
\item{input}{A dot list}

\item{disregard.parallels}{Logical:  Should the presence of parallel arrays
invalidate potential efficiency peers? See the vignettes for more details.}
}
\value{
A list of the all array paths between each pair of arrays.
}
\description{
Find arrays valid for efficiency calculation
}
\keyword{internal}
