% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CJS.R
\name{dualArrayCJS}
\alias{dualArrayCJS}
\title{Calculate estimated last-array efficiency}
\usage{
dualArrayCJS(input)
}
\arguments{
\item{input}{A presence/absence matrix.}
}
\value{
A summary of the CJS results
}
\description{
The calculations are based on 'Using mark-recapture models to estimate survival from telemetry data' by Perry et al. 2012
}
\keyword{internal}
