% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CJS.R
\name{dualMatrix}
\alias{dualMatrix}
\title{Compile detection matrix for last array}
\usage{
dualMatrix(array, replicates, spatial, detections.list)
}
\arguments{
\item{replicates}{A list of arrays containing the respective replicate stations.}

\item{spatial}{The spatial data frame.}

\item{detections.list}{A list of the detections split by each target tag.}
}
\value{
A matrix of detection histories per tag for the last array.
}
\description{
Compile detection matrix for last array
}
\keyword{internal}
