% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CJS.R
\name{efficiencyMatrix}
\alias{efficiencyMatrix}
\title{Compile inter-array detection matrix}
\usage{
efficiencyMatrix(spatial, simple.movements)
}
\arguments{
\item{simple.movements}{A list of valid-only movements for each fish}
}
\value{
a matrix of detection histories per fish.
}
\description{
Compile inter-array detection matrix
}
\keyword{internal}
