% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CJS.R
\name{efficiencyMatrix}
\alias{efficiencyMatrix}
\title{Compile inter-array detection matrix}
\usage{
efficiencyMatrix(movements, arrays, paths, dotmat)
}
\arguments{
\item{movements}{A list of movements for each target tag, created by groupMovements.}
}
\value{
a matrix of detection histories per fish.
}
\description{
Compile inter-array detection matrix
}
\keyword{internal}
