% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CJS.R
\name{efficiencyMatrix}
\alias{efficiencyMatrix}
\title{Compile inter-array detection matrix}
\usage{
efficiencyMatrix(spatial, simple.movements, minimum.detections)
}
\arguments{
\item{simple.movements}{A list of valid-only movements for each fish}

\item{minimum.detections}{The minimum number of times a tag must have been recorded during the study period for it to be considered a true tag and not random noise.}
}
\value{
a matrix of detection histories per fish.
}
\description{
Compile inter-array detection matrix
}
\keyword{internal}
