% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CJS.R
\name{efficiencyMatrix}
\alias{efficiencyMatrix}
\title{Compile inter-array detection matrix}
\usage{
efficiencyMatrix(movements, arrays, paths, dotmat)
}
\arguments{
\item{movements}{The movements table.}

\item{arrays}{A list containing information for each array.}

\item{paths}{A list containing the shortest paths between arrays with distance > 1.}

\item{dotmat}{The matrix of distances between arrays.}
}
\value{
A matrix of detection histories per fish.
}
\description{
Compile inter-array detection matrix
}
\keyword{internal}
