% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{emergencyBreak}
\alias{emergencyBreak}
\title{Standard procedure when aborting}
\usage{
emergencyBreak(the.function.call)
}
\value{
No return value, called for side effects.
}
\description{
Wraps up the report in R's temporary folder before the function end.
}
\keyword{internal}
