% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{emptyMatrix}
\alias{emptyMatrix}
\title{Create a Template Distances Matrix}
\usage{
emptyMatrix(file = "spatial.csv")
}
\arguments{
\item{file}{The path to the file containing the spatial information.}
}
\value{
An empty matrix with the rows and columns required to
operate with the 'spatial.csv' file present in the current
work directory.
}
\description{
Creates an empty matrix based on the local 'spatial.csv' file and saves it to 'distances.csv' so the
user can manually fill it.
}
\details{
It is highly recommended to read the vignette regarding distances matrix before running this function.
You can find it by running \code{vignette('a-2_distances_matrix', 'actel')} or \code{browseVignettes('actel')}
}
\examples{
# This function requires the presence of a file with spatial information

# Let's create a dummy 'spatial.csv' file in R's temporary directory
write.csv(example.spatial, file = paste0(tempdir(), "/spatial.csv"))

# run emptyMatrix on the temporary spatial.csv file
emptyMatrix(paste0(tempdir(), "/spatial.csv"))

}
