% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unknownReceiversCheck.R
\name{emptyReceiversCheck}
\alias{emptyReceiversCheck}
\title{Check for unknown receivers in the raw detections}
\usage{
emptyReceiversCheck(spatial, detections)
}
\arguments{
\item{detections}{A dataframe with all the detections. Supplied by loadDetections.}
}
\description{
Check for unknown receivers in the raw detections
}
\keyword{internal}
