% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/z_examples.R
\name{example.detections}
\alias{example.detections}
\title{Example detection data}
\format{
A data frame with 14549 rows and 4 variables:
\describe{
\item{Timestamp}{The date and time of the detections}
\item{Receiver}{The ALS serial number}
\item{CodeSpace}{The code space of the detected tag}
\item{Signal}{The signal of the detected tag}
\item{Sensor.Value}{The data recorded by the sensor (dummy values)}
\item{Sensor.Unit}{The unit of the sensor data}
}
}
\source{
Data collected by the authors.
}
\description{
A dataset containing the detections of the deployed ALS, for the 60 fish.
}
\keyword{internal}
