% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/z_examples.R
\docType{data}
\name{example.results}
\alias{example.results}
\title{Example migration results}
\format{
A list of outputs from migration()
}
\source{
Data collected by the authors.
}
\usage{
example.results
}
\description{
A list with the results of a migration analysis ran on the example data.
Note: Many objects were trimmed to reduce package size. Use exampleWorkspace()
To run an analysis on the example data and obtain a full results object.
}
\keyword{internal}
