% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/z_examples.R
\name{example.spatial}
\alias{example.spatial}
\title{Example spatial data}
\format{
A data frame with 18 rows and 6 variables:
\describe{
\item{Station.name}{The name of the hydrophone station or release site}
\item{Latitude}{The latitude of the hydrophone station or release site in WGS84}
\item{Longitude}{The longitude of the hydrophone station or release site in WGS84}
\item{x}{The x coordinate of the hydrophone station or release site in EPSG 32632}
\item{y}{The y coordinate of the hydrophone station or release site in EPSG 32632}
\item{Array}{If documenting a hydrophone station, the array to which the station belongs.
If documenting a release site, the first array(s) where the fish is expected to be detected.}
\item{Section}{The Section to which the hydrophone station belongs (irrelevant for the release sites).}
\item{Type}{The type of spatial object (must be either 'Hydrophone' or 'Release')}
}
}
\source{
Data collected by the authors.
}
\description{
A dataset containing the positions of the deployed hydrophone stations and release sites.
}
\keyword{internal}
