% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/z_examples.R
\name{exampleWorkspace}
\alias{exampleWorkspace}
\title{Create an example Workspace}
\usage{
exampleWorkspace(example.spatial = example.spatial,
  example.biometrics = example.biometrics,
  example.detections = example.detections,
  example.deployments = example.deployments)
}
\arguments{
\item{example.spatial}{A dataset containing the positions of the deployed ALS and release site.}

\item{example.biometrics}{A dataset containing the positions of the deployed ALS and release site.}

\item{example.detections}{A dataset containing the detections of the deployed ALS, for the 60 fish.}

\item{example.deployments}{A dataset containing the deployed receivers.}
}
\description{
Creates a ready-to-run workspace with example data.
}
