% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/z_examples.R
\name{exampleWorkspace}
\alias{exampleWorkspace}
\title{Create an example Workspace}
\usage{
exampleWorkspace(spatial = example.spatial,
  biometrics = example.biometrics, detections = example.detections,
  deployments = example.deployments)
}
\arguments{
\item{spatial}{A dataset containing the positions of the deployed ALS and release site.}

\item{biometrics}{A dataset containing the positions of the deployed ALS and release site.}

\item{detections}{A dataset containing the detections of the deployed ALS, for the 60 fish.}

\item{deployments}{A dataset containing the deployed receivers.}
}
\description{
Creates a ready-to-run workspace with example data.
}
