% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/z_examples.R
\name{exampleWorkspace}
\alias{exampleWorkspace}
\title{Create an example Workspace}
\usage{
exampleWorkspace(
  spatial = example.spatial,
  biometrics = example.biometrics,
  detections = example.detections,
  deployments = example.deployments
)
}
\arguments{
\item{spatial, biometrics, detections, deployments}{Datasets provided with the package.}
}
\description{
Creates a ready-to-run workspace with example data.
}
