% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/z_examples.R
\name{exampleWorkspace}
\alias{exampleWorkspace}
\title{Deploy Example Data}
\usage{
exampleWorkspace(dir)
}
\arguments{
\item{dir}{The name of the target directory. Will be created if not present.}
}
\value{
No return value, called for side effects.
}
\description{
Creates a ready-to-run workspace with example data.
}
\examples{
\donttest{
# deploy a minimal dataset to try actel!
exampleWorkspace(paste0(tempdir(), "/exampleWorkspace"))
}

}
