% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load.R
\name{excludeTags}
\alias{excludeTags}
\title{Collect summary information on the tags detected but that are not part of the study.}
\usage{
excludeTags(input, exclude.tags)
}
\arguments{
\item{input}{list of detections}

\item{exclude.tags}{A vector of tags that should be excluded from the
detection data before any analyses are performed. Intended to be used if
stray tags from a different code space but with the same signal as a target
tag are detected in the study area.}
}
\value{
A list of detections for each tag that does not contain the excluded tags.
}
\description{
Collect summary information on the tags detected but that are not part of the study.
}
\keyword{internal}
