% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/explore.R
\name{explore}
\alias{explore}
\title{Explorative Analysis}
\usage{
explore(
  tz = NULL,
  datapack = NULL,
  max.interval = 60,
  minimum.detections,
  min.total.detections = 2,
  min.per.event = 1,
  start.time = NULL,
  stop.time = NULL,
  speed.method = c("last to first", "last to last"),
  speed.warning = NULL,
  speed.error = NULL,
  jump.warning = 2,
  jump.error = 3,
  inactive.warning = NULL,
  inactive.error = NULL,
  exclude.tags = NULL,
  override = NULL,
  report = FALSE,
  auto.open = TRUE,
  discard.orphans = FALSE,
  discard.first = NULL,
  save.detections = FALSE,
  GUI = c("needed", "always", "never"),
  save.tables.locally = FALSE,
  print.releases = TRUE,
  detections.y.axis = c("auto", "stations", "arrays")
)
}
\arguments{
\item{tz}{The time zone of the study area. Must match one of the values
present in \code{\link[base]{timezones}}.}

\item{datapack}{A data bundle pre-compiled through the function \code{\link{preload}}.
May be used to run actel analyses based on R objects, rather than input files.}

\item{max.interval}{The number of minutes that must pass between detections
for a new event to be created. Defaults to 60.}

\item{minimum.detections}{DEPRECATED. Please use the arguments min.total.detections
and min.per.event instead.}

\item{min.total.detections}{Minimum number of times a tag must have
been detected during the study period for the detections to be considered true
and not just random noise. Defaults to 2.}

\item{min.per.event}{Minimum number of detections an event must have to be
deemed valid. For analyses with both array and section events, a vector of
two values can be provided. If only one value is provided, the same threshold
applies for both types of events. Defaults to 1.}

\item{start.time}{Detection data prior to the timestamp set in
\code{start.time} (in YYYY-MM-DD HH:MM:SS format) is not considered during
the analysis.}

\item{stop.time}{Detection data posterior to the timestamp set in
\code{stop.time} (in YYYY-MM-DD HH:MM:SS format) is not considered during
the analysis.}

\item{speed.method}{Can take two forms: 'last to first' or 'last to last'.
If 'last to first' (default), the last detection on the previous array is matched
to the first detection on the target array to perform the calculations.
If 'last to last', the last detection on ´the previous array is matched to the
last detection on the target array to perform the calculations.}

\item{speed.warning}{If a tag moves at a speed equal or greater than
\code{speed.warning} (in metres per second), a warning is issued. If left
NULL (default), no warnings are issued. Must be equal to or lower than \code{speed.error}}

\item{speed.error}{If a tag moves at a speed equal or greater than
\code{speed.error} (in metres per second), user intervention is suggested.
If left NULL (default), user intervention is never suggested.}

\item{jump.warning}{If a tag crosses a number of arrays equal or greater
than \code{jump.warning} without being detected, a warning is issued. Defaults
to 2. To disable jump warnings, set to Inf. Must be equal to or lower than \code{jump.error}.}

\item{jump.error}{If a tag crosses a number of arrays equal or greater than
\code{jump.error} without being detected, user intervention is suggested.
Defaults to 3. To disable user intervention suggestions, set to Inf.}

\item{inactive.warning}{If a tag spends a number of days equal or greater
than \code{inactive.warning} in a given array at the tail of the respective
detections, a warning is issued. If left NULL (default), no warnings are
issued. Must be equal to or lower than \code{innactive.error}.}

\item{inactive.error}{If a tag spends a number of days equal or greater than
\code{inactive.error} in a given array at the tail of the respective
detections, user intervention is suggested. If left NULL (default), user
intervention is never suggested.}

\item{exclude.tags}{A vector of tags that should be excluded from the
detection data before any analyses are performed. Intended to be used if
stray tags from a different code space but with the same signal as a target
tag are detected in the study area.}

\item{override}{A vector of signals for which the user intends to manually
define which movement events are valid and invalid.}

\item{report}{Logical. Should an HTML report be created at the end of the
analysis? NOTE: Setting report to TRUE will generate an HTML file in the current
directory. Additionally, if auto.open = TRUE (default), the web browser will
automatically be launched to open the report once the function terminates.}

\item{auto.open}{Logical: Should the report be automatically opened once the
analysis is over? Defaults to TRUE. NOTE: If report = TRUE and auto.open = TRUE,
the web browser will automatically be launched to open the report once the
function terminates.}

\item{discard.orphans}{Logical: Should actel automatically discard
detections that do not fall within receiver deployment periods, or that
were recorded before the respective animals were released?}

\item{discard.first}{A threshold amount of time (in hours) that must pass after
release for the respective detections to be valid. Set to 0 to discard only
the release-to-first-detection calculations.}

\item{save.detections}{Logical: Should the processed detections be saved for
future runs?}

\item{GUI}{One of "needed", "always" or "never". If "needed", a new window is
opened to inspect the movements only when the movements table is too big to be
displayed in R's console. If "always", a graphical interface is always created
when the possibility to invalidate events emerges. If "never", a graphical
interface is never invoked. In this case, if the table to be displayed does
not fit in R's console, a temporary file will be saved and the user will be
prompted to open that file and examine it. Defaults to "needed".}

\item{save.tables.locally}{Logical: If a table must be temporarily stored into a file
for user inspection, should it be saved in the current working directory, or
in R's temporary folder?}

\item{print.releases}{Logical: Should the release sites be printed in the
study area diagrams?}

\item{detections.y.axis}{The type of y axis desired for the individual
detection plots. While the argument defaults to "auto", it can be hard-set
to one of "stations" or "arrays".}
}
\value{
A list containing:
\itemize{
\item \code{bio}: A copy of the biometrics input;
\item \code{detections}: A list containing all detections for each target tag;
\item \code{valid.detections}: A list containing the valid detections for each target tag;
\item \code{spatial}: A list containing the spatial information used during the analysis;
\item \code{deployments}: A data frame containing the deployments of each receiver;
\item \code{arrays}: A list containing the array details used during the analysis;
\item \code{movements}: A list containing all movement events for each target tag;
\item \code{valid.movements}: A list containing the valid movement events for each target tag;
\item \code{times}: A data frame containing all arrival times (per tag) at each array;
\item \code{rsp.info}: A list containing containing appendix information for the RSP package;
\item \code{dist.mat}: A matrix containing the distance matrix used in the analysis (if a valid
distance matrix was supplied)
}
}
\description{
\code{explore} allows you to quickly get a summary of your data. You can use
\code{explore} to get a general feel for the study results, and check if the
input files are behaving as expected. It is also a good candidate if you just
want to validate your detections for later use in other analyses.
}
\examples{
\donttest{
# Start by moving to a temporary directory
old.wd <- getwd()
setwd(tempdir())

# Deploy the example workspace
exampleWorkspace("explore_example")

# Move your R session into the example workspace
setwd("explore_example")

# run the explore analysis. Ensure the tz argument
# matches the time zone of the study area. For the
# example dataset, tz = "Europe/Copenhagen"
results <- explore(tz = "Europe/Copenhagen")

# to obtain an HTML report, run the analysis with report = TRUE

# return to original working directory
setwd(old.wd)
rm(old.wd)
}

}
\seealso{
\code{\link{migration}}, \code{\link{residency}}
}
