% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{fact}
\alias{fact}
\title{nearsq helper}
\usage{
fact(n)
}
\arguments{
\item{n}{number of plots to fit.}
}
\value{
An integer
}
\description{
Finds the largest x for which n \%\% x == 0
}
\details{
Obtained here: \url{https://stackoverflow.com/questions/32017327/calculate-the-optimal-grid-layout-dimensions-for-a-given-amount-of-plots-in-r}
}
\keyword{internal}
