% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load.R
\name{findDirectChains}
\alias{findDirectChains}
\title{Find all arrays linked to an array in a given direction}
\usage{
findDirectChains(array, array.list, direction = c("before", "after"))
}
\arguments{
\item{array}{The array for which to find peers}

\item{array.list}{An array list}

\item{direction}{The direction in which to expand the chain ("before" or "after")}
}
\value{
The array list with all linked arrays.
}
\description{
Find all arrays linked to an array in a given direction
}
\keyword{internal}
