% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load.R
\name{findPeers}
\alias{findPeers}
\title{Find efficiency peers for each array}
\usage{
findPeers(
  array,
  array.list,
  peer.direction = c("before", "after"),
  disregard.parallels
)
}
\arguments{
\item{array}{The array for which to find peers}

\item{array.list}{An array list}

\item{peer.direction}{The direction of peers to be found ("before" or "after")}
}
\value{
The array list with efficiency peers.
}
\description{
Find efficiency peers for each array
}
\keyword{internal}
