% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/residency.R
\name{findSecondsPerSection}
\alias{findSecondsPerSection}
\title{Calculate number of seconds at each location per day}
\usage{
findSecondsPerSection(res, day, the.range)
}
\arguments{
\item{res}{a residency list}

\item{day}{the day being analysed (a Date object)}

\item{the.range}{the first and last day for the specific fish}
}
\value{
A data frame containing the number of seconds spent at each location for a specific day
}
\description{
Calculate number of seconds at each location per day
}
\keyword{internal}
