% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CJS.R
\name{getDualMatrices}
\alias{getDualMatrices}
\title{Prepare intra-array matrices for selected arrays}
\usage{
getDualMatrices(replicates, CJS = NULL, spatial, detections.list)
}
\arguments{
\item{replicates}{A list of arrays containing the respective replicate stations.}

\item{CJS}{The overall CJS results.}

\item{spatial}{The spatial data frame.}

\item{detections.list}{A list of the detections split by each target tag.}
}
\value{
A list of dual matrices for each relevant array.
}
\description{
Prepare intra-array matrices for selected arrays
}
\keyword{internal}
