% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/z_deprecated.R
\name{getEstimate}
\alias{getEstimate}
\title{Get efficiency estimate for last array}
\usage{
getEstimate(spatial, detections.list, replicate)
}
\arguments{
\item{detections.list}{A list of the detections split by each target tag, created by splitDetections.}

\item{replicate}{A vector of station standard names to use as a replicate of the last array, for efficiency estimations.}
}
\value{
the modified CJS model for the last array
}
\description{
Get efficiency estimate for last array
}
\keyword{internal}
