% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{getTimes}
\alias{getTimes}
\title{Extract time stamp of valid entry or exit in each array}
\usage{
getTimes(
  movements,
  spatial,
  type = c("arrival", "departure"),
  events = c("first", "all", "last")
)
}
\arguments{
\item{movements}{The list of valid movements.}

\item{spatial}{The list of spatial objects.}

\item{type}{The point to be recorded: one of "arrival" or "departure".}

\item{events}{The number of events to record. if "one" and type is "arrival", the very first arrival is returned;
if "one" and type is "departure", the very last departure is returned.}
}
\value{
A data frame with the timestamps for each fish (rows) and array (columns)
}
\description{
Extract time stamp of valid entry or exit in each array
}
\keyword{internal}
