% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/residency.R
\name{globalRatios}
\alias{globalRatios}
\title{Calculate number/percentage of tags at each location for each day}
\usage{
globalRatios(positions, section.order)
}
\arguments{
\item{positions}{a positions table, supplied by dailyPositions}
}
\value{
A list containing:
\itemize{
\item \code{absolutes}: A data frame containing the absolute number of tags at each location per day,
\item \code{percentages}: A data frame containing the percentage of tag relative to the total
number of active tags at each location per day.
}
}
\description{
Calculate number/percentage of tags at each location for each day
}
\keyword{internal}
