% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check.R
\name{graphicalInvalidate}
\alias{graphicalInvalidate}
\title{Opens a new winder that allows the user to determine movement event invalidity}
\usage{
graphicalInvalidate(moves, fish, trigger)
}
\arguments{
\item{moves}{The movements table}

\item{fish}{The fish being analysed}

\item{trigger}{The warning/message that triggered the interaction}
}
\value{
the movement table with valid/invalid notes
}
\description{
Opens a new winder that allows the user to determine movement event invalidity
}
\keyword{internal}
