% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/user_interaction.R
\name{graphicalInvalidate}
\alias{graphicalInvalidate}
\title{Opens a new window that allows the user to determine movement event invalidity}
\usage{
graphicalInvalidate(detections, moves, tag, trigger)
}
\arguments{
\item{detections}{The detections data.frame for a specific tag.}

\item{tag}{The tag being analysed.}

\item{trigger}{The message/warning that triggered the interaction}
}
\value{
A data frame with the movement events for the target tag and an updated 'Valid' column.
}
\description{
Opens a new window that allows the user to determine movement event invalidity
}
\keyword{internal}
