% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/movements.R
\name{groupMovements}
\alias{groupMovements}
\title{Group movements}
\usage{
groupMovements(
  detections.list,
  bio,
  spatial,
  speed.method,
  max.interval,
  tz,
  dist.mat
)
}
\arguments{
\item{detections.list}{A list of the detections split by each target tag.}

\item{bio}{The biometrics table.}

\item{spatial}{The spatial list.}

\item{speed.method}{Can take two forms: 'last to first' or 'last to last'.
If 'last to first' (default), the last detection on the previous array is matched
to the first detection on the target array to perform the calculations.
If 'last to last', the last detection on ´the previous array is matched to the
last detection on the target array to perform the calculations.}

\item{max.interval}{The number of minutes that must pass between detections
for a new event to be created. Defaults to 60.}

\item{tz}{The time zone of the study area. Must match one of the values
present in \code{\link[base]{timezones}}.}

\item{dist.mat}{A matrix of the distances between the deployed ALS.}
}
\value{
A list containing the movement events for each tag.
}
\description{
Crawls trough the detections of each tag and groups them based on ALS arrays and time requirements.
}
\keyword{internal}
