% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helperFunctions.R
\name{havingIP}
\alias{havingIP}
\title{Check internet connection}
\usage{
havingIP()
}
\description{
For an unknown reason, R will crash if it attemps to open the html report in a pc with no internet connection.
The report opens, but the progress freezes and R eventually crashes. Can be removed once figured out why this
is happening.
}
\keyword{internal}
