% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CJS.R
\name{includeMissing}
\alias{includeMissing}
\title{Include tags that were never detected}
\usage{
includeMissing(x, status.df)
}
\arguments{
\item{x}{an efficiency matrix}

\item{status.df}{The main results table.}
}
\value{
A matrix of detection histories per tag, including tags that were never detected.
}
\description{
Include tags that were never detected
}
\keyword{internal}
