% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check.R
\name{includeUnknownReceiver}
\alias{includeUnknownReceiver}
\title{Temporarily include missing receivers in the spatial object}
\usage{
includeUnknownReceiver(spatial, deployments, unknown.receivers)
}
\arguments{
\item{spatial}{A list of spatial objects in the study area.}

\item{deployments}{a list of deployments.}

\item{unknown.receivers}{serial number of the receivers to be included.}
}
\value{
A list containing:
\itemize{
\item \code{spatial}: A list containing the spatial elements of the study, with data for the unknown receivers, if relevant.
\item \code{deployments}: A list containing the receiver deployments, with data for the unknown receivers, if relevant.
}
}
\description{
Temporarily include missing receivers in the spatial object
}
\keyword{internal}
