% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check.R
\name{invalidateEvents}
\alias{invalidateEvents}
\title{Allow the user to determine a given movement event invalid}
\usage{
invalidateEvents(movements, fish)
}
\arguments{
\item{movements}{The movements table}

\item{fish}{The fish being analysed}
}
\value{
the movement table with valid/invalid notes
}
\description{
Allow the user to determine a given movement event invalid
}
\keyword{internal}
