% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/user_interaction.R
\name{invalidateEvents}
\alias{invalidateEvents}
\title{Allow the user to determine a given movement event invalid}
\usage{
invalidateEvents(displayed.moves, all.moves, detections, tag, GUI)
}
\arguments{
\item{tag}{The tag being analysed}

\item{GUI}{One of "needed", "always" or "never". If "needed", a new window is
opened to inspect the movements only if the movements table is too big to be
displayed in R's console. If "always", a graphical interface is always created
when the possibility to invalidate events emerges. If "never", a graphical
interface is never invoked. In this case, if the table to be displayed does
not fit in R's console, a temporary file will be saved and the user will be
prompted to open and examine that file. Defaults to "needed".}
}
\value{
A data frame with the movement events for the target tag and an updated 'Valid' column.
}
\description{
Allow the user to determine a given movement event invalid
}
\keyword{internal}
