% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assembleEfficiency.R
\name{lastMatrix}
\alias{lastMatrix}
\title{Compile detection matrix for last array}
\usage{
lastMatrix(spatial, detections.list, replicate)
}
\arguments{
\item{spatial}{A list of the spatial objects, created by assembleSpatial.}

\item{detections.list}{A list of the detections split by each target tag, created by splitDetections.}

\item{replicate}{The Standard.Name of the stations to use as a replicate of the last array to estimate last detection efficiency.}
}
\value{
a matrix of detection histories per fish for the last array.
}
\description{
Compile detection matrix for last array
}
\keyword{internal}
