% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CJSFunctions.R
\name{lastMatrix}
\alias{lastMatrix}
\title{Compile detection matrix for last array}
\usage{
lastMatrix(spatial, detections.list, replicate)
}
\arguments{
\item{detections.list}{A list of the detections split by each target tag, created by splitDetections.}
}
\value{
a matrix of detection histories per fish for the last array.
}
\description{
Compile detection matrix for last array
}
\keyword{internal}
