% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{listToTable}
\alias{listToTable}
\title{Convert a list of data frames with identical columns into a single data frame}
\usage{
listToTable(input)
}
\arguments{
\item{input}{a list of data frames}
}
\value{
a data frame with all lists combined
}
\description{
Convert a list of data frames with identical columns into a single data frame
}
