% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load.R
\name{loadBio}
\alias{loadBio}
\title{Import biometrics}
\usage{
loadBio(file)
}
\arguments{
\item{file}{an input file with biometric data.}
}
\value{
The biometrics table
}
\description{
Import biometrics
}
\keyword{internal}
