% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load.R
\name{loadBio}
\alias{loadBio}
\title{Load Biometrics file}
\usage{
loadBio(file, tz)
}
\arguments{
\item{file}{an input file with biometric data.}

\item{tz}{The time zone of the study area. Must match one of the values
present in \code{\link[base]{timezones}}.}
}
\value{
The biometrics table
}
\description{
Load Biometrics file
}
\keyword{internal}
