% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadFunctions.R
\name{loadDeployments}
\alias{loadDeployments}
\title{Load deployments file and Check the structure}
\usage{
loadDeployments(file, tz.study.area)
}
\arguments{
\item{file}{an input file with spatial data.}
}
\value{
The deployments dataframe
}
\description{
Load deployments file and Check the structure
}
\keyword{internal}
