% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load.R
\name{loadDeployments}
\alias{loadDeployments}
\title{Load deployments file and Check the structure}
\usage{
loadDeployments(input, tz)
}
\arguments{
\item{input}{an input data frame or file with deployment data.}

\item{tz}{The time zone of the study area. Must match one of the values
present in \code{\link[base]{timezones}}.}
}
\value{
A data frame with the deployments information
}
\description{
Load deployments file and Check the structure
}
\keyword{internal}
