% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load.R
\name{loadDistances}
\alias{loadDistances}
\title{Load distances matrix}
\usage{
loadDistances(input = "distances.csv", spatial)
}
\arguments{
\item{input}{Either a path to a csv file containing a distances matrix,
or an R object containing a distances matrix.}

\item{spatial}{A list of spatial objects in the study area.}
}
\value{
A matrix of the distances (in metres) between stations
(if a 'distances.csv' is present)
}
\description{
Load distances matrix
}
\keyword{internal}
