% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load.R
\name{loadDot}
\alias{loadDot}
\title{Load spatial.dot}
\usage{
loadDot(string = NULL, input = NULL, spatial, sections)
}
\arguments{
\item{input}{The name of the dot file}

\item{spatial}{A spatial data frame}

\item{sections}{The sections in which the study was divided. Must be coincident with the names given to the ALS arrays. (i.e. if an array is 'River1', then the respective section is 'River')}
}
\value{
The dot list
}
\description{
Load spatial.dot
}
\keyword{internal}
