% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load.R
\name{loadDot}
\alias{loadDot}
\title{Load spatial.dot}
\usage{
loadDot(
  string = NULL,
  input = NULL,
  spatial,
  sections = NULL,
  disregard.parallels
)
}
\arguments{
\item{string}{A string of dot connections.}

\item{input}{The name of a file containing dot connections.}

\item{spatial}{A spatial data frame.}

\item{sections}{The sections in which the study area is divided. Must be
coincident with the names given to the arrays. See the vignettes for more
details.}

\item{disregard.parallels}{Logical:  Should the presence of parallel arrays
invalidate potential efficiency peers? See the vignettes for more details.}
}
\value{
A list containing:
\itemize{
\item \code{dot}: A data frame containing the connections between arrays
\item \code{arrays}: A list containing detailed information on the arrays
\item \code{dotmat}: A matrix of the distance (in number of arrays) between pairs of arrays
\item \code{paths}: A list of the all array paths between each pair of arrays.
}
}
\description{
Load spatial.dot
}
\keyword{internal}
