% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load.R
\name{loadDot}
\alias{loadDot}
\title{Load spatial.dot}
\usage{
loadDot(
  string = NULL,
  input = NULL,
  spatial,
  disregard.parallels,
  preloading = FALSE
)
}
\arguments{
\item{string}{A string of dot connections.}

\item{input}{The name of a file containing dot connections.}

\item{spatial}{A spatial data frame.}

\item{disregard.parallels}{Logical:  Should the presence of parallel arrays
invalidate potential efficiency peers? See the vignettes for more details.}

\item{preloading}{Logical: Is this function being called from preload()?}
}
\value{
A list containing:
\itemize{
\item \code{dot}: A data frame containing the connections between arrays
\item \code{arrays}: A list containing detailed information on the arrays
\item \code{dotmat}: A matrix of the distance (in number of arrays) between pairs of arrays
\item \code{paths}: A list of the all array paths between each pair of arrays.
}
}
\description{
Load spatial.dot
}
\keyword{internal}
