% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load.R
\name{loadSpatial}
\alias{loadSpatial}
\title{Load Spatial File}
\usage{
loadSpatial(file = "spatial.csv")
}
\arguments{
\item{file}{an input file with spatial data in the actel format.}
}
\value{
A data frame with the spatial information present in 'spatial.csv' and the Standard.name column.
}
\description{
Loads a spatial file prepared for actel and appends the Standard.name column. Additionally,
performs a series of quality checks on the contents of the target file.
}
\examples{
# This function requires the presence of a file with spatial information

# Let's create a dummy 'spatial.csv' file in R's temporary directory
write.csv(example.spatial, file = paste0(tempdir(), "/spatial.csv"), row.names = FALSE)

# run loadSpatial on the temporary spatial.csv file
loadSpatial(file = paste0(tempdir(), '/spatial.csv'))

}
