% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load.R
\name{loadSpatial}
\alias{loadSpatial}
\title{Load Spatial file and Check the structure}
\usage{
loadSpatial(file = "spatial.csv", verbose = FALSE)
}
\arguments{
\item{file}{an input file with spatial data.}

\item{verbose}{Logical: If TRUE, the appendTo function is enabled.}
}
\value{
The spatial dataframe
}
\description{
Load Spatial file and Check the structure
}
