% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load.R
\name{loadStudyData}
\alias{loadStudyData}
\title{Load, structure and check the inputs}
\usage{
loadStudyData(tz.study.area, override = NULL, start.timestamp,
  end.timestamp, sections = NULL, exclude.tags)
}
\arguments{
\item{tz.study.area}{The time-zone of the study area. Necessary to convert the ALS time data, which is in UTC.}

\item{override}{A list of tags for which the user intends to manually define entering and leaving points for each study section.}

\item{start.timestamp}{Detection data prior to this date is not analysed. Improves processing time when loading large amounts of detection data.}

\item{end.timestamp}{Detection data posterior to this date is not analysed. Improves processing time when loading large amounts of detection data.}

\item{sections}{The sections in which the study was divided. Must be coincident with the names given to the ALS arrays. (i.e. if an array is 'River1', then the respective section is 'River')}

\item{exclude.tags}{A list of tags that should be excluded from the detection data before any analyses are performed. Intended to be used if stray tags from a different code space but with the same signal as a target tag are detected in the study area.}
}
\value{
a list of objects to be used for data analysis
}
\description{
Load, structure and check the inputs
}
\keyword{internal}
