% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CJS.R
\name{mbGroupCJS}
\alias{mbGroupCJS}
\title{Calculate CJS for each group for each array}
\usage{
mbGroupCJS(m, status.df, fixed.efficiency = NULL)
}
\arguments{
\item{m}{The presence/absence matrices.}

\item{status.df}{The main results table.}

\item{fixed.efficiency}{A vector of fixed efficiency estimates from a more complete CJS model.}
}
\value{
A list of CJS results
}
\description{
Calculate CJS for each group for each array
}
\keyword{internal}
