% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CJS.R
\name{mbGroupCJS}
\alias{mbGroupCJS}
\title{Calculate CJS for each group for each array}
\usage{
mbGroupCJS(mat, status.df, fixed.efficiency = NULL)
}
\arguments{
\item{status.df}{The main results table.}

\item{fixed.efficiency}{A vector of fixed efficiency estimates [0, 1]. \code{length(fixed.efficiency)} must match \code{ncol(input)}.}
}
\value{
A list containing CJS results split by group.
}
\description{
Calculate CJS for each group for each array
}
\keyword{internal}
