% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CJS.R
\name{mbSplitCJS}
\alias{mbSplitCJS}
\title{Calculate CJS for each group.release combination}
\usage{
mbSplitCJS(mat, fixed.efficiency = NULL)
}
\arguments{
\item{fixed.efficiency}{A vector of fixed efficiency estimates [0, 1]. \code{length(fixed.efficiency)} must match \code{ncol(input)}.}
}
\value{
A list containing CJS results split by group and release site
}
\description{
Calculate CJS for each group.release combination
}
\keyword{internal}
