% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/migration.R
\name{migration}
\alias{migration}
\title{Migration Analysis}
\usage{
migration(
  path = NULL,
  sections,
  success.arrays = NULL,
  minimum.detections = 2,
  maximum.time = 60,
  max.interval = 60,
  speed.method = c("last to first", "first to first"),
  speed.warning = NULL,
  speed.error = NULL,
  if.last.skip.section = TRUE,
  tz.study.area = NULL,
  tz = NULL,
  start.time = NULL,
  start.timestamp = NULL,
  stop.time = NULL,
  end.timestamp = NULL,
  report = TRUE,
  override = NULL,
  exclude.tags = NULL,
  replicates = NULL,
  disregard.parallels = TRUE,
  jump.warning = 2,
  jump.error = 3,
  inactive.warning = NULL,
  inactive.error = NULL,
  debug = FALSE
)
}
\arguments{
\item{path}{Path to the folder containing the data. If left NULL (default),
the analysis runs in the current folder.}

\item{sections}{The sections in which the study area is divided. Must be
coincident with the names given to the arrays. See the vignettes for more
details.}

\item{success.arrays}{The arrays that mark the end of the study area. If a
fish crosses one of these arrays, it is considered to have successfully
migrated through the study area.}

\item{minimum.detections}{For tags with only one movement event, defines the
minimum number of times a tag must have been recorded during the study
period for it to be considered true detections and not random noise.
Defaults to 2.}

\item{maximum.time}{DEPRECATED. See \code{max.interval}.}

\item{max.interval}{The number of minutes that must pass between detections
for a new event to be created. Defaults to 60.}

\item{speed.method}{Can take two forms: 'last to first' or 'first to first'.
If 'last to first' (default), the last detection on a given array is matched
to the first detection on the next array to perform the calculations.
If 'first to first', the first detection on a given array is matched to the
first detection on the next array to perform the calculations.}

\item{speed.warning}{If a fish moves at a speed equal or greater than
\code{speed.warning} (in metres per second), a warning is issued. If left
NULL (default), no warnings are issued.}

\item{speed.error}{If a fish moves at a speed equal or greater than
\code{speed.error} (in metres per second), user intervention is suggested.
If left NULL (default), user intervention is never suggested.}

\item{if.last.skip.section}{Logical: Should a fish detected at the last array
of a given section be considered to have disappeared in the next section?}

\item{tz.study.area}{DEPRECATED. See \code{tz}.}

\item{tz}{The time zone of the study area. Must match one of the values
present in \code{\link[base]{timezones}}.}

\item{start.time}{Detection data prior to the timestamp set in
\code{start.time} (in YYYY-MM-DD HH:MM:SS format) is not considered during
the analysis.}

\item{start.timestamp}{DEPRECATED. See \code{start.time}.}

\item{stop.time}{Detection data posterior to the timestamp set in
\code{stop.time} (in YYYY-MM-DD HH:MM:SS format) is not considered during
the analysis.}

\item{end.timestamp}{DEPRECATED See argument \code{stop.time}}

\item{report}{Logical. Should an HTML report be created at the end of the
analysis?}

\item{override}{A vector of tags for which the user intends to manually
define which movement events are valid and invalid.}

\item{exclude.tags}{A vector of tags that should be excluded from the
detection data before any analyses are performed. Intended to be used if
stray tags from a different code space but with the same signal as a target
tag are detected in the study area.}

\item{replicates}{A list containing, for each array to which intra-array
efficiency is to be calculated: The standard names of the stations to be
used as a replicate. See the vignettes for more details.}

\item{disregard.parallels}{Logical:  Should the presence of parallel arrays
invalidate potential efficiency peers? See the vignettes for more details.}

\item{jump.warning}{If a fish crosses a number of arrays equal or greater
than \code{jump.error} without being detected, a warning is issued. If left
NULL (default), no warnings are issued.}

\item{jump.error}{If a fish crosses a number of arrays equal or greater than
\code{jump.error} without being detected, user intervention is suggested.
If left NULL (default), user intervention is never suggested.}

\item{inactive.warning}{If a fish spends a number of days equal or greater
than \code{inactive.error} in a given array at the tail of the respective
detections, a warning is issued. If left NULL (default), no warnings are
issued.}

\item{inactive.error}{If a fish spends a number of days equal or greater than
\code{inactive.error} in a given array at the tail of the respective
detections, user intervention is suggested. If left NULL (default), user
intervention is never suggested.}

\item{debug}{Logical: Should temporary files be kept at the end of the
analysis?}
}
\value{
A list containing:
\itemize{
\item \code{detections}: All detections for each target fish;
\item \code{valid.detections}: Valid detections for each target fish;
\item \code{spatial}: The spatial information used during the analysis;
\item \code{deployments}: The deployments of each receiver;
\item \code{arrays}: The array details used during the analysis;
\item \code{movements}: All movement events for each target fish;
\item \code{valid.movements}: Valid movemenet events for each target fish;
\item \code{section.movements}: Valid section shifts for each target fish;
\item \code{status.df}: Summary information for each fish, including the
following columns:
\itemize{
\item \emph{Time.until.[section]}: Time spent between leaving one
section and reaching the next.
\item \emph{Speed.to.[section]}: Average speed from one section to the
next (if a distance matrix is provided)
\item \emph{First.station.[section]}: Standard name of the first station
where the fish was detected in a given section
\item \emph{Arrived.[section]}: Arrival time at a given section
\item \emph{Time.in.[section]}: Total time spent within a given section
\item \emph{Last.station.[section]}: Standard name of the last station
where the fish was detected in a given section
\item \emph{Left.[section]}: Departure time from a givem section
\item \emph{Very.last.array}: Last array where the fish was detected
\item \emph{Status}: Fate assigned to the fish
\item \emph{Valid.detections}: Number of valid detections
\item \emph{Invalid.detections}: Number of invalid detections
\item \emph{Backwards.movements}: Number of backward movement events
\item \emph{Max.cons.back.moves}: Longest successive backwards movements
\item \emph{P.type}: Type of processing:
\itemize{
\item 'Skipped' if no data was found for the fish,
\item 'Auto' if no user interaction was required,
\item 'Manual' if user interaction was suggested and the user made
changes to the validity of the events,
\item 'Overridden' if the user listed the fish in the
\code{override} argument.
}
\item \emph{Comments}: Comments left by the user during the analysis
}
\item \code{section.overview}: Summary table of the number of fish that
disappeared in each section;
\item \code{array.overview}: Number of known and estimated fish to have
passed through each array;
\item \code{matrices}: CJS matrices used for the efficiency calculations;
\item \code{overall.CJS}: Results of the inter-array CJS calculations;
\item \code{intra.array.CJS}: Results of the intra-array CJS calculations;
\item \code{times}: All arrival times (per fish) at each array;
\item \code{rsp.info}: Appendix information for the RSP package;
\item \code{dist.mat}: The distance matrix used in the analysis (if a valid
distance matrix was supplied)
}
}
\description{
The \code{migration} analysis runs the same initial checks as \code{explore},
but on top of it, it analyses the fish behaviour. By selecting the arrays
that lead to success, you can define whether or not your fish survived the
migration. Additional plots help you find out if some fish has been acting
odd. Multiple options allow you to tweak the analysis to fit your study
perfectly.
}
