% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/movements.R
\name{movementSpeeds}
\alias{movementSpeeds}
\title{Calculate time and speed}
\usage{
movementSpeeds(movements, speed.method, dist.mat, silent = TRUE)
}
\arguments{
\item{movements}{A list of movements for each target tag, created by groupMovements.}

\item{speed.method}{Can take two forms: 'last to first' or 'first to first'.
If 'last to first' (default), the last detection on a given array is matched
to the first detection on the next array to perform the calculations.
If 'first to first', the first detection on a given array is matched to the
first detection on the next array to perform the calculations.}

\item{dist.mat}{A matrix of the distances between the deployed ALS.}

\item{silent}{logical: If TRUE, debug messages are issued (only works within actel)}
}
\value{
The movement data frame with time and speed calculations
}
\description{
Triggers movementTimes and also calculates the speed between events.
}
\keyword{internal}
