% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/movements.R
\name{movementSpeeds}
\alias{movementSpeeds}
\title{Calculate time and speed}
\usage{
movementSpeeds(movements, speed.method, dist.mat)
}
\arguments{
\item{movements}{A list of movements for each target tag, created by groupMovements.}

\item{speed.method}{Can take two forms: 'last to first' or 'last to last'.
If 'last to first' (default), the last detection on the previous array is matched
to the first detection on the target array to perform the calculations.
If 'last to last', the last detection on ´the previous array is matched to the
last detection on the target array to perform the calculations.}

\item{dist.mat}{A matrix of the distances between the deployed ALS.}
}
\value{
The movement data frame with speed calculations for the target fish.
}
\description{
Triggers movementTimes and also calculates the speed between events.
}
\keyword{internal}
