% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/movements.R
\name{movementTimes}
\alias{movementTimes}
\title{Calculate movement times}
\usage{
movementTimes(movements, type = c("array", "section"))
}
\arguments{
\item{movements}{the movement dataframe for the tag being analysed.}

\item{type}{The type of movements being analysed. One of "array" or "section".}
}
\value{
The movement data frame with time calculations for the target tag.
}
\description{
Determines the duration of an event and the time from an event to the next.
}
\keyword{internal}
