% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CJS.R
\name{oneWayMoves}
\alias{oneWayMoves}
\title{Trim movements table to contain only uni-directional movements}
\usage{
oneWayMoves(movements, arrays)
}
\arguments{
\item{arrays}{A list containing information for each array.}
}
\value{
A data frame with the uni-directional movements for the target tags.
}
\description{
Trim movements table to contain only uni-directional movements
}
\keyword{internal}
