% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check.R
\name{overrideValidityChecks}
\alias{overrideValidityChecks}
\title{Skips all validity checks for a fish and allows the user to freely invalidate events}
\usage{
overrideValidityChecks(moves, fish)
}
\arguments{
\item{moves}{The movements table}

\item{fish}{The fish being analysed}
}
\description{
Skips all validity checks for a fish and allows the user to freely invalidate events
}
\keyword{internal}
