% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/printFunctions.R
\name{prepareAdditions}
\alias{prepareAdditions}
\title{Summarise places were more fish were released}
\usage{
prepareAdditions(group.CJS, progression)
}
\arguments{
\item{group.CJS}{a list of CJS's for each group, with release sites merged}

\item{progression}{the progression dataframe, provided by compileProgressionDataFrame}
}
\value{
The additions dataframe, or NULL if additions should be aborted.
}
\description{
The "additions" will give the labels
}
\keyword{internal}
