% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{prepareTotals}
\alias{prepareTotals}
\title{Compile the totals at each entry point per group}
\usage{
prepareTotals(group.CJS)
}
\value{
the totals data frame
}
\description{
Compile the totals at each entry point per group
}
\keyword{internal}
