% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{printArrayOverview}
\alias{printArrayOverview}
\title{print Rmd fragment for inclusion in the report}
\usage{
printArrayOverview(array.overview)
}
\arguments{
\item{array.overview}{a list of absolute detection numbers for each group}
}
\value{
An rmd string to be included in the report.
}
\description{
print Rmd fragment for inclusion in the report
}
\keyword{internal}
