% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/printFunctions.R
\name{printBiometrics}
\alias{printBiometrics}
\title{Print biometric graphics}
\usage{
printBiometrics(bio)
}
\value{
string to be included in printRmd
}
\description{
Searches for columns containing biometric data and prints graphics per fish group
}
\keyword{internal}
