% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{printBiometrics}
\alias{printBiometrics}
\title{Print biometric graphics}
\usage{
printBiometrics(bio)
}
\arguments{
\item{bio}{A table with the tags and biometrics of the studied animals.}
}
\value{
A string of file locations in rmd syntax, to be included in printRmd
}
\description{
Searches for columns containing biometric data and prints graphics per animal group
}
\keyword{internal}
