% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{printCircular}
\alias{printCircular}
\title{Print circular graphics for each array}
\usage{
printCircular(times, bio, suffix = NULL)
}
\arguments{
\item{times}{a (list of) circular object(s)}

\item{bio}{A table with the tags and biometrics of the studied animals.}
}
\value{
A string of file locations in rmd syntax, to be included in printRmd
}
\description{
Prints the time of first entry point on each of the arrays. Contains functions adapted from the circular R package.
}
\details{
For more details about the original function, visit the circular package homepage at \url{https://github.com/cran/circular}
}
\keyword{internal}
