% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{printDot}
\alias{printDot}
\title{Print DOT diagram}
\usage{
printDot(dot, spatial, print.releases)
}
\arguments{
\item{dot}{a dot data frame}

\item{print.releases}{Logical: Should the release sites be printed in the
study area diagrams?}
}
\value{
No return value, called to plot and save graphic.
}
\description{
Print DOT diagram
}
\keyword{internal}
