% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/printFunctions.R
\name{printDotplots}
\alias{printDotplots}
\title{Print dotplots}
\usage{
printDotplots(status.df, invalid.dist)
}
\arguments{
\item{status.df}{A dataframe with all the final data for each fish, created by assembleOutput.}

\item{invalid.dist}{Wether or not the distance matrix suplied is valid for the study area.}
}
\description{
Prints dotplots of multiple variables.
}
\keyword{internal}
