% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/printFunctions.R
\name{printDotplots}
\alias{printDotplots}
\title{Print dotplots}
\usage{
printDotplots(status.df, invalid.dist)
}
\arguments{
\item{invalid.dist}{Whether or not the distance matrix supplied is valid for the study area.}
}
\description{
Prints dotplots of multiple variables.
}
\keyword{internal}
