% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{printDotplots}
\alias{printDotplots}
\title{Print dotplots}
\usage{
printDotplots(status.df, valid.dist)
}
\arguments{
\item{valid.dist}{Logical: Is a valid distances matrix being used?}
}
\value{
No return value, called to plot and save graphic.
}
\description{
Prints dotplots of multiple variables.
}
\keyword{internal}
