% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{printEfficiency}
\alias{printEfficiency}
\title{Print efficiency fragment}
\usage{
printEfficiency(intra.CJS, type = c("migration", "residency"))
}
\arguments{
\item{intra.CJS}{The output of the getEstimate calculations.}
}
\description{
Prints the ALS inter-array efficiency for inclusion in printRmd.
}
\keyword{internal}
