% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/printFunctions.R
\name{printEfficiency}
\alias{printEfficiency}
\title{Print efficiency tables}
\usage{
printEfficiency(efficiency)
}
\arguments{
\item{efficiency}{A list containing the intersection and interarray efficiency. Supplied by assembleEfficiency.}
}
\description{
Prints the ALS intersection and interarray efficiency in LaTeX format for inclusion in printLaTeX.
}
\keyword{internal}
