% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{printEfficiency}
\alias{printEfficiency}
\title{Print efficiency fragment}
\usage{
printEfficiency(overall.CJS, last.array.results)
}
\arguments{
\item{overall.CJS}{a single CJS with all the groups and release sites merged}

\item{last.array.results}{The output of the getEstimate calculations.}
}
\description{
Prints the ALS inter-array efficiency for inclusion in printRmd.
}
\keyword{internal}
