% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/explore.R
\name{printExploreRmd}
\alias{printExploreRmd}
\title{Print Rmd report}
\usage{
printExploreRmd(
  override.fragment,
  biometric.fragment,
  individual.plots,
  circular.plots,
  sensor.plots,
  spatial,
  deployments,
  detections,
  valid.detections,
  detections.y.axis
)
}
\arguments{
\item{override.fragment}{Rmarkdown string specifying the type of report for the header.}

\item{biometric.fragment}{Rmarkdown string specifying the biometric graphics drawn.}

\item{individual.plots}{Rmarkdown string specifying the name of the individual plots.}

\item{circular.plots}{Rmarkdown string specifying the name of the circular plots.}

\item{sensor.plots}{Rmarkdown string specifying the name of the sensor plots.}

\item{detections}{All the detections used in the study}

\item{valid.detections}{The valid detections used in the study}
}
\value{
No return value, called for side effects.
}
\description{
Creates a Rmd report and converts it to html.
}
\keyword{internal}
