% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{printGlobalRatios}
\alias{printGlobalRatios}
\title{print the distribution of fish per location}
\usage{
printGlobalRatios(global.ratios, daily.ratios, sections)
}
\arguments{
\item{global.ratios}{the global ratios}

\item{daily.ratios}{the daily ratios}

\item{sections}{The sections in which the study area is divided. Must be
coincident with the names given to the arrays. See the vignettes for more
details.}
}
\value{
No return value, called to plot and save graphic.
}
\description{
print the distribution of fish per location
}
\keyword{internal}
