% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{printGlobalRatios}
\alias{printGlobalRatios}
\title{print the distribution of tags per location}
\usage{
printGlobalRatios(global.ratios, group.ratios, time.ratios, spatial, rsp.info)
}
\arguments{
\item{global.ratios}{the global ratios}

\item{group.ratios}{the global ratios}

\item{time.ratios}{the daily ratios}
}
\value{
No return value, called to plot and save graphic.
}
\description{
print the distribution of tags per location
}
\keyword{internal}
