% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{printIndividualResidency}
\alias{printIndividualResidency}
\title{print the individual locations per day}
\usage{
printIndividualResidency(ratios, global.ratios, spatial, rsp.info)
}
\arguments{
\item{ratios}{the daily ratios}

\item{spatial}{The spatial list.}
}
\value{
A string of file locations in rmd syntax, to be included in printRmd
}
\description{
print the individual locations per day
}
\keyword{internal}
