% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/printFunctions.R
\name{printIndividuals}
\alias{printIndividuals}
\title{Print individual graphics}
\usage{
printIndividuals(redraw, detections.list, status.df, tz.study.area)
}
\arguments{
\item{redraw}{When report is set to TRUE, controls if files created in a previous run of the function should be overriden.}

\item{detections.list}{A list of the detections split by each target tag, created by splitDetections.}

\item{status.df}{A dataframe with all the final data for each fish, created by assembleOutput.}

\item{tz.study.area}{The timezone of the study area. Necessary to conver the ALS time data, which is in UTC.}
}
\value{
LaTeX string to be included in .printLatex
}
\description{
Prints the individual detections for each fish, overlaying the points in time considered crucial during the analysis.
}
\keyword{internal}
