% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{printIndividuals}
\alias{printIndividuals}
\title{Print individual graphics}
\usage{
printIndividuals(
  detections.list,
  movements,
  valid.movements,
  spatial,
  status.df = NULL,
  rsp.info,
  type = c("stations", "arrays"),
  extension = "png"
)
}
\arguments{
\item{detections.list}{A list of the detections split by each target tag, created by splitDetections.}

\item{movements}{A list of movements for each target tag, created by groupMovements.}

\item{spatial}{The spatial data frame.}

\item{status.df}{The main results table.}

\item{type}{The type of y axis desired. One of "stations" (default) or "arrays".}

\item{extension}{the format of the generated graphics}
}
\value{
A string of file locations in rmd syntax, to be included in printRmd
}
\description{
Prints the individual detections for each tag, overlaying the points in time considered crucial during the analysis.
}
\keyword{internal}
