% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{printIndividuals}
\alias{printIndividuals}
\title{Print individual graphics}
\usage{
printIndividuals(
  detections.list,
  bio,
  status.df = NULL,
  tz,
  movements,
  valid.movements = NULL,
  extension = "png",
  spatial
)
}
\arguments{
\item{detections.list}{A list of the detections split by each target tag, created by splitDetections.}

\item{bio}{A table with the tags and biometrics of the studied fish.}

\item{status.df}{The main results table.}

\item{tz}{The time zone of the study area. Must match one of the values
present in \code{\link[base]{timezones}}.}

\item{movements}{A list of movements for each target tag, created by groupMovements.}

\item{extension}{the format of the generated graphics}

\item{spatial}{The spatial data frame.}
}
\value{
A string of file locations in rmd syntax, to be included in printRmd
}
\description{
Prints the individual detections for each fish, overlaying the points in time considered crucial during the analysis.
}
\keyword{internal}
