% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/migration.R
\name{printMigrationRmd}
\alias{printMigrationRmd}
\title{Print Rmd report}
\usage{
printMigrationRmd(
  override.fragment,
  biometric.fragment,
  section.overview,
  efficiency.fragment,
  display.progression,
  array.overview.fragment,
  survival.graph.size,
  individual.plots,
  circular.plots,
  sensor.plots,
  spatial,
  deployments,
  valid.detections,
  detections,
  detections.y.axis
)
}
\arguments{
\item{override.fragment}{Rmarkdown string specifying the type of report for the header.}

\item{biometric.fragment}{Rmarkdown string specifying the biometric graphics drawn.}

\item{section.overview}{A summary table with the number of tags that disappeared/moved onwards at each section.}

\item{efficiency.fragment}{Rmarkdown string specifying the efficiency results.}

\item{display.progression}{Logical. If TRUE, the progression plot has been created and can be displayed.}

\item{array.overview.fragment}{Rmarkdown string specifying the array overview results.}

\item{survival.graph.size}{Rmarkdown string specifying the type size of the survival graphics.}

\item{individual.plots}{Rmarkdown string specifying the name of the individual plots.}

\item{circular.plots}{Rmarkdown string specifying the name of the circular plots.}

\item{sensor.plots}{Rmarkdown string specifying the name of the sensor plots.}
}
\value{
No return value, called for side effects.
}
\description{
Creates a Rmd report and converts it to html.
}
\keyword{internal}
