% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/printFunctions.R
\name{printProgression}
\alias{printProgression}
\title{Print progression graphic}
\usage{
printProgression(status.df, overall.CJS, split.CJS, group.CJS)
}
\arguments{
\item{overall.CJS}{a single CJS with all the groups and release sites merged}

\item{split.CJS}{a list of CJS's for each group.release combination}

\item{group.CJS}{a list of CJS's for each group, with release sites merged}
}
\description{
Prints progression graphics per fish group and release site.
}
\keyword{internal}
