% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{printProgression}
\alias{printProgression}
\title{Print progression diagram}
\usage{
printProgression(dot, overall.CJS, spatial, status.df, print.releases)
}
\arguments{
\item{dot}{a dot data frame}

\item{overall.CJS}{a single CJS with all the groups and release sites merged}

\item{spatial}{A list of spatial objects in the study area}

\item{status.df}{A data frame with the final migration results}

\item{print.releases}{Logical: Should the release sites be printed in the
study area diagrams?}
}
\value{
No return value, called to plot and save graphic.
}
\description{
Print progression diagram
}
\keyword{internal}
