% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/residency.R
\name{printResidencyRmd}
\alias{printResidencyRmd}
\title{Print Rmd report}
\usage{
printResidencyRmd(biometric.fragment, individual.plots, spatial,
  deployments)
}
\arguments{
\item{biometric.fragment}{Rmarkdown string specifying the biometric graphics drawn.}

\item{individual.plots}{Rmarkdown string specifying the name of the individual plots.}
}
\description{
Creates a Rmd report and converts it to hmtl.
}
\keyword{internal}
