% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/residency.R
\name{printResidencyRmd}
\alias{printResidencyRmd}
\title{Print Rmd report}
\usage{
printResidencyRmd(
  override.fragment,
  biometric.fragment,
  efficiency.fragment,
  individual.detection.plots,
  individual.residency.plots,
  array.circular.plots,
  section.arrival.circular.plots,
  section.departure.circular.plots,
  sensor.plots,
  spatial,
  deployments,
  detections,
  valid.detections,
  last.seen,
  detections.y.axis
)
}
\arguments{
\item{override.fragment}{Rmarkdown string specifying the type of report for the header.}

\item{biometric.fragment}{Rmarkdown string specifying the biometric graphics drawn.}

\item{efficiency.fragment}{Rmarkdown string specifying the efficiency results.}

\item{sensor.plots}{Rmarkdown string specifying the name of the sensor plots.}
}
\value{
No return value, called for side effects.
}
\description{
Creates a Rmd report and converts it to html.
}
\keyword{internal}
