% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{printSurvivalGraphic}
\alias{printSurvivalGraphic}
\title{Print survival graphic}
\usage{
printSurvivalGraphic(section.overview)
}
\arguments{
\item{section.overview}{A data frame containing the survival per animal group present in the biometrics. Supplied by assembleOverview.}
}
\value{
No return value, called to plot and save graphic.
}
\description{
Prints survival graphics per animal group.
}
\keyword{internal}
