% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadFunctions.R
\name{processThelmaFile}
\alias{processThelmaFile}
\title{Thelma files}
\usage{
processThelmaFile(input)
}
\arguments{
\item{input}{the file name, supplied by findFiles.}
}
\value{
A dataframe of standardized detections from the input file.
}
\description{
Processes Thelma ALS files.
}
\keyword{internal}
