% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load.R
\name{processVemcoFile}
\alias{processVemcoFile}
\title{Vemco files}
\usage{
processVemcoFile(input)
}
\arguments{
\item{input}{the detections data frame.}
}
\value{
A data frame of standardized detections from the input file.
}
\description{
Processes Vemco ALS files
}
\keyword{internal}
