% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load.R
\name{readDot}
\alias{readDot}
\title{Read dot file}
\usage{
readDot(input = NULL, string = NULL)
}
\arguments{
\item{input}{A simple .dot file}
}
\value{
A table with A to B rows
}
\description{
Read dot file
}
\keyword{internal}
