% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CJS.R
\name{recalculateCJS}
\alias{recalculateCJS}
\title{Combine CJS model results}
\usage{
recalculateCJS(input, estimate = NULL, fixed.efficiency = NULL)
}
\arguments{
\item{input}{A list of CJS results to be combined}

\item{estimate}{an estimate of the last array's efficiency}

\item{fixed.efficiency}{A vector of fixed efficiency estimates from a more complete CJS model.}
}
\value{
The combined CJS
}
\description{
Combine CJS model results
}
\keyword{internal}
