% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/residency.R
\name{res_assembleOutput}
\alias{res_assembleOutput}
\title{Create status.df}
\usage{
res_assembleOutput(res.df, bio, spatial, tz)
}
\arguments{
\item{bio}{A table with the tags and biometrics of the studied animals.}

\item{spatial}{The spatial list.}

\item{tz}{The time zone of the study area. Must match one of the values
present in \code{\link[base]{timezones}}.}
}
\value{
A data frame containing all the final data for each tag.
}
\description{
Combines the timetable and the original biometrics.
}
\keyword{internal}
