% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{ringsAbs}
\alias{ringsAbs}
\title{Draw rings at absolute points}
\usage{
ringsAbs(plot.params, border, rings.lty, ring.text, ring.text.pos,
  ring.text.cex)
}
\arguments{
\item{border}{the colour to draw the border. The default, NULL, means to use par("fg"). Use border = NA to omit borders.}

\item{rings.lty}{line type of the rings, See help on par.}

\item{ring.text}{logical: if notes should be displayed.}

\item{ring.text.pos}{The position of the rings' text. Ignored if ring.text is set to FALSE.}

\item{ring.text.cex}{The size of the ring's text. Ignored if ring.text is set to FALSE.}
}
\description{
Adapted from RosediagRad (from the circular package) to draw rings inside the circular plot. Called if rings = TRUE in myRoseDiag
}
\details{
For more details about the circular package, visit its homepage at https://github.com/cran/circular
}
\keyword{internal}
