% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/residency.R
\name{sectionMovements}
\alias{sectionMovements}
\title{Compress array-movements into section-movements}
\usage{
sectionMovements(movements, sections)
}
\arguments{
\item{movements}{A list of movements for each target tag, created by groupMovements.}

\item{sections}{The sections in which the study was divided. Must be coincident with the names given to the ALS arrays. (i.e. if an array is 'River1', then the respective section is 'River')}
}
\value{
the section movements
}
\description{
Compress array-movements into section-movements
}
\keyword{internal}
