% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/movements.R
\name{sectionMovements}
\alias{sectionMovements}
\title{Compress array-movements into section-movements}
\usage{
sectionMovements(movements, spatial, valid.dist)
}
\arguments{
\item{movements}{A list of movements for each target tag, created by groupMovements.}

\item{spatial}{The spatial data list.}

\item{valid.dist}{Logical: Is a valid distances matrix being used?}
}
\value{
A data frame containing the section movements for the target tag.
}
\description{
Compress array-movements into section-movements
}
\keyword{internal}
