% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/movements.R
\name{sectionMovements}
\alias{sectionMovements}
\title{Compress array-movements into section-movements}
\usage{
sectionMovements(movements, sections, invalid.dist)
}
\arguments{
\item{movements}{A list of movements for each target tag, created by groupMovements.}

\item{sections}{The sections in which the study area is divided. Must be
coincident with the names given to the arrays. See the vignettes for more
details.}

\item{invalid.dist}{Whether or not the distance matrix supplied is valid for the study area.}
}
\value{
A data frame containing the section movements for the target fish.
}
\description{
Compress array-movements into section-movements
}
\keyword{internal}
