% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load.R
\name{setSpatialStandards}
\alias{setSpatialStandards}
\title{Create Standard Names for spatial elements}
\usage{
setSpatialStandards(input)
}
\arguments{
\item{input}{A data frame with spatial information.}
}
\value{
A data frame with the same information as the input plus a Standard.name column.
}
\description{
Includes standard names and also reprints 'spatial.csv'
}
\keyword{internal}
