% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CJS.R
\name{simpleCJS}
\alias{simpleCJS}
\title{Compute simple CJS model}
\usage{
simpleCJS(input, estimate = NULL, fixed.efficiency = NULL,
  silent = TRUE)
}
\arguments{
\item{input}{a detection matrix}

\item{estimate}{an estimate of the last array's efficiency}

\item{fixed.efficiency}{A vector of fixed efficiency estimates from a more complete CJS model.}

\item{silent}{logical: if TRUE, suppress messages.}
}
\value{
A summary of the CJS results
}
\description{
The calculations are based on 'Using mark-recapture models to estimate survival from telemetry data' by Perry et al. 2012
}
