% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/groupMovements.R
\name{simplifyMovements}
\alias{simplifyMovements}
\title{Removes invalid events}
\usage{
simplifyMovements(movements, bio, speed.method, dist.mat, invalid.dist)
}
\arguments{
\item{movements}{A list of movements for each target tag, created by groupMovements.}

\item{bio}{A table with the tags and biometrics of the studied fish.}

\item{speed.method}{One of 'last to first' or 'first to first'. In the former, the last detection on a given array/section is matched to the first detection on the next array/section (default). If changed to 'first to first', the first detection on two consecutive arrays/sections are used to perform the calculations.}

\item{dist.mat}{A matrix of the distances between the deployed ALS.}

\item{invalid.dist}{Whether or not the distance matrix supplied is valid for the study area.}
}
\value{
The movement dataframe containing only valid events
}
\description{
Remove invalid movement events and recalculte times/speeds.
}
\keyword{internal}
