% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/movements.R
\name{speedReleaseToFirst}
\alias{speedReleaseToFirst}
\title{Calculate time and speed since release.}
\usage{
speedReleaseToFirst(
  fish,
  bio,
  movements,
  dist.mat,
  speed.method,
  invalid.dist = FALSE
)
}
\arguments{
\item{fish}{The tag ID of the fish currently being analysed}

\item{bio}{A table with the tags and biometrics of the studied fish.}

\item{movements}{A list of movements for each target tag, created by groupMovements.}

\item{dist.mat}{A matrix of the distances between the deployed ALS.}

\item{speed.method}{Can take two forms: 'last to first' or 'last to last'.
If 'last to first' (default), the last detection on the previous array is matched
to the first detection on the target array to perform the calculations.
If 'last to last', the last detection on ´the previous array is matched to the
last detection on the target array to perform the calculations.}

\item{invalid.dist}{Whether or not the distance matrix supplied is valid for the study area.}
}
\value{
The movement data frame containing time and speed from release to first event.
}
\description{
Calculate time and speed since release.
}
\keyword{internal}
