% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assembleEfficiency.R
\name{splitCJS}
\alias{splitCJS}
\title{Split CJS matrix and calculate separate CJS}
\usage{
splitCJS(input, by, estimate = NULL)
}
\arguments{
\item{input}{The CJS matrix to be split}

\item{by}{a grouping vector with the same length has the number of rows in input}

\item{estimate}{an estimate of the last array's efficiency}
}
\value{
the split CJS
}
\description{
Split CJS matrix and calculate separate CJS
}
