% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load.R
\name{splitDetections}
\alias{splitDetections}
\title{Split detections by tag}
\usage{
splitDetections(detections, bio, exclude.tags = NULL)
}
\arguments{
\item{detections}{A data frame with all the detections. Supplied by loadDetections.}

\item{bio}{A table with the tags and biometrics of the studied animals.}

\item{exclude.tags}{A vector of tags that should be excluded from the
detection data before any analyses are performed. Intended to be used if
stray tags from a different code space but with the same signal as a target
tag are detected in the study area.}
}
\value{
A list of detections for each tag.
}
\description{
Splits the detections' table by tags and selects only detections from target tags
}
\keyword{internal}
