% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadFunctions.R
\name{standardizeStations}
\alias{standardizeStations}
\title{Standardize Stations}
\usage{
standardizeStations(input, spatial)
}
\arguments{
\item{input}{A dataframe of standardized detections.}

\item{spatial}{A list of spatial objects in the study area}
}
\value{
A dataframe with standardized station names.
}
\description{
Matches the ALS serial number to the stations in the study area, standardizing the ALS station names.
}
\keyword{internal}
