% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{std.error.circular}
\alias{std.error.circular}
\title{Calculate the standard error of the mean}
\usage{
std.error.circular(x, na.rm = TRUE, silent = FALSE)
}
\arguments{
\item{x}{input data}

\item{na.rm}{logical: if TRUE, missing values are removed.}

\item{silent}{logica: if TRUE, The number of NA's removed is not displayed.}
}
\value{
SDM
}
\description{
Calculate the standard error of the mean
}
