% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{stripCodeSpaces}
\alias{stripCodeSpaces}
\title{Remove Code Spaces from transmitter names}
\usage{
stripCodeSpaces(input)
}
\arguments{
\item{input}{A vector of transmitter names}
}
\description{
Remove Code Spaces from transmitter names
}
\keyword{internal}
