% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helperFunctions.R
\name{timeConverter}
\alias{timeConverter}
\title{Convert hh:mm:ss time to hh.hhh}
\usage{
timeConverter(input)
}
\arguments{
\item{input}{a string containing either hours:minutes or hours:minutes:seconds data.}
}
\value{
Decimal hour equivalent (single value or vector)
}
\description{
Converts hour:minute:second text strings to decimal hours
}
\keyword{internal}
