% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check.R
\name{transferValidity}
\alias{transferValidity}
\title{Transfer validity updates from valid movements to all movements}
\usage{
transferValidity(from, to)
}
\arguments{
\item{from}{The valid movements table with newly invalidated moves}

\item{to}{The target movements table}
}
\value{
A data frame with the movement events for the target fish and an updated 'Valid' column.
}
\description{
Transfer validity updates from valid movements to all movements
}
\keyword{internal}
