% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{transitionLayer}
\alias{transitionLayer}
\title{Calculate Transition Layer}
\usage{
transitionLayer(
  path = ".",
  shape,
  size,
  EPSGcode,
  coord.x = NULL,
  coord.y = NULL,
  buffer = NULL,
  directions = c(16, 8, 4),
  force = FALSE
)
}
\arguments{
\item{path}{The system path to the 'shape' file. Defaults to the current directory.}

\item{shape}{A shape file projected in a metric coordinate system.}

\item{size}{The pixel size, in metres.}

\item{EPSGcode}{The EPSG code of the shape file's coordinate system. DO NOT USE degree-based
coordinate systems.}

\item{coord.x, coord.y}{The names of the columns containing the x and y positions of the stations
in the spatial.csv file. Must be in the same coordinate system as the shape file.}

\item{buffer}{Artificially expand the shape file edges. Can be a single value (applied to all edges)
or four values (xmin, xmax, ymin, ymax).}

\item{directions}{The number of directions considered for every movement situation during cost
calculation. See the vignettes for more details.}

\item{force}{Logical: Should the process continue even if the transition layer has 2000 pixels
on one or both axes?}
}
\value{
A TransitionLayer object.
}
\description{
\code{transitionLayer()} imports a shape file into R and prepares it to be used in distance
estimations. Adapted from Grant Adams' script "distance to closest mpa". This function creates
a 'transition.layer.RData' file in the current directory, which is then used by \code{\link{distancesMatrix}}.
}
\details{
It is highly recommended to read the vignette regarding distances matrix before running this function.
You can find it by running \code{vignette('a-2_distances_matrix', 'actel')} or \code{browseVignettes('actel')}
}
\examples{
\donttest{
# check if R can run the distance functions
aux <- c(
  length(suppressWarnings(packageDescription("raster"))),
  length(suppressWarnings(packageDescription("gdistance"))),
  length(suppressWarnings(packageDescription("sp"))),
  length(suppressWarnings(packageDescription("tools"))),
  length(suppressWarnings(packageDescription("rgdal"))))
missing.packages <- sapply(aux, function(x) x == 1)

if (any(missing.packages)) {
  message("Sorry, this function requires packages '", 
    paste(c("raster", "gdistance", "sp", "tools", "rgdal")[missing.packages], collapse = "', '"), 
    "' to operate. Please install ", ifelse(sum(missing.packages) > 1, "them", "it"), 
    " before proceeding.")
} else {
  if (suppressWarnings(require("rgdal"))) {
    # Fetch actel's example shapefile location
    aux <- system.file("example_shapefile", package = "actel")[1]

    # import the shape file
    x <- transitionLayer(path = aux, shape = "example_shapefile.shp", size = 20, EPSGcode = 32632)
  } else {
    message("Sorry, it appears that rgdal is not being able to load.")
  }
}
rm(aux, missing.packages)
}
}
