% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distanceMatrix.R
\name{transitionLayer}
\alias{transitionLayer}
\title{Calculate distance matrix}
\usage{
transitionLayer(shape, size, EPSGcode, directions = c(16, 8, 4),
  force = FALSE)
}
\arguments{
\item{shape}{A shape file from which to create the transition layer.}

\item{size}{The pixel size, in metres.}

\item{EPSGcode}{The EPSG code corresponding to the coordinate system of the input data. All inputs must be in the same metric system. DO NOT use degree-based coordinates.}

\item{directions}{The number of directions considered for every movement situation during cost calculation.}

\item{force}{logical: if TRUE, allows producing transition layers with more than 2000 pixels on one or both axes.}
}
\description{
Adapted from Grant Adams' script "distance to closest mpa". (grant.adams@eagles.usm.edu)
}
\details{
https://cran.r-project.org/web/packages/gdistance/vignettes/gdistance1.pdf
}
