% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{transitionLayer}
\alias{transitionLayer}
\title{Calculate Transition Layer}
\usage{
transitionLayer(
  shape,
  size,
  EPSGcode,
  coord.x = NULL,
  coord.y = NULL,
  buffer = NULL,
  directions = c(16, 8, 4),
  force = FALSE
)
}
\arguments{
\item{shape}{A shape file projected in a metric coordinate system.}

\item{size}{The pixel size, in metres.}

\item{EPSGcode}{The EPSG code of the shape file's coordinate system. DO NOT USE degree-based coordinate systems.}

\item{coord.x, coord.y}{The names of the columns containing the x and y positions of the stations in the spatial.csv file. Must be in the same coordinate system as the shape file.}

\item{buffer}{Artificially expand the shape file edges. Can be a single value (applied to all edges) or four values (xmin, xmax, ymin, ymax).}

\item{directions}{The number of directions considered for every movement situation during cost calculation. See the vignettes for more details.}

\item{force}{Logical: Should the process continue even if the transition layer has 2000 pixels on one or both axes?}
}
\value{
A RData file with the transition layer is stored in the current directory.
}
\description{
\code{transitionLayer} imports a shape file into R and prepares it to be used in distance
estimations. Adapted from Grant Adams' script "distance to closest mpa".
(grant.adams@eagles.usm.edu)
}
\details{
\url{https://cran.r-project.org/web/packages/gdistance/vignettes/gdistance1.pdf}
}
