% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/z_deprecated.R
\name{unknownReceiversCheckA}
\alias{unknownReceiversCheckA}
\title{Check for unknown receivers in the raw detections}
\usage{
unknownReceiversCheckA(detections, spatial)
}
\arguments{
\item{detections}{A data frame with all the detections. Supplied by loadDetections.}
}
\description{
Check for unknown receivers in the raw detections
}
\keyword{internal}
