% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/movements.R
\name{updateValidity}
\alias{updateValidity}
\title{update array-movement validity based on the section-movements}
\usage{
updateValidity(arrmoves, secmoves)
}
\value{
A list with the updated array movements.
}
\description{
update array-movement validity based on the section-movements
}
\keyword{internal}
