% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/explore.R
\name{validateDetections}
\alias{validateDetections}
\title{Compare original detections with the valid movements and exclude invalid detections}
\usage{
validateDetections(detections.list, movements)
}
\arguments{
\item{detections.list}{The list of detections per tag}

\item{movements}{The list of movements to be matched}
}
\value{
A list containing the valid detections per tag.
}
\description{
Compare original detections with the valid movements and exclude invalid detections
}
\keyword{internal}
